### 2023-02-26 #5190 @editorialbot
[PRE REVIEW]: SLEPLET: Slepian scale-discretised wavelets in Python
**Submitting author:** @paddyroddy (<a href="http://orcid.org/0000-0002-6271-1700">Patrick James Roddy</a>)
**Repository:** https://github.com/astro-informatics/sleplet
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.0
