### 2023-02-27 f4d0352
Add JOSS badge to README
