### 2023-03-02 7c2c7ce
Don't rely too much on `ruff` `ignore`
