+++ src/sleplet/plotting/polar_cap/simons_5_3.py
@@ -40,7 +40,7 @@ def _create_plot(ax: npt.NDArray, position: tuple[int, int], theta_max: int) -> 
     slepian = SlepianPolarCap(L, np.deg2rad(theta_max))
     axs = ax[position]
     legend = "full" if position == LEGEND_POS else False
-    assert isinstance(slepian.order, np.ndarray)
+    assert isinstance(slepian.order, np.ndarray)  # noqa: S101
     orders = np.abs(slepian.order[:RANKS])
     labels = np.array([f"$\\pm${m}" if m != 0 else m for m in orders])
     idx = np.argsort(orders)
