+++ pyproject.toml
@@ -86,14 +86,9 @@ force-exclude = true
 ignore = [
     "EM101",
     "EM102",
-    "N801",
-    "N802",
     "N803",
     "N805",
     "N806",
-    "N812",
-    "PLR0913",
-    "S101",
 ]
 select = [
     "A",
