### 2023-03-02 85c0783
Don't rely too much on `ruff` `ignore` (#75)
Use `noqa` more instead
