+++ tests/test_noise.py
@@ -19,7 +19,7 @@ def test_denoising_earth_axisymmetric_wavelets() -> None:
     fun_noised = Earth(L, noise=SNR_IN)
     aw = AxisymmetricWavelets(L, B=B, j_min=J_MIN)
     _, noised_snr, denoised_snr = denoising_axisym(fun, fun_noised, aw, SNR_IN, N_SIGMA)
-    assert isinstance(noised_snr, float)
+    assert isinstance(noised_snr, float)  # noqa: S101
     assert_array_less(noised_snr, denoised_snr)
 
 
