+++ src/sleplet/utils/mesh_methods.py
@@ -4,7 +4,7 @@ import numpy as np
 import tomli
 from igl import average_onto_faces, cotmatrix, read_triangle_mesh, upsample
 from numpy import typing as npt
-from scipy.sparse import linalg as LA_sparse
+from scipy.sparse import linalg as LA_sparse  # noqa: N812
 
 from sleplet.utils.config import settings
 from sleplet.utils.integration_methods import integrate_whole_mesh
