+++ pyproject.toml
@@ -118,7 +118,10 @@ select = [
 ]
 isort.known-first-party = ["sleplet"]
 mccabe.max-complexity = 18
-pep8-naming.classmethod-decorators = ["classmethod", "pydantic.validator"]
+pep8-naming.classmethod-decorators = [
+    "classmethod",
+    "pydantic.validator",
+]
 
 [tool.setuptools_scm]
 
