+++ pyproject.toml
@@ -116,6 +116,9 @@ select = [
     "W",
     "YTT",
 ]
+per-file-ignores = {"tests/*" = [
+    "S101",
+]}
 isort.known-first-party = ["sleplet"]
 mccabe.max-complexity = 18
 pep8-naming.classmethod-decorators = [
