### 2023-03-06 #5221 @editorialbot
[REVIEW]: SLEPLET: Slepian scale-discretised wavelets in Python
**Submitting author:** @paddyroddy (<a href="http://orcid.org/0000-0002-6271-1700">Patrick James Roddy</a>)
**Repository:** https://github.com/astro-informatics/sleplet
**Branch with paper.md** (empty if default branch): main
**Version:** v1.3.6
