### 2023-03-07 #5190 @louvetv
Hello

@vissarion <https://github.com/vissarion> could be ok for me depending
of the deadline for the review ?

<https://github.com/vissarion>
Le 06/03/2023 à 11:46, Vissarion Fisikopoulos a écrit :
>
> 👋 @max-little <https://github.com/max-little>, @mh-skjelvareid
> <https://github.com/mh-skjelvareid>, @klb2 <https://github.com/klb2>,
> @Saran-nns <https://github.com/Saran-nns>, @louvetv
> <https://github.com/louvetv>, @jerodway <https://github.com/jerodway>,
> @edeno <https://github.com/edeno> would any of you be willing to
> review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and
> follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
