+++ pyproject.toml
@@ -130,6 +130,8 @@ pep8-naming.classmethod-decorators = [
 
 [tool.tomlsort]
 sort_inline_arrays = false
+sort_inline_tables = true
+sort_table_keys = true
 spaces_indent_inline_array = 4
 trailing_comma_inline_array = true
 
