+++ pyproject.toml
@@ -89,6 +89,9 @@ ignore = [
     "N803",
     "N806",
 ]
+per-file-ignores = {"tests/*" = [
+    "S101",
+]}
 select = [
     "A",
     "B",
@@ -116,9 +119,6 @@ select = [
     "W",
     "YTT",
 ]
-per-file-ignores = {"tests/*" = [
-    "S101",
-]}
 isort.known-first-party = ["sleplet"]
 mccabe.max-complexity = 18
 pep8-naming.classmethod-decorators = [
@@ -130,6 +130,8 @@ pep8-naming.classmethod-decorators = [
 
 [tool.tomlsort]
 sort_inline_arrays = false
+sort_inline_tables = true
+sort_table_keys = true
 spaces_indent_inline_array = 4
 trailing_comma_inline_array = true
 
