### 2023-03-08 #5221 @klb2
I currently have issues installing the required packages `pyssht` and `pys2let` on my machine. I have opened a bug report in their projects.

After I am able to resolve these issues, I will continue my review of the submitted `sleplet` package.
