### 2023-03-13 #5221 @klb2
The developers of the `pyssht` and `pys2let` packages fixed the issue and I was now able to install the `sleplet` package.

I will continue my review this week.
