### 2023-03-14 #5221 @paddyroddy
> From the JOSS guidelines (https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain):
>
> > As this short list shows, JOSS papers are only expected to contain a limited set of metadata (see example below), a Statement of need, Summary, Acknowledgements, and References sections. You can look at an [example accepted paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.
>
> Based on this, I think that the "Demonstration" section should be completely removed from the paper and the "Conclusions" section should be merged into the "Summary" and "Statement of Needs" sections.

Have done
