### 2023-03-23 #5221 @klb2
In the paper, it is stated that
> To the author’s knowledge, there is no public software which allows one to compute Slepian wavelets (or similar approach) on the sphere or general manifolds/meshes.
> `SHTools` (Wieczorek & Meschede, 2018) and `slepian_alpha` (Simons et al., n.d.) are examples of codes which allow one to compute Slepian functions on the sphere. In conjunction with `SSHT` (McEwen & Wiaux, 2011), `S2LET` (Leistedt, B. et al., 2013) may be used to develop scale-discretised wavelets on the whole sphere.

and later

> Whilst Slepian wavelets may be trivially computed from a set of Slepian functions, the computation of the spherical Slepian functions themselves are computationally complex

Based on this, it is unclear for me what the advantages of the `sleplet` package are over the existing ones like `SHTools` and `slepian_alpha`. It is said in the paper that these two packages allow computing the Slepian functions and that it is trivial to compute the Slepian wavelets from the Slepian functions.
