+++ utils/mesh_methods.py
@@ -1,0 +1,143 @@
+from pathlib import Path
+
+import numpy as np
+import tomli
+from igl import average_onto_faces, cotmatrix, read_triangle_mesh, upsample
+from numpy import typing as npt
+from scipy.sparse import linalg as LA_sparse  # noqa: N812
+
+from sleplet.utils.config import settings
+from sleplet.utils.integration_methods import integrate_whole_mesh
+from sleplet.utils.logger import logger
+
+_file_location = Path(__file__).resolve()
+_meshes_path = _file_location.parents[1] / "data" / "meshes"
+
+
+def average_functions_on_vertices_to_faces(
+    faces: npt.NDArray[np.int_],
+    functions_on_vertices: npt.NDArray[np.complex_ | np.float_],
+) -> npt.NDArray[np.float_]:
+    """
+    the integrals require all functions to be defined on faces
+    this method handles an arbitrary number of functions
+    """
+    logger.info("converting function on vertices to faces")
+    # handle the case of a 1D array
+    array_is_1d = len(functions_on_vertices.shape) == 1
+    if array_is_1d:
+        functions_on_vertices = functions_on_vertices.reshape(1, -1)
+
+    functions_on_faces = np.zeros((functions_on_vertices.shape[0], faces.shape[0]))
+    for i, f in enumerate(functions_on_vertices):
+        functions_on_faces[i] = average_onto_faces(faces, f)
+
+    # put the vector back in 1D form
+    if array_is_1d:
+        functions_on_faces = functions_on_faces.reshape(-1)
+    return functions_on_faces
+
+
+def create_mesh_region(
+    mesh_config: dict, vertices: npt.NDArray[np.float_]
+) -> npt.NDArray[np.bool_]:
+    """
+    creates the boolean region for the given mesh
+    """
+    return (
+        (vertices[:, 0] >= mesh_config["XMIN"])
+        & (vertices[:, 0] <= mesh_config["XMAX"])
+        & (vertices[:, 1] >= mesh_config["YMIN"])
+        & (vertices[:, 1] <= mesh_config["YMAX"])
+        & (vertices[:, 2] >= mesh_config["ZMIN"])
+        & (vertices[:, 2] <= mesh_config["ZMAX"])
+    )
+
+
+def extract_mesh_config(mesh_name: str) -> dict:
+    """
+    reads in the given mesh region settings file
+    """
+    with open(_meshes_path / "regions" / f"{mesh_name}.toml", "rb") as f:
+        return tomli.load(f)
+
+
+def mesh_eigendecomposition(
+    name: str,
+    vertices: npt.NDArray[np.float_],
+    faces: npt.NDArray[np.int_],
+    *,
+    number_basis_functions: int | None = None,
+) -> tuple[npt.NDArray[np.float_], npt.NDArray[np.float_], int]:
+    """
+    computes the eigendecomposition of the mesh represented
+    as a graph if already computed then it loads the data
+    """
+    # determine number of basis functions
+    if number_basis_functions is None:
+        number_basis_functions = vertices.shape[0] // 4
+    logger.info(
+        f"finding {number_basis_functions}/{vertices.shape[0]} "
+        f"basis functions of {name} mesh"
+    )
+
+    # create filenames
+    eigd_loc = (
+        _meshes_path
+        / "laplacians"
+        / "basis_functions"
+        / f"{name}_b{number_basis_functions}"
+    )
+    eval_loc = eigd_loc / "eigenvalues.npy"
+    evec_loc = eigd_loc / "eigenvectors.npy"
+
+    if eval_loc.exists() and evec_loc.exists():
[…]
