### 2023-03-23 a7c576c
Install numpy for docs
