+++ .github/workflows/documentation.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install pdoc3
+          python -m pip install numpy pdoc3
 
       - name: Create documentation
         run: pdoc --html src/sleplet
