+++ pyproject.toml
@@ -132,11 +132,10 @@ pep8-naming.classmethod-decorators = [
 [tool.setuptools_scm]
 
 [tool.tomlsort]
-sort_inline_arrays = false
-sort_inline_tables = true
-sort_table_keys = true
+all = true
 spaces_indent_inline_array = 4
 trailing_comma_inline_array = true
+overrides."tool.coverage.paths.source".inline_arrays = false
 
 [tool.tox]
 legacy_tox_ini = """
