+++ .github/workflows/documentation.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip -e .[docs]
+          python -m pip install -e .[docs]
 
       - name: Create documentation
         run: pdoc --html src/sleplet
