+++ .github/workflows/documentation.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Set up python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.x"
+          python-version: "3.10"
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
 
