### 2023-03-23 #5221 @paddyroddy
> > Is there any documentation besides the examples in the README? I could not find an API documentation that documents all of the individual functions of the package.
>
> I hadn't realised full API documentation was required. I can look into this.

I've made some: https://astro-informatics.github.io/sleplet/
