+++ README.md
@@ -11,6 +11,7 @@
 Run
 
 ```sh
+git lfs pull
 pip install -e .
 ```
 
