+++ src/sleplet/__init__.py
@@ -1,10 +1,3 @@
 """
-`SLEPLET` is a Python package for the construction of Slepian wavelets in the
-spherical and manifold (via meshes) settings. The API of `SLEPLET` has been
-designed in an object-orientated manner and is easily extendible. Upon
-installation, `SLEPLET` comes with two command line interfaces - `sphere` and
-`mesh` - which allows one to easily generate plots on the sphere and a set of
-meshes using `plotly`.
-
 .. include:: ../../README.md
 """
