### 2023-03-23 df43f4d
Add README and Git LFS note (#85)
