+++ README.md
@@ -6,11 +6,19 @@
 [![tests](https://github.com/astro-informatics/sleplet/actions/workflows/tests.yml/badge.svg)](https://github.com/astro-informatics/sleplet/actions/workflows/tests.yml)
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
 
+`SLEPLET` is a Python package for the construction of Slepian wavelets in the
+spherical and manifold (via meshes) settings. The API of `SLEPLET` has been
+designed in an object-orientated manner and is easily extendible. Upon
+installation, `SLEPLET` comes with two command line interfaces - `sphere` and
+`mesh` - which allows one to easily generate plots on the sphere and a set of
+meshes using `plotly`.
+
 ## Installation
 
 Run
 
 ```sh
+git lfs pull
 pip install -e .
 ```
 
