+++ src/sleplet/slepian/slepian_region/slepian_limit_lat_lon.py
@@ -21,8 +21,7 @@ from sleplet.utils.vars import (
     THETA_MIN_DEFAULT,
 )
 
-_file_location = Path(__file__).resolve()
-_eigen_path = _file_location.parents[2] / "data" / "slepian" / "eigensolutions"
+_data_path = Path(__file__).resolve().parents[2] / "data"
 
 
 @dataclass(config=Validation, kw_only=True)
@@ -55,13 +54,20 @@ class SlepianLimitLatLon(SlepianFunctions):
         )
 
     def _create_matrix_location(self) -> Path:
-        return _eigen_path / f"D_{self.region.name_ending}_L{self.L}_N{self.N}"
+        return (
+            _data_path
+            / f"slepian_eigensolutions_D_{self.region.name_ending}_L{self.L}_N{self.N}"
+        )
 
     def _solve_eigenproblem(
         self,
     ) -> tuple[npt.NDArray[np.float_], npt.NDArray[np.complex_]]:
-        eval_loc = self.matrix_location / "eigenvalues.npy"
-        evec_loc = self.matrix_location / "eigenvectors.npy"
+        eval_loc = self.matrix_location.with_name(
+            f"{self.matrix_location.stem}_eigenvalues.npy"
+        )
+        evec_loc = self.matrix_location.with_name(
+            f"{self.matrix_location.stem}_eigenvectors.npy"
+        )
         if eval_loc.exists() and evec_loc.exists():
             logger.info("binaries found - loading...")
             return np.load(eval_loc), np.load(evec_loc)
