### 2023-03-24 66223bd
Adjust other pathlib stuff
