+++ src/sleplet/plotting/mesh/mesh_tiling.py
@@ -13,8 +13,7 @@ from sleplet.utils.class_lists import MESHES
 from sleplet.utils.plot_methods import save_plot
 from sleplet.utils.wavelet_methods import create_kappas
 
-file_location = Path(__file__).resolve()
-fig_path = file_location.parents[2] / "figures"
+fig_path = Path(__file__).resolve().parents[2] / "figures"
 sns.set(context="paper")
 
 B = 3
