+++ src/sleplet/plotting/wavelets/axisymmetric_tiling.py
@@ -8,8 +8,7 @@ from scipy.interpolate import pchip
 from sleplet.utils.plot_methods import save_plot
 from sleplet.utils.wavelet_methods import create_kappas
 
-file_location = Path(__file__).resolve()
-fig_path = file_location.parents[2] / "figures"
+fig_path = Path(__file__).resolve().parents[2] / "figures"
 sns.set(context="paper")
 
 B = 3
