+++ src/sleplet/plotting/polar_cap/slepian_error.py
@@ -17,8 +17,7 @@ from sleplet.utils.vars import SAMPLING_SCHEME
 
 L = 16
 
-file_location = Path(__file__).resolve()
-fig_path = file_location.parents[2] / "figures"
+fig_path = Path(__file__).resolve().parents[2] / "figures"
 sns.set(context="paper")
 
 
