+++ src/sleplet/utils/config.py
@@ -2,8 +2,7 @@ from pathlib import Path
 
 import tomli
 
-_file_location = Path(__file__).resolve()
-_settings_file = _file_location.parents[1] / "config" / "settings.toml"
+_settings_path = Path(__file__).resolve().parents[1] / "config"
 
-with open(_settings_file, "rb") as f:
+with open(_settings_path / "settings.toml", "rb") as f:
     settings = tomli.load(f)
