+++ src/sleplet/slepian/slepian_region/slepian_arbitrary.py
@@ -31,8 +31,7 @@ from sleplet.utils.slepian_arbitrary_methods import clean_evals_and_evecs
 from sleplet.utils.validation import Validation
 from sleplet.utils.vars import L_MAX_DEFAULT, L_MIN_DEFAULT
 
-_file_location = Path(__file__).resolve()
-_data_path = _file_location.parents[2] / "data"
+_data_path = Path(__file__).resolve().parents[2] / "data"
 
 
 @dataclass(config=Validation)
