+++ src/sleplet/data/other/earth/create_south_america_mask.py
@@ -7,8 +7,7 @@ from sleplet.data.other.earth.create_earth_flm import create_flm
 from sleplet.utils.plot_methods import rotate_earth_to_south_america
 from sleplet.utils.vars import SAMPLING_SCHEME, SOUTH_AMERICA_RANGE
 
-_file_location = Path(__file__).resolve()
-_mask_path = _file_location.parents[3] / "data" / "slepian" / "masks"
+_data_path = Path(__file__).resolve().parents[3] / "data"
 
 
 def create_mask(L: int) -> None:
@@ -20,7 +19,7 @@ def create_mask(L: int) -> None:
     earth_f = ssht.inverse(rot_flm, L, Reality=True, Method=SAMPLING_SCHEME)
     thetas, _ = ssht.sample_positions(L, Grid=True, Method=SAMPLING_SCHEME)
     mask = (thetas <= SOUTH_AMERICA_RANGE) & (earth_f >= 0)
-    np.save(_mask_path / f"south_america_L{L}.npy", mask)
+    np.save(_data_path / f"slepian_masks_south_america_L{L}.npy", mask)
 
 
 if __name__ == "__main__":
