+++ .github/workflows/tests.yml
@@ -17,14 +17,14 @@ jobs:
         uses: nschloe/action-cached-lfs-checkout@v1
         with:
           include: |
-            *bird_b697*,\
             *D_polar40_L16_N30*,\
             *D_south_america_L16_N10*,\
             EGM2008_Topography_flms_L2190.mat,\
-            *polar40_L32*,\
+            meshes_laplacians_*bird_b697*,\
+            slepian_masks_polar40_L32*,\
             slepian_masks_south_america_L16*,\
             slepian_masks_south_america_L32*,\
-            *theta30-60_phi30-60_L32*
+            slepian_masks_theta30-60_phi30-60_L32*
 
       - name: Cache tox
         uses: actions/cache@v3
