### 2023-03-24 08fda89
Move data files to one directory (#86)
First part in using `pooch`
