+++ src/sleplet/plotting/polar_cap/eigenfunctions.py
@@ -23,7 +23,7 @@ SIGNS = [1, -1, 1, -1, 1, -1]
 
 file_location = Path(__file__).resolve()
 fig_path = file_location.parents[2] / "figures"
-data_path = file_location.parents[2] / "data" / "slepian" / "eigensolutions"
+data_path = file_location.parents[2] / "data"
 sns.set(context="paper")
 
 
@@ -80,7 +80,10 @@ def _find_p_value(rank: int, shannon: int) -> int:
     """
     method to find the effective p rank of the Slepian function
     """
-    orders = np.load(data_path / f"D_polar{THETA_MAX}_L{L}_N{shannon}" / "orders.npy")
+    orders = np.load(
+        data_path
+        / f"slepian_eigensolutions_D_polar{THETA_MAX}_L{L}_N{shannon}_orders.npy"
+    )
     return np.where(orders == ORDER)[0][rank]
 
 
