+++ src/sleplet/plotting/polar_cap/slepian_coefficients.py
@@ -14,8 +14,7 @@ from sleplet.utils.plot_methods import save_plot
 
 L = 16
 
-file_location = Path(__file__).resolve()
-fig_path = file_location.parents[2] / "figures"
+fig_path = Path(__file__).resolve().parents[2] / "figures"
 sns.set(context="paper")
 
 
