+++ src/sleplet/slepian/slepian_region/slepian_arbitrary.py
@@ -31,8 +31,7 @@ from sleplet.utils.slepian_arbitrary_methods import clean_evals_and_evecs
 from sleplet.utils.validation import Validation
 from sleplet.utils.vars import L_MAX_DEFAULT, L_MIN_DEFAULT
 
-_file_location = Path(__file__).resolve()
-_slepian_path = _file_location.parents[2] / "data" / "slepian"
+_data_path = Path(__file__).resolve().parents[2] / "data"
 
 
 @dataclass(config=Validation)
@@ -61,14 +60,19 @@ class SlepianArbitrary(SlepianFunctions):
 
     def _create_matrix_location(self) -> Path:
         return (
-            _slepian_path / "eigensolutions" / f"D_{self.mask_name}_L{self.L}_N{self.N}"
+            _data_path
+            / f"slepian_eigensolutions_D_{self.mask_name}_L{self.L}_N{self.N}"
         )
 
     def _solve_eigenproblem(
         self,
     ) -> tuple[npt.NDArray[np.float_], npt.NDArray[np.complex_]]:
-        eval_loc = self.matrix_location / "eigenvalues.npy"
-        evec_loc = self.matrix_location / "eigenvectors.npy"
+        eval_loc = self.matrix_location.with_name(
+            f"{self.matrix_location.name}_eigenvalues.npy"
+        )
+        evec_loc = self.matrix_location.with_name(
+            f"{self.matrix_location.name}_eigenvectors.npy"
+        )
         if not eval_loc.exists() or not evec_loc.exists():
             return self._solve_D_matrix(eval_loc, evec_loc)
 
