+++ pyproject.toml
@@ -26,6 +26,7 @@ dependencies = [
     "numpy>=1.23.5",
     "pandas>=1.5.3",
     "plotly>=5.13.0",
+    "pooch>=1.7.0",
     "pydantic>=1.10.4",
     "pys2let>=2.2.4",
     "pyssht>=1.5.1",
