### 2023-03-24 8e67e51
Remove Slepian masks
