+++ src/sleplet/functions/flm/africa.py
@@ -1,19 +1,16 @@
-from pathlib import Path
-
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 from sleplet.data.other.earth.create_earth_flm import create_flm
+from sleplet.data.setup_pooch import POOCH
 from sleplet.functions.f_lm import F_LM
 from sleplet.utils.harmonic_methods import ensure_f_bandlimited
 from sleplet.utils.plot_methods import rotate_earth_to_africa
 from sleplet.utils.string_methods import convert_camel_case_to_snake_case
 from sleplet.utils.validation import Validation
 from sleplet.utils.vars import SAMPLING_SCHEME
-
-_data_path = Path(__file__).resolve().parents[2] / "data"
 
 
 @dataclass(config=Validation)
@@ -52,5 +49,5 @@ class Africa(F_LM):
         earth_f = ssht.inverse(
             rot_flm, self.L, Reality=self.reality, Method=SAMPLING_SCHEME
         )
-        mask = np.load(_data_path / f"slepian_masks_{self.name}_L{self.L}.npy")
+        mask = np.load(POOCH.fetch(f"slepian_masks_{self.name}_L{self.L}.npy"))
         return np.where(mask, earth_f, 0)
