+++ src/sleplet/functions/flm/south_america.py
@@ -6,6 +6,7 @@ from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 from sleplet.data.other.earth.create_earth_flm import create_flm
+from sleplet.data.setup_pooch import POOCH
 from sleplet.functions.f_lm import F_LM
 from sleplet.utils.harmonic_methods import ensure_f_bandlimited
 from sleplet.utils.plot_methods import rotate_earth_to_south_america
@@ -52,5 +53,5 @@ class SouthAmerica(F_LM):
         earth_f = ssht.inverse(
             rot_flm, self.L, Reality=self.reality, Method=SAMPLING_SCHEME
         )
-        mask = np.load(_data_path / f"slepian_masks_{self.name}_L{self.L}.npy")
+        mask = np.load(POOCH.fetch(f"slepian_masks_{self.name}_L{self.L}.npy"))
         return np.where(mask, earth_f, 0)
