### 2023-03-24 6eefa4a
Fix mask method
