+++ src/sleplet/functions/flm/south_america.py
@@ -53,5 +53,10 @@ class SouthAmerica(F_LM):
         earth_f = ssht.inverse(
             rot_flm, self.L, Reality=self.reality, Method=SAMPLING_SCHEME
         )
-        mask = np.load(POOCH.fetch(f"slepian_masks_{self.name}_L{self.L}.npy"))
+        mask_name = f"slepian_masks_{self.name}_L{self.L}.npy"
+        mask = np.load(
+            POOCH.fetch(mask_name)
+            if mask_name in POOCH.registry
+            else _data_path / mask_name
+        )
         return np.where(mask, earth_f, 0)
