+++ src/sleplet/data/setup_pooch.py
@@ -1,3 +1,6 @@
+import os
+from pathlib import Path
+
 import pooch
 
 from sleplet.utils.vars import ZENODO_DATA_DOI
@@ -9,3 +12,10 @@ POOCH = pooch.create(
     registry=None,
 )
 POOCH.load_registry_from_doi()
+
+
+def find_on_pooch_then_local(data_path: Path, filename: str) -> os.PathLike:
+    """
+    find a file on POOCH first and if not look in data folder
+    """
+    return POOCH.fetch(filename) if filename in POOCH.registry else data_path / filename
