+++ src/sleplet/functions/flm/africa.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
@@ -13,8 +11,6 @@ from sleplet.utils.plot_methods import rotate_earth_to_africa
 from sleplet.utils.string_methods import convert_camel_case_to_snake_case
 from sleplet.utils.validation import Validation
 from sleplet.utils.vars import SAMPLING_SCHEME
-
-_data_path = Path(__file__).resolve().parents[2] / "data"
 
 
 @dataclass(config=Validation)
