### 2023-03-24 4841e33
Print pooch cache path
