### 2023-03-25 0589fc0
Remove pooch from pip
