+++ .github/workflows/tests.yml
@@ -34,7 +34,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install pooch tox tox-gh-actions
+          python -m pip install tox tox-gh-actions
 
       - name: Test with tox
         run: tox
