+++ src/sleplet/plotting/polar_cap/eigenfunctions.py
@@ -22,9 +22,7 @@ RESOLUTION = calc_plot_resolution(L)
 SIGNS = [1, -1, 1, -1, 1, -1]
 
 
-file_location = Path(__file__).resolve()
-fig_path = file_location.parents[2] / "figures"
-data_path = file_location.parents[2] / "data"
+fig_path = Path(__file__).resolve().parents[2] / "figures"
 sns.set(context="paper")
 
 
