### 2023-03-25 5111df0
Use find pooch method everywhere
