### 2023-03-25 4673dc7
Remove Path object
