+++ src/sleplet/utils/mesh_methods.py
@@ -6,6 +6,7 @@ from igl import average_onto_faces, cotmatrix, read_triangle_mesh, upsample
 from numpy import typing as npt
 from scipy.sparse import linalg as LA_sparse  # noqa: N812
 
+from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.utils.config import settings
 from sleplet.utils.integration_methods import integrate_whole_mesh
 from sleplet.utils.logger import logger
@@ -81,18 +82,14 @@ def mesh_eigendecomposition(
     )
 
     # create filenames
-    eigd_loc = (
-        _data_path
-        / f"meshes_laplacians_basis_functions_{name}_b{number_basis_functions}"
-    )
-    eval_loc = eigd_loc.with_name(f"{eigd_loc.stem}_eigenvalues.npy")
-    evec_loc = eigd_loc.with_name(f"{eigd_loc.stem}_eigenvectors.npy")
+    eigd_loc = f"meshes_laplacians_basis_functions_{name}_b{number_basis_functions}"
+    eval_loc = f"{eigd_loc}_eigenvalues.npy"
+    evec_loc = f"{eigd_loc}_eigenvectors.npy"
 
-    if eval_loc.exists() and evec_loc.exists():
-        logger.info("binaries found - loading...")
-        eigenvalues = np.load(eval_loc)
-        eigenvectors = np.load(evec_loc)
-    else:
+    try:
+        eigenvalues = np.load(find_on_pooch_then_local(eval_loc))
+        eigenvectors = np.load(find_on_pooch_then_local(evec_loc))
+    except TypeError:
         laplacian = _mesh_laplacian(vertices, faces)
         eigenvalues, eigenvectors = LA_sparse.eigsh(
             laplacian, k=number_basis_functions, which="LM", sigma=0
@@ -100,8 +97,8 @@ def mesh_eigendecomposition(
         eigenvectors = _orthonormalise_basis_functions(vertices, faces, eigenvectors.T)
         if settings["SAVE_MATRICES"]:
             logger.info("saving binaries...")
-            np.save(eval_loc, eigenvalues)
-            np.save(evec_loc, eigenvectors)
+            np.save(_data_path / eval_loc, eigenvalues)
+            np.save(_data_path / evec_loc, eigenvectors)
     return eigenvalues, eigenvectors, number_basis_functions
 
 
