+++ src/sleplet/__init__.py
@@ -1,3 +1,7 @@
 """
 .. include:: ../../README.md
 """
+
+from sleplet._version import version
+
+__version__ = version
