### 2023-03-25 47143f4
Add version and fix PyPI (#88)
