+++ src/sleplet/__init__.py
@@ -1,3 +1,5 @@
 """
 .. include:: ../../README.md
 """
+
+from sleplet._version import __version__  # noqa: F401
