+++ pyproject.toml
@@ -131,6 +131,8 @@ pep8-naming.classmethod-decorators = [
 ]
 
 [tool.setuptools_scm]
+local_scheme = "no-local-version"
+write_to = "src/sleplet/_version.py"
 
 [tool.tomlsort]
 all = true
