### 2023-03-25 0dcbeaf
Move `logger` to top level
