+++ src/sleplet/utils/slepian_methods.py
@@ -2,6 +2,7 @@ import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
+from sleplet import logger
 from sleplet.meshes.classes.mesh_slepian import MeshSlepian
 from sleplet.meshes.classes.mesh_slepian_decomposition import MeshSlepianDecomposition
 from sleplet.slepian.slepian_decomposition import SlepianDecomposition
@@ -10,7 +11,6 @@ from sleplet.slepian.slepian_region.slepian_arbitrary import SlepianArbitrary
 from sleplet.slepian.slepian_region.slepian_limit_lat_lon import SlepianLimitLatLon
 from sleplet.slepian.slepian_region.slepian_polar_cap import SlepianPolarCap
 from sleplet.utils.harmonic_methods import boost_coefficient_resolution, mesh_inverse
-from sleplet.utils.logger import logger
 from sleplet.utils.region import Region
 from sleplet.utils.vars import SAMPLING_SCHEME
 
