+++ src/sleplet/plotting/create_plot_sphere.py
@@ -10,8 +10,8 @@ from plotly.graph_objs import Figure, Surface
 from plotly.graph_objs.surface import Lighting
 from pydantic.dataclasses import dataclass
 
+from sleplet import logger
 from sleplet.utils.config import settings
-from sleplet.utils.logger import logger
 from sleplet.utils.plot_methods import (
     boost_field,
     calc_plot_resolution,
