+++ src/sleplet/plotting/misc/translation_normalisation.py
@@ -5,8 +5,8 @@ import pyssht as ssht
 import seaborn as sns
 from matplotlib import pyplot as plt
 
+from sleplet import logger
 from sleplet.functions.flm.harmonic_gaussian import HarmonicGaussian
-from sleplet.utils.logger import logger
 from sleplet.utils.plot_methods import save_plot
 from sleplet.utils.vars import ALPHA_DEFAULT, SAMPLING_SCHEME
 
