+++ src/sleplet/slepian/slepian_region/slepian_arbitrary.py
@@ -9,6 +9,7 @@ from numpy import typing as npt
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 
+from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.slepian.slepian_functions import SlepianFunctions
 from sleplet.utils.array_methods import fill_upper_triangle_of_hermitian_matrix
@@ -18,7 +19,6 @@ from sleplet.utils.integration_methods import (
     calc_integration_weight,
     integrate_region_sphere,
 )
-from sleplet.utils.logger import logger
 from sleplet.utils.mask_methods import create_mask_region
 from sleplet.utils.parallel_methods import (
     attach_to_shared_memory_block,
