+++ src/sleplet/meshes/classes/mesh_slepian_decomposition.py
@@ -4,13 +4,13 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+from sleplet import logger
 from sleplet.meshes.classes.mesh_slepian import MeshSlepian
 from sleplet.utils.harmonic_methods import mesh_inverse
 from sleplet.utils.integration_methods import (
     integrate_region_mesh,
     integrate_whole_mesh,
 )
-from sleplet.utils.logger import logger
 from sleplet.utils.validation import Validation
 
 
