+++ src/sleplet/plotting/polar_cap/simons_5_3.py
@@ -5,9 +5,9 @@ import seaborn as sns
 from matplotlib import pyplot as plt
 from numpy import typing as npt
 
+from sleplet import logger
 from sleplet.plotting.inputs import TEXT_BOX
 from sleplet.slepian.slepian_region.slepian_polar_cap import SlepianPolarCap
-from sleplet.utils.logger import logger
 from sleplet.utils.plot_methods import save_plot
 
 L = 16
