+++ src/sleplet/functions/flm/directional_spin_wavelets.py
@@ -5,8 +5,8 @@ from pydantic import validator
 from pydantic.dataclasses import dataclass
 from pys2let import pys2let_j_max, wavelet_tiling
 
+from sleplet import logger
 from sleplet.functions.f_lm import F_LM
-from sleplet.utils.logger import logger
 from sleplet.utils.string_methods import (
     convert_camel_case_to_snake_case,
     filename_args,
