### 2023-03-25 890bde3
Move logger to top level (#89)
