+++ src/sleplet/utils/mask_methods.py
@@ -2,10 +2,10 @@ import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
+from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.meshes.classes.mesh import Mesh
 from sleplet.utils.harmonic_methods import mesh_forward, mesh_inverse
-from sleplet.utils.logger import logger
 from sleplet.utils.region import Region
 from sleplet.utils.vars import SAMPLING_SCHEME
 
