+++ src/sleplet/scripts/axisymmetric_wavelet_covariance.py
@@ -2,10 +2,10 @@ import numpy as np
 import pyssht as ssht
 from numpy.random import default_rng
 
+from sleplet import logger
 from sleplet.functions.flm.axisymmetric_wavelets import AxisymmetricWavelets
 from sleplet.utils.bool_methods import is_ergodic
 from sleplet.utils.harmonic_methods import compute_random_signal
-from sleplet.utils.logger import logger
 from sleplet.utils.vars import RANDOM_SEED, SAMPLING_SCHEME
 from sleplet.utils.wavelet_methods import (
     axisymmetric_wavelet_forward,
