+++ src/sleplet/utils/logger.py
@@ -1,8 +1,0 @@
-import logging
-
-logging.basicConfig(
-    format="[%(asctime)s] [%(levelname)s] --- %(message)s (%(filename)s:%(lineno)s)",
-    datefmt="%Y-%m-%d %H:%M:%S",
-    level=logging.INFO,
-)
-logger = logging.getLogger(__name__)
