+++ src/sleplet/scripts/plotting_on_mesh.py
@@ -4,13 +4,13 @@ from argparse import ArgumentParser, Namespace
 import numpy as np
 from numpy import typing as npt
 
+from sleplet import logger
 from sleplet.meshes.classes.mesh import Mesh
 from sleplet.meshes.mesh_coefficients import MeshCoefficients
 from sleplet.meshes.mesh_slepian_coefficients import MeshSlepianCoefficients
 from sleplet.plotting.create_plot_mesh import Plot
 from sleplet.utils.class_lists import MESH_COEFFICIENTS, MESHES
 from sleplet.utils.harmonic_methods import mesh_inverse
-from sleplet.utils.logger import logger
 from sleplet.utils.slepian_methods import slepian_mesh_inverse
 from sleplet.utils.string_methods import convert_classes_list_to_snake_case
 
