+++ src/sleplet/meshes/harmonic_coefficients/mesh_basis_functions.py
@@ -3,9 +3,9 @@ from numpy import typing as npt
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 
+from sleplet import logger
 from sleplet.meshes.mesh_harmonic_coefficients import MeshHarmonicCoefficients
 from sleplet.utils.harmonic_methods import mesh_forward
-from sleplet.utils.logger import logger
 from sleplet.utils.validation import Validation
 
 
