+++ src/sleplet/utils/mesh_methods.py
@@ -6,10 +6,10 @@ from igl import average_onto_faces, cotmatrix, read_triangle_mesh, upsample
 from numpy import typing as npt
 from scipy.sparse import linalg as LA_sparse  # noqa: N812
 
+from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.utils.config import settings
 from sleplet.utils.integration_methods import integrate_whole_mesh
-from sleplet.utils.logger import logger
 
 _data_path = Path(__file__).resolve().parents[1] / "data"
 
