+++ src/sleplet/__init__.py
@@ -2,4 +2,13 @@
 .. include:: ../../README.md
 """
 
+import logging
+
 from sleplet._version import __version__  # noqa: F401
+
+logging.basicConfig(
+    format="[%(asctime)s] [%(levelname)s] --- %(message)s (%(filename)s:%(lineno)s)",
+    datefmt="%Y-%m-%d %H:%M:%S",
+    level=logging.INFO,
+)
+logger = logging.getLogger(__name__)
