### 2023-03-25 1605501
Generate masks on the fly
