### 2023-03-25 035fceb
Start type guarding
