+++ src/sleplet/meshes/classes/mesh_slepian_decomposition.py
@@ -1,17 +1,20 @@
 from dataclasses import KW_ONLY
+from typing import TYPE_CHECKING
 
 import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 from sleplet import logger
-from sleplet.meshes.classes.mesh_slepian import MeshSlepian
 from sleplet.utils.harmonic_methods import mesh_inverse
 from sleplet.utils.integration_methods import (
     integrate_region_mesh,
     integrate_whole_mesh,
 )
 from sleplet.utils.validation import Validation
+
+if TYPE_CHECKING:
+    from sleplet.meshes.classes.mesh_slepian import MeshSlepian
 
 
 @dataclass(config=Validation)
