+++ src/sleplet/meshes/classes/mesh_slepian.py
@@ -1,5 +1,6 @@
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 import numpy as np
 from numpy import linalg as LA  # noqa: N812
@@ -8,7 +9,6 @@ from pydantic.dataclasses import dataclass
 
 from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
-from sleplet.meshes.classes.mesh import Mesh
 from sleplet.utils.array_methods import fill_upper_triangle_of_hermitian_matrix
 from sleplet.utils.config import settings
 from sleplet.utils.integration_methods import integrate_region_mesh
@@ -21,6 +21,10 @@ from sleplet.utils.parallel_methods import (
 )
 from sleplet.utils.slepian_arbitrary_methods import compute_mesh_shannon
 from sleplet.utils.validation import Validation
+
+if TYPE_CHECKING:
+    from sleplet.meshes.classes.mesh import Mesh
+
 
 _data_path = Path(__file__).resolve().parents[2] / "data"
 
