### 2023-03-25 2797029
More type guarding
