+++ src/sleplet/utils/harmonic_methods.py
@@ -1,14 +1,16 @@
 from collections.abc import Callable
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 from numpy.random import Generator
 
-from sleplet.meshes.classes.mesh import Mesh
 from sleplet.utils.integration_methods import integrate_whole_mesh
 from sleplet.utils.vars import SAMPLING_SCHEME
+
+if TYPE_CHECKING:
+    from sleplet.meshes.classes.mesh import Mesh
 
 
 def create_spherical_harmonic(L: int, ind: int) -> npt.NDArray[np.complex_]:
