+++ src/sleplet/slepian/slepian_decomposition.py
@@ -1,4 +1,5 @@
 from dataclasses import KW_ONLY
+from typing import TYPE_CHECKING
 
 import numpy as np
 import pyssht as ssht
@@ -6,7 +7,6 @@ from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 from sleplet import logger
-from sleplet.slepian.slepian_functions import SlepianFunctions
 from sleplet.utils.integration_methods import (
     calc_integration_weight,
     integrate_region_sphere,
@@ -14,6 +14,9 @@ from sleplet.utils.integration_methods import (
 )
 from sleplet.utils.validation import Validation
 from sleplet.utils.vars import SAMPLING_SCHEME
+
+if TYPE_CHECKING:
+    from sleplet.slepian.slepian_functions import SlepianFunctions
 
 
 @dataclass(config=Validation)
