+++ src/sleplet/slepian/slepian_functions.py
@@ -1,12 +1,15 @@
 from abc import abstractmethod
+from typing import TYPE_CHECKING
 
 import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 from sleplet import logger
-from sleplet.utils.region import Region
 from sleplet.utils.validation import Validation
+
+if TYPE_CHECKING:
+    from sleplet.utils.region import Region
 
 
 @dataclass(config=Validation)
