+++ src/sleplet/utils/denoising.py
@@ -1,15 +1,9 @@
+from typing import TYPE_CHECKING
+
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
-from sleplet.functions.coefficients import Coefficients
-from sleplet.functions.f_p import F_P
-from sleplet.functions.flm.axisymmetric_wavelets import AxisymmetricWavelets
-from sleplet.functions.fp.slepian_wavelets import SlepianWavelets
-from sleplet.meshes.mesh_coefficients import MeshCoefficients
-from sleplet.meshes.slepian_coefficients.mesh_slepian_wavelets import (
-    MeshSlepianWavelets,
-)
 from sleplet.utils.noise import (
     compute_sigma_j,
     compute_sigma_noise,
@@ -30,6 +24,16 @@ from sleplet.utils.wavelet_methods import (
     slepian_wavelet_forward,
     slepian_wavelet_inverse,
 )
+
+if TYPE_CHECKING:
+    from sleplet.functions.coefficients import Coefficients
+    from sleplet.functions.f_p import F_P
+    from sleplet.functions.flm.axisymmetric_wavelets import AxisymmetricWavelets
+    from sleplet.functions.fp.slepian_wavelets import SlepianWavelets
+    from sleplet.meshes.mesh_coefficients import MeshCoefficients
+    from sleplet.meshes.slepian_coefficients.mesh_slepian_wavelets import (
+        MeshSlepianWavelets,
+    )
 
 
 def denoising_axisym(
