+++ src/sleplet/meshes/classes/mesh_slepian_decomposition.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from dataclasses import KW_ONLY
 from typing import TYPE_CHECKING
 
