+++ src/sleplet/meshes/mesh_coefficients.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from abc import abstractmethod
 from dataclasses import KW_ONLY
 from typing import TYPE_CHECKING
