+++ src/sleplet/plotting/create_plot_mesh.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from dataclasses import KW_ONLY
 from pathlib import Path
 from typing import TYPE_CHECKING
