+++ src/sleplet/utils/class_lists.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from glob import glob
 from pathlib import Path
 from typing import TYPE_CHECKING
