### 2023-03-27 09dc703
Move to mask methods
