### 2023-03-27 36b789e
Use `hasattr` rather than checking if `F_P`
