+++ src/sleplet/functions/flm/south_america.py
@@ -6,8 +6,10 @@ from pydantic.dataclasses import dataclass
 from sleplet.data.other.earth.create_earth_flm import create_flm
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.functions.f_lm import F_LM
-from sleplet.utils.harmonic_methods import ensure_f_bandlimited
-from sleplet.utils.plot_methods import rotate_earth_to_south_america
+from sleplet.utils.harmonic_methods import (
+    ensure_f_bandlimited,
+    rotate_earth_to_south_america,
+)
 from sleplet.utils.string_methods import convert_camel_case_to_snake_case
 from sleplet.utils.validation import Validation
 from sleplet.utils.vars import SAMPLING_SCHEME
