+++ src/sleplet/functions/flm/noise_earth.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
