+++ src/sleplet/plotting/arbitrary/south_america/wavelet_reconstruction_south_america.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from sleplet.functions.fp.slepian_wavelet_coefficients_south_america import (
     SlepianWaveletCoefficientsSouthAmerica,
 )
