+++ src/sleplet/utils/string_methods.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import re
 from fractions import Fraction
 
