+++ src/sleplet/utils/region.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 
