### 2023-03-27 e2e8035
Revert "Add future annotations everywhere"
This reverts commit 079d67f4dccceeabad1a8c251f1b64e6adeafcb2.
