+++ src/sleplet/plotting/arbitrary/south_america/slepian_reconstruction_south_america.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from sleplet.functions.fp.slepian_south_america import SlepianSouthAmerica
 from sleplet.plotting.create_plot_sphere import Plot
 from sleplet.utils.region import Region
