+++ src/sleplet/scripts/slepian_energy_leakage.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import numpy as np
 from numpy.testing import assert_allclose
 
