+++ src/sleplet/utils/integration_methods.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from functools import reduce
 from typing import Any
 
