+++ src/sleplet/plotting/misc/denoising_axisym_south_america.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from sleplet.functions.flm.axisymmetric_wavelets import AxisymmetricWavelets
 from sleplet.functions.flm.south_america import SouthAmerica
 from sleplet.plotting.create_plot_sphere import Plot
