+++ src/sleplet/slepian/slepian_region/slepian_arbitrary.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import KW_ONLY
 from pathlib import Path
