+++ src/sleplet/functions/coefficients.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from abc import abstractmethod
 from dataclasses import KW_ONLY
 
