+++ src/sleplet/scripts/plotting_on_sphere.py
@@ -1,4 +1,6 @@
 #!/usr/bin/env python
+from __future__ import annotations
+
 from argparse import ArgumentParser, Namespace
 from typing import TYPE_CHECKING
 
