### 2023-03-27 4389560
Remove `TYPE_CHECKING` projects
