+++ src/sleplet/utils/plot_methods.py
@@ -1,7 +1,4 @@
-from __future__ import annotations
-
 from pathlib import Path
-from typing import TYPE_CHECKING
 
 import numpy as np
 import pyssht as ssht
@@ -10,18 +7,16 @@ from matplotlib import pyplot as plt
 from numpy import typing as npt
 
 from sleplet import logger
+from sleplet.functions.coefficients import Coefficients
 from sleplet.utils.config import settings
 from sleplet.utils.harmonic_methods import invert_flm_boosted
 from sleplet.utils.mask_methods import create_mask_region
+from sleplet.utils.region import Region
 from sleplet.utils.slepian_methods import slepian_inverse
 from sleplet.utils.vars import (
     SAMPLING_SCHEME,
     SPHERE_UNSEEN,
 )
-
-if TYPE_CHECKING:
-    from sleplet.functions.coefficients import Coefficients
-    from sleplet.utils.region import Region
 
 
 def calc_plot_resolution(L: int) -> int:
