+++ src/sleplet/meshes/classes/mesh_slepian.py
@@ -66,9 +66,8 @@ class MeshSlepian:
                 self.slepian_eigenvalues,
                 self.slepian_functions,
             ) = self._clean_evals_and_evecs(LA.eigh(D))
-            if settings["SAVE_MATRICES"]:
-                np.save(_data_path / eval_loc, self.slepian_eigenvalues)
-                np.save(_data_path / evec_loc, self.slepian_functions[: self.N])
+            np.save(_data_path / eval_loc, self.slepian_eigenvalues)
+            np.save(_data_path / evec_loc, self.slepian_functions[: self.N])
 
     def _create_D_matrix(self) -> npt.NDArray[np.float_]:  # noqa: N802
         """
