+++ src/sleplet/slepian/slepian_region/slepian_limit_lat_lon.py
@@ -10,7 +10,6 @@ from pydantic.dataclasses import dataclass
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.slepian.slepian_functions import SlepianFunctions
 from sleplet.utils.array_methods import fill_upper_triangle_of_hermitian_matrix
-from sleplet.utils.config import settings
 from sleplet.utils.mask_methods import create_mask_region
 from sleplet.utils.region import Region
 from sleplet.utils.validation import Validation
@@ -68,9 +67,8 @@ class SlepianLimitLatLon(SlepianFunctions):
         except TypeError:
             K = self._create_K_matrix()
             eigenvalues, eigenvectors = self._clean_evals_and_evecs(LA.eigh(K))
-            if settings["SAVE_MATRICES"]:
-                np.save(_data_path / eval_loc, eigenvalues)
-                np.save(_data_path / evec_loc, eigenvectors[: self.N])
+            np.save(_data_path / eval_loc, eigenvalues)
+            np.save(_data_path / evec_loc, eigenvectors[: self.N])
             return eigenvalues, eigenvectors
 
     def _create_K_matrix(self) -> npt.NDArray[np.complex_]:  # noqa: N802
