### 2023-03-27 342de30
Move private methods from README
