+++ src/sleplet/scripts/plotting_on_mesh.py
@@ -117,7 +117,7 @@ def _coefficients_to_field(
     )
 
 
-def _compute_amplitude_for_noisy_plots(f: MeshCoefficients) -> float | None:
+def compute_amplitude_for_noisy_plots(f: MeshCoefficients) -> float | None:
     """
     for the noised plots fix the amplitude to the initial data
     """
@@ -146,7 +146,7 @@ def main() -> None:
     )
 
     # custom amplitude for noisy plots
-    amplitude = _compute_amplitude_for_noisy_plots(f)
+    amplitude = compute_amplitude_for_noisy_plots(f)
 
     # perform plot
     plot(f, normalise=not args.unnormalise, amplitude=amplitude)
