### 2023-03-27 3c03317
Generate masks on the fly (#92)
