+++ src/sleplet/utils/mask_methods.py
@@ -1,13 +1,23 @@
+from pathlib import Path
+
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
 from sleplet import logger
+from sleplet.data.other.earth.create_earth_flm import create_flm
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.meshes.classes.mesh import Mesh
-from sleplet.utils.harmonic_methods import mesh_forward, mesh_inverse
+from sleplet.utils.harmonic_methods import (
+    mesh_forward,
+    mesh_inverse,
+    rotate_earth_to_africa,
+    rotate_earth_to_south_america,
+)
 from sleplet.utils.region import Region
-from sleplet.utils.vars import SAMPLING_SCHEME
+from sleplet.utils.vars import AFRICA_RANGE, SAMPLING_SCHEME, SOUTH_AMERICA_RANGE
+
+_data_path = Path(__file__).resolve().parents[1] / "data"
 
 
 def create_mask_region(L: int, region: Region) -> npt.NDArray[np.float_]:
@@ -25,7 +35,7 @@ def create_mask_region(L: int, region: Region) -> npt.NDArray[np.float_]:
         case "arbitrary":
             logger.info("loading and checking shape of provided mask")
             name = f"{region.mask_name}_L{L}.npy"
-            mask = _load_mask(name)
+            mask = _load_mask(L, name)
             assert mask.shape == thetas.shape, (  # noqa: S101
                 f"mask {name} has shape {mask.shape} which does not match "
                 f"the provided L={L}, the shape should be {thetas.shape}"
@@ -49,14 +59,12 @@ def create_mask_region(L: int, region: Region) -> npt.NDArray[np.float_]:
     return mask
 
 
-def _load_mask(mask_name: str) -> npt.NDArray[np.float_]:
+def _load_mask(L: int, mask_name: str) -> npt.NDArray[np.float_]:
     """
     attempts to read the mask from the config file
     """
     mask = find_on_pooch_then_local(f"slepian_masks_{mask_name}")
-    if mask is None:
-        raise FileNotFoundError(f"can not find the file: '{mask_name}'")
-    return np.load(mask)
+    return create_mask(L, mask_name) if mask is None else np.load(mask)
 
 
 def ensure_masked_flm_bandlimited(
@@ -121,3 +129,42 @@ def convert_region_on_vertices_to_faces(mesh: Mesh) -> npt.NDArray[np.float_]:
     region_on_faces = np.zeros(mesh.faces.shape[0])
     region_on_faces[faces_in_region] = 1
     return region_on_faces
+
+
+def _create_africa_mask(
+    L: int, earth_flm: npt.NDArray[np.complex_]
+) -> npt.NDArray[np.float_]:
+    """
+    creates the Africa region mask
+    """
+    rot_flm = rotate_earth_to_africa(earth_flm, L)
+    earth_f = ssht.inverse(rot_flm, L, Reality=True, Method=SAMPLING_SCHEME)
+    thetas, _ = ssht.sample_positions(L, Grid=True, Method=SAMPLING_SCHEME)
+    return (thetas <= AFRICA_RANGE) & (earth_f >= 0)
+
+
+def _create_south_america_mask(
+    L: int, earth_flm: npt.NDArray[np.complex_]
+) -> npt.NDArray[np.float_]:
+    """
+    creates the Africa region mask
+    """
+    rot_flm = rotate_earth_to_south_america(earth_flm, L)
+    earth_f = ssht.inverse(rot_flm, L, Reality=True, Method=SAMPLING_SCHEME)
[…]
