+++ src/sleplet/config/settings.toml
@@ -8,7 +8,6 @@ PHI_MIN = 0
 POLAR_GAP = false
 SAMPLES = 2
 SAVE_FIG = false
-SAVE_MATRICES = false
 SLEPIAN_MASK = "south_america"
 THETA_MAX = 180
 THETA_MIN = 0

[…]
