+++ src/sleplet/utils/plot_methods.py
@@ -73,11 +73,10 @@ def save_plot(path: Path, name: str) -> None:
     helper method to save plots
     """
     plt.tight_layout()
-    if settings["SAVE_FIG"]:
-        for file_type in {"png", "pdf"}:
-            logger.info(f"saving {file_type}")
-            filename = path / file_type / f"{name}.{file_type}"
-            plt.savefig(filename, bbox_inches="tight")
+    for file_type in {"png", "pdf"}:
+        logger.info(f"saving {file_type}")
+        filename = path / file_type / f"{name}.{file_type}"
+        plt.savefig(filename, bbox_inches="tight")
     if settings["AUTO_OPEN"]:
         plt.show()
 
