+++ src/sleplet/config/settings.toml
@@ -7,7 +7,6 @@ PHI_MAX = 360
 PHI_MIN = 0
 POLAR_GAP = false
 SAMPLES = 2
-SAVE_FIG = false
 SLEPIAN_MASK = "south_america"
 THETA_MAX = 180
 THETA_MIN = 0
