+++ src/sleplet/utils/vars.py
@@ -18,8 +18,6 @@ BETA_DEFAULT: float = 0.125
 SOUTH_AMERICA_ALPHA: float = np.deg2rad(54)
 SOUTH_AMERICA_BETA: float = np.deg2rad(108)
 SOUTH_AMERICA_GAMMA: float = np.deg2rad(63)
-L_MAX_DEFAULT: int = -1
-L_MIN_DEFAULT: int = 0
 MESH_CBAR_FONT_SIZE: int = 32
 MESH_CBAR_LEN: float = 0.95
 MESH_UNSEEN: float = -1e5  # kaleido bug
