+++ src/sleplet/utils/parallel_methods.py
@@ -2,16 +2,12 @@ import numpy as np
 from multiprocess.shared_memory import SharedMemory
 from numpy import typing as npt
 
-from sleplet.utils.vars import L_MIN_DEFAULT
 
-
-def split_arr_into_chunks(
-    arr_max: int, ncpu: int, *, arr_min: int = L_MIN_DEFAULT
-) -> list[npt.NDArray[np.int_]]:
+def split_arr_into_chunks(arr_max: int, ncpu: int) -> list[npt.NDArray[np.int_]]:
     """
     split L into a list of arrays for parallelism
     """
-    arr = np.arange(arr_min, arr_max)
+    arr = np.arange(arr_max)
     size = len(arr)
     arr[size // 2 : size] = arr[size // 2 : size][::-1]
     return [np.sort(arr[i::ncpu]) for i in range(ncpu)]
