+++ src/sleplet/config/settings.toml
@@ -3,7 +3,6 @@ NCPU = 4
 PHI_MAX = 360
 PHI_MIN = 0
 POLAR_GAP = false
-SAMPLES = 2
 SLEPIAN_MASK = "south_america"
 THETA_MAX = 180
 THETA_MIN = 0
