+++ src/sleplet/plotting/create_plot_mesh.py
@@ -12,7 +12,6 @@ from pydantic.dataclasses import dataclass
 
 from sleplet import logger
 from sleplet.meshes.classes.mesh import Mesh
-from sleplet.utils.config import settings
 from sleplet.utils.mask_methods import convert_region_on_vertices_to_faces
 from sleplet.utils.mesh_methods import average_functions_on_vertices_to_faces
 from sleplet.utils.plot_methods import convert_colourscale, normalise_function
@@ -22,7 +21,11 @@ from sleplet.utils.plotly_methods import (
     create_tick_mark,
 )
 from sleplet.utils.validation import Validation
-from sleplet.utils.vars import MESH_CBAR_FONT_SIZE, MESH_CBAR_LEN, MESH_UNSEEN
+from sleplet.utils.vars import (
+    MESH_CBAR_FONT_SIZE,
+    MESH_CBAR_LEN,
+    MESH_UNSEEN,
+)
 
 _fig_path = Path(__file__).resolve().parents[1] / "figures"
 
@@ -86,14 +89,13 @@ class Plot:
 
         fig = Figure(data=data, layout=layout)
 
-        # create html and open if auto_open is true
         html_filename = str(_fig_path / "html" / f"{self.filename}.html")
 
-        py.plot(fig, filename=html_filename, auto_open=settings["AUTO_OPEN"])
+        py.plot(fig, filename=html_filename)
 
         for file_type in {"png", "pdf"}:
-            logger.info(f"saving {file_type}")
             filename = str(_fig_path / file_type / f"{self.filename}.{file_type}")
+            logger.info(f"saving {filename}")
             fig.write_image(filename, engine="kaleido")
 
     def _prepare_field(
