### 2023-03-27 c9dcac0
Remove settings toml and use environment variables (#93)
