+++ src/sleplet/plotting/inputs.py
@@ -1,6 +1,0 @@
-TEXT_BOX: dict[str, str | float] = {"boxstyle": "round", "color": "w"}
-THETA_0: int = 0
-THETA_1: int = 50
-THETA_MAX: int = 40
-PHI_0: int = 0
-PHI_1: int = 120
