### 2023-03-27 f3b866f
Fix figure paths
