+++ examples/polar_cap/slepian_error.py
@@ -17,7 +17,7 @@ from sleplet.utils.vars import SAMPLING_SCHEME
 L = 16
 THETA_MAX = 40
 
-fig_path = Path(__file__).resolve().parents[2] / "figures"
+fig_path = Path(__file__).resolve().parents[3] / "figures"
 sns.set(context="paper")
 
 
