### 2023-03-27 b0877eb
Move scripts and fix SA & Africa
