### 2023-03-27 aa26ae0
Correct figure paths again
