+++ examples/wavelets/axisymmetric_tiling.py
@@ -8,7 +8,7 @@ from scipy.interpolate import pchip
 from sleplet.utils.plot_methods import save_plot
 from sleplet.utils.wavelet_methods import create_kappas
 
-fig_path = Path(__file__).resolve().parents[3] / "figures"
+fig_path = Path(__file__).resolve().parents[2] / "figures"
 sns.set(context="paper")
 
 B = 3
