+++ examples/mesh/mesh_tiling.py
@@ -13,7 +13,7 @@ from sleplet.utils.class_lists import MESHES
 from sleplet.utils.plot_methods import save_plot
 from sleplet.utils.wavelet_methods import create_kappas
 
-fig_path = Path(__file__).resolve().parents[3] / "figures"
+fig_path = Path(__file__).resolve().parents[2] / "figures"
 sns.set(context="paper")
 
 B = 3
