### 2023-03-27 7d213b0
Fix examples saving figures
