### 2023-03-27 59c157e
Add progress bar requirement
