+++ pyproject.toml
@@ -38,6 +38,7 @@ dependencies = [
     "scipy>=1.10.0",
     "seaborn>=0.12.2",
     "tomli>=2.0.1",
+    "tqdm>=4.65.0",
 ]
 description = "Slepian scale-discretised wavelets in Python"
 dynamic = [
