### 2023-03-27 4e85654
Add printing environment variables
