+++ README.md
@@ -63,7 +63,7 @@ default region is the `south_america` arbitrary region.
 
 To recreate the figures from the below papers, one may use the CLI or the API.
 For those which don't use the `mesh` or `sphere` CLIs, the relevant API code
-isn't provided as it is contained within the source code.
+isn't provided as it is contained within the [examples folder](./examples).
 
 ### Sifting Convolution on the Sphere
 
