### 2023-03-27 b8981a6
Take examples out of the module and put in their own folder (#95)
