+++ examples/arbitrary/africa/eigenvalues_africa.py
@@ -7,7 +7,7 @@ from matplotlib import pyplot as plt
 from sleplet.slepian.slepian_region.slepian_arbitrary import SlepianArbitrary
 from sleplet.utils.plot_methods import save_plot
 
-fig_path = Path(__file__).resolve().parents[3] / "figures"
+fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "figures"
 sns.set(context="paper")
 
 L = 128
@@ -17,14 +17,14 @@ def main() -> None:
     """
     plots the tiling of the Slepian line
     """
-    slepian = SlepianArbitrary(L, "south_america")
+    slepian = SlepianArbitrary(L, "africa")
     p_range = np.arange(0, L**2)
     plt.semilogx(p_range, slepian.eigenvalues, "k.")
     plt.axvline(slepian.N, c="k", ls="--", alpha=0.8)
     plt.annotate(
         f"N={slepian.N}",
         xy=(slepian.N, 1),
-        xytext=(17, 3),
+        xytext=(19, 3),
         ha="center",
         textcoords="offset points",
         annotation_clip=False,
@@ -33,7 +33,7 @@ def main() -> None:
     plt.xticks(ticks, ticks)
     plt.xlabel(r"$p$")
     plt.ylabel(r"$\mu$")
-    save_plot(fig_path, f"south_america_eigenvalues_L{L}")
+    save_plot(fig_path, f"africa_eigenvalues_L{L}")
 
 
 if __name__ == "__main__":
