+++ .github/workflows/tests_and_deploy.yml
@@ -38,3 +38,34 @@ jobs:
 
       - name: Test with tox
         run: tox
+
+  deploy:
+    needs: tests
+    runs-on: "ubuntu-latest"
+    if: |
+      github.event_name == 'push'
+      && github.ref_type == 'tag'
+
+    steps:
+      - name: Checkout source
+        uses: actions/checkout@v3
+
+      - name: Set up python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.10"
+          cache: "pip"
+          cache-dependency-path: "pyproject.toml"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install build
+
+      - name: Build distributions
+        run: python -m build
+
+      - name: Publish package to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          password: ${{ secrets.PYPI_API_TOKEN }}
