+++ .github/workflows/test_and_deploy.yml
@@ -44,7 +44,7 @@ jobs:
     runs-on: "ubuntu-latest"
     if: |
       github.event_name == 'push'
-      && github.ref_type == 'tag'
+      && startsWith(github.ref, 'refs/tags')
 
     steps:
       - name: Checkout source
