+++ .github/workflows/test_and_deploy.yml
@@ -41,11 +41,10 @@ jobs:
 
   deploy:
     needs: test
-    runs-on: "ubuntu-latest"
     if: |
       github.event_name == 'push'
-      && startsWith(github.ref, 'refs/tags')
-
+      && github.ref_type == 'tag'
+    runs-on: "ubuntu-latest"
     steps:
       - name: Checkout source
         uses: actions/checkout@v3
