+++ .github/workflows/test_and_deploy.yml
@@ -41,9 +41,7 @@ jobs:
 
   deploy:
     needs: test
-    if: |
-      github.event_name == 'push'
-      && github.ref_type == 'tag'
+    if: github.event_name == 'push' && github.ref_type == 'tag'
     runs-on: "ubuntu-latest"
     steps:
       - name: Checkout source
