+++ README.md
@@ -3,7 +3,7 @@
 [![SLEPLET](https://zenodo.org/badge/DOI/10.5281/zenodo.7268074.svg)](https://doi.org/10.5281/zenodo.7268074)
 [![Python](https://img.shields.io/badge/python-3.10-orange.svg)](https://www.python.org/downloads/release/python-3100/)
 [![Documentation](https://github.com/astro-informatics/sleplet/actions/workflows/documentation.yml/badge.svg)](https://astro-informatics.github.io/sleplet)
-[![Test & Deploy](https://github.com/astro-informatics/sleplet/actions/workflows/test_and_deploy.yml/badge.svg)](https://github.com/astro-informatics/sleplet/actions/workflows/test_and_deploy.yml)
+[![Test](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml/badge.svg)](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml)
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
 
 `SLEPLET` is a Python package for the construction of Slepian wavelets in the
