+++ .github/workflows/deploy.yml
@@ -7,9 +7,7 @@ on:
 
 jobs:
   deploy:
-    needs: test
     runs-on: "ubuntu-latest"
-
     steps:
       - name: Checkout source
         uses: actions/checkout@v3
