+++ README.md
@@ -1,9 +1,10 @@
 # SLEPLET
 
+![PyPI](https://img.shields.io/pypi/v/sleplet?label=pypi%20package)
 [![SLEPLET](https://zenodo.org/badge/DOI/10.5281/zenodo.7268074.svg)](https://doi.org/10.5281/zenodo.7268074)
 [![Python](https://img.shields.io/badge/python-3.10-orange.svg)](https://www.python.org/downloads/release/python-3100/)
 [![Documentation](https://github.com/astro-informatics/sleplet/actions/workflows/documentation.yml/badge.svg)](https://astro-informatics.github.io/sleplet)
-[![tests](https://github.com/astro-informatics/sleplet/actions/workflows/tests.yml/badge.svg)](https://github.com/astro-informatics/sleplet/actions/workflows/tests.yml)
+[![Test](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml/badge.svg)](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml)
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
 
 `SLEPLET` is a Python package for the construction of Slepian wavelets in the
@@ -15,7 +16,15 @@ meshes using `plotly`.
 
 ## Installation
 
-Run
+The recommended way to install `SLEPLET` is via
+[pip](https://pypi.org/project/pip/)
+
+```sh
+pip install sleplet
+```
+
+To install the latest development version of `SLEPLET` clone this repository
+and run
 
 ```sh
 pip install -e .
