+++ pyproject.toml
@@ -3,6 +3,7 @@ build-backend = "setuptools.build_meta"
 requires = [
     "setuptools",
     "setuptools-scm",
+    "wheel",
 ]
 
 [project]
