+++ README.md
@@ -1,6 +1,7 @@
 # SLEPLET
 
-[![PyPI](https://img.shields.io/pypi/pyversions/sleplet)](https://pypi.org/project/sleplet/)
+[![PyPI](https://badge.fury.io/py/sleplet.svg)](https://pypi.org/project/sleplet)
+[![Python](https://img.shields.io/pypi/pyversions/sleplet)](https://www.python.org)
 [![Documentation](https://github.com/astro-informatics/sleplet/actions/workflows/documentation.yml/badge.svg)](https://astro-informatics.github.io/sleplet)
 [![Zenodo](https://zenodo.org/badge/DOI/10.5281/zenodo.7268074.svg)](https://doi.org/10.5281/zenodo.7268074)
 [![Test](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml/badge.svg)](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml)
