+++ src/sleplet/__init__.py
@@ -1,5 +1,6 @@
 """
-API Documentation of the `SLEPLET` package.
+API of the `SLEPLET` package. See [PyPI](https://pypi.org/project/sleplet) for
+the rest of the documentation.
 
 [![PyPI](https://badge.fury.io/py/sleplet.svg)](https://pypi.org/project/sleplet)
 """
