+++ pyproject.toml
@@ -69,13 +69,11 @@ optional-dependencies.docs = [
     "pdoc3",
     "rundoc",
 ]
-
-[project.urls]
-Homepage = "https://github.com/astro-informatics/sleplet"
-Download = "https://pypi.org/project/sleplet"
-"Bug Tracker" = "https://github.com/astro-informatics/sleplet/issues"
-Documentation = "https://astro-informatics.github.io/sleplet"
-"Source Code" = "https://github.com/astro-informatics/sleplet"
+urls."Bug Tracker" = "https://github.com/astro-informatics/sleplet/issues"
+urls.Documentation = "https://astro-informatics.github.io/sleplet"
+urls.Download = "https://pypi.org/project/sleplet"
+urls.Homepage = "https://github.com/astro-informatics/sleplet"
+urls."Source Code" = "https://github.com/astro-informatics/sleplet"
 
 [tool.coverage]
 report = {skip_covered = true, sort = "cover"}
@@ -154,7 +152,6 @@ write_to = "src/sleplet/_version.py"
 all = true
 spaces_indent_inline_array = 4
 trailing_comma_inline_array = true
-overrides."project.urls".table_keys = false
 overrides."tool.coverage.paths.source".inline_arrays = false
 
 [tool.tox]
