### 2023-03-28 1a64f6e
Pick default mesh for examples
