+++ examples/mesh/mesh_slepian_eigenvalues.py
@@ -50,6 +50,9 @@ if __name__ == "__main__":
         type=valid_meshes,
         choices=MESHES,
         help="mesh to plot",
+        default="homer",
+        const="homer",
+        nargs="?",
     )
     parser.add_argument(
         "--basis",
