+++ examples/mesh/denoising_slepian_mesh.py
@@ -54,6 +54,9 @@ if __name__ == "__main__":
         type=valid_meshes,
         choices=MESHES,
         help="mesh to plot",
+        default="homer",
+        const="homer",
+        nargs="?",
     )
     parser.add_argument(
         "--noise",
