### 2023-03-28 3b6e92e
Make sure plots don't hang
