### 2023-03-28 dc1e14d
Add action to check examples in folder & README (#101)
As `workflow_dispatch` as it will take a long time
