+++ src/sleplet/__init__.py
@@ -1,5 +1,8 @@
 """
-.. include:: ../../README.md
+API of the `SLEPLET` package. See [PyPI](https://pypi.org/project/sleplet) for
+the rest of the documentation.
+
+[![PyPI](https://badge.fury.io/py/sleplet.svg)](https://pypi.org/project/sleplet)
 """
 
 import logging
