+++ .github/workflows/examples.yml
@@ -3,9 +3,6 @@ name: Examples
 on:
   workflow_dispatch:
 
-env:
-  PYTHON_VERSION: "3.10"
-
 jobs:
   readme:
     runs-on: ubuntu-latest
@@ -16,7 +13,7 @@ jobs:
       - name: Set up python
         uses: actions/setup-python@v4
         with:
-          python-version: $PYTHON_VERSION
+          python-version: "3.10"
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
 
@@ -37,7 +34,7 @@ jobs:
       - name: Set up python
         uses: actions/setup-python@v4
         with:
-          python-version: $PYTHON_VERSION
+          python-version: "3.10"
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
 
