+++ .github/workflows/examples.yml
@@ -48,5 +48,4 @@ jobs:
         run: |
           find examples -name '*.py' \
           -exec echo {} \; \
-          -exec echo  \; \
           -exec python {} \;
