+++ README.md
@@ -19,14 +19,14 @@ meshes using `plotly`.
 The recommended way to install `SLEPLET` is via
 [pip](https://pypi.org/project/pip/)
 
-```bash
+```sh
 pip install sleplet
 ```
 
 To install the latest development version of `SLEPLET` clone this repository
 and run
 
-```bash
+```sh
 pip install -e .
 ```
 
@@ -51,13 +51,15 @@ such it is recommended to stick to the powers of two up to `L=128`.
 When it comes to selecting a Slepian region the order precedence is
 [polar cap region](https://doi.org/10.1111/j.1365-246X.2006.03065.x) >
 [limited latitude longitude region](https://doi.org/10.1109/TSP.2016.2646668) >
-arbitrary region, [as seen in the code](https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/utils/region.py). The
-default region is the `south_america` arbitrary region.
+arbitrary region,
+[as seen in the code](https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/utils/region.py). The default region is the `south_america` arbitrary region.
 
 - `POLAR_GAP`
-  - for a Slepian `polar cap region`, when set in conjunction with `THETA_MAX` but without the other `PHI`/`THETA` variables
+  - for a Slepian `polar cap region`, when set in conjunction with `THETA_MAX`
+    but without the other `PHI`/`THETA` variables
 - `THETA_MAX`
-  - for a Slepian `polar cap region`, when set without the other `PHI`/`THETA` variables
+  - for a Slepian `polar cap region`, when set without the other `PHI`/`THETA`
+    variables
   - for a Slepian `limited latitude longitude region`
 - `THETA_MIN`
   - for a Slepian `limited latitude longitude region`
@@ -72,7 +74,8 @@ default region is the `south_america` arbitrary region.
 
 To recreate the figures from the below papers, one may use the CLI or the API.
 For those which don't use the `mesh` or `sphere` CLIs, the relevant API code
-isn't provided as it is contained within the [examples folder](https://github.com/astro-informatics/sleplet/tree/main/examples).
+isn't provided as it is contained within the
+[examples folder](https://github.com/astro-informatics/sleplet/tree/main/examples).
 
 ### Sifting Convolution on the Sphere
 
@@ -80,7 +83,7 @@ isn't provided as it is contained within the [examples folder](https://github.co
 
 #### Fig. 1
 
-```bash
+```sh
 for ell in $(seq 2 -1 1); do
     sphere harmonic_gaussian -a 0.75 -b 0.125 -e ${ell} 1 -L 128 -m translate -o
 done
@@ -102,7 +105,7 @@ for ell in range(2, 0, -1):
 
 #### Fig. 2
 
-```bash
+```sh
 sphere earth -L 128
 ```
 
@@ -121,7 +124,7 @@ Plot(f_sphere, f.L, "fig_2").execute()
 
 #### Fig. 3
 
-```bash
+```sh
 for ell in $(seq 2 -1 1); do
     sphere harmonic_gaussian -c earth -e ${ell} 1 -L 128
 done
[…]
