+++ .github/workflows/examples.yml
@@ -20,10 +20,10 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install -e .[docs]
+          python -m pip install tox tox-gh-actions
 
-      - name: Run rundoc on the README
-        run: python -m rundoc run README.md
+      - name: Run examples in the README
+        run: tox -e py310-readme
 
   examples:
     runs-on: ubuntu-latest
@@ -43,7 +43,7 @@ jobs:
           python -m pip install --upgrade pip
           python -m pip install -e .[docs]
 
-      - name: Execute all examples in folders
+      - name: Run examples in the examples folder
         run: |
           find examples -name '*.py' \
           -exec echo {} \; \
