+++ pyproject.toml
@@ -68,6 +68,9 @@ optional-dependencies.dev = [
 optional-dependencies.docs = [
     "pdoc3",
 ]
+optional-dependencies.readme = [
+    "pytest-codebase",
+]
 urls.Code = "https://github.com/astro-informatics/sleplet"
 urls.Documentation = "https://astro-informatics.github.io/sleplet"
 urls.Download = "https://pypi.org/project/sleplet"
@@ -91,6 +94,9 @@ plugins = [
 
 [tool.pytest.ini_options]
 addopts = "--color=yes -v"
+testpaths = [
+    "tests",
+]
 
 [tool.ruff]
 fix = true
@@ -154,27 +160,16 @@ overrides."tool.coverage.paths.source".inline_arrays = false
 legacy_tox_ini = """
     [gh-actions]
     python =
-        3.10: py310, py310-readme
+        3.10: py310
 
-    [testenv:py310]
+    [testenv]
     commands =
-        pytest --cov --slow {posargs:tests}
+        pytest --cov --slow
     deps =
         pytest-cov
         pytest-skip-slow
-    description =
-        run unit tests
-
-    [testenv:py310-readme]
-    commands =
-        pytest --codeblocks {posargs:README.md}
-    deps =
-        pytest-codeblocks
-    description =
-        run examples in README.md
 
     [tox]
     env_list =
-        py310{,-readme}
-    isolated_build = True
+        py310
 """
