+++ .github/workflows/examples.yml
@@ -45,6 +45,8 @@ jobs:
 
       - name: Run examples in the examples folder
         run: |
-          find examples -name '*.py' \
-          -exec echo {} \; \
-          -exec python {} \;
+          find examples -name "*.py" \
+          -exec sh -c 'for f; \
+          do echo $f; \
+          python "$f" || exit 1\
+          ; done' sh {} +
