+++ src/sleplet/functions/flm/africa.py
@@ -3,7 +3,7 @@ import pyssht as ssht
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
-from sleplet.data.other.earth.create_earth_flm import create_flm
+from sleplet.data.create_earth_flm import create_flm
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.functions.f_lm import F_LM
 from sleplet.utils.harmonic_methods import ensure_f_bandlimited, rotate_earth_to_africa
