+++ pyproject.toml
@@ -29,7 +29,6 @@ dependencies = [
     "multiprocess>=0.70.14",
     "numba>=0.56.4",
     "numpy>=1.23.5",
-    "pandas>=1.5.3",
     "plotly>=5.13.0",
     "pooch>=1.7.0",
     "pydantic>=1.10.4",
