### 2023-03-28 86b3dd9
Make `vars` private
