+++ src/sleplet/slepian/slepian_region/slepian_polar_cap.py
@@ -13,6 +13,7 @@ from pydantic.dataclasses import dataclass
 from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.slepian.slepian_functions import SlepianFunctions
+from sleplet.utils._vars import NCPU
 from sleplet.utils.harmonic_methods import create_emm_vector
 from sleplet.utils.mask_methods import create_mask_region
 from sleplet.utils.parallel_methods import (
@@ -24,7 +25,6 @@ from sleplet.utils.parallel_methods import (
 )
 from sleplet.utils.region import Region
 from sleplet.utils.validation import Validation
-from sleplet.utils.vars import NCPU
 
 L_SAVE_ALL = 16
 
