+++ src/sleplet/plotting/create_plot_mesh.py
@@ -12,6 +12,11 @@ from pydantic.dataclasses import dataclass
 
 from sleplet import logger
 from sleplet.meshes.classes.mesh import Mesh
+from sleplet.utils._vars import (
+    MESH_CBAR_FONT_SIZE,
+    MESH_CBAR_LEN,
+    MESH_UNSEEN,
+)
 from sleplet.utils.mask_methods import convert_region_on_vertices_to_faces
 from sleplet.utils.mesh_methods import average_functions_on_vertices_to_faces
 from sleplet.utils.plot_methods import convert_colourscale, normalise_function
@@ -21,11 +26,6 @@ from sleplet.utils.plotly_methods import (
     create_tick_mark,
 )
 from sleplet.utils.validation import Validation
-from sleplet.utils.vars import (
-    MESH_CBAR_FONT_SIZE,
-    MESH_CBAR_LEN,
-    MESH_UNSEEN,
-)
 
 _fig_path = Path(__file__).resolve().parents[1] / "figures"
 
