### 2023-03-28 fa9513a
Make _array_methods private
