+++ src/sleplet/functions/flm/ridgelets.py
@@ -8,12 +8,12 @@ from scipy.special import gammaln
 
 from sleplet import logger
 from sleplet.functions.f_lm import F_LM
+from sleplet.utils._validation import Validation
 from sleplet.utils.string_methods import (
     convert_camel_case_to_snake_case,
     filename_args,
     wavelet_ending,
 )
-from sleplet.utils.validation import Validation
 from sleplet.utils.wavelet_methods import create_kappas
 
 
