### 2023-03-28 b9bb0ff
Make `convolution` private
