+++ src/sleplet/utils/wavelet_methods.py
@@ -4,7 +4,7 @@ from numpy import typing as npt
 from pys2let import axisym_wav_l
 
 from sleplet.slepian.slepian_functions import SlepianFunctions
-from sleplet.utils.convolution_methods import sifting_convolution
+from sleplet.utils._convolution_methods import sifting_convolution
 from sleplet.utils.slepian_methods import compute_s_p_omega
 
 
