### 2023-03-28 9ca52b1
Make bool methods private
