+++ src/sleplet/utils/region.py
@@ -1,6 +1,7 @@
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 
+from sleplet.utils._bool_methods import is_limited_lat_lon, is_polar_cap
 from sleplet.utils._validation import Validation
 from sleplet.utils._vars import (
     PHI_MAX_DEFAULT,
@@ -8,7 +9,6 @@ from sleplet.utils._vars import (
     THETA_MAX_DEFAULT,
     THETA_MIN_DEFAULT,
 )
-from sleplet.utils.bool_methods import is_limited_lat_lon, is_polar_cap
 from sleplet.utils.string_methods import angle_as_degree, multiples_of_pi
 
 
