### 2023-03-28 3bfe0d9
Make integration methods private
