+++ src/sleplet/meshes/classes/mesh_slepian.py
@@ -10,9 +10,9 @@ from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.meshes.classes.mesh import Mesh
 from sleplet.utils._array_methods import fill_upper_triangle_of_hermitian_matrix
+from sleplet.utils._integration_methods import integrate_region_mesh
 from sleplet.utils._validation import Validation
 from sleplet.utils._vars import NCPU
-from sleplet.utils.integration_methods import integrate_region_mesh
 from sleplet.utils.parallel_methods import (
     attach_to_shared_memory_block,
     create_shared_memory_array,
