### 2023-03-28 6a8deaf
Make mask methods private
