+++ src/sleplet/functions/flm/africa.py
@@ -6,10 +6,10 @@ from pydantic.dataclasses import dataclass
 from sleplet.data.create_earth_flm import create_flm
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.functions.f_lm import F_LM
+from sleplet.utils._mask_methods import create_mask
 from sleplet.utils._validation import Validation
 from sleplet.utils._vars import SAMPLING_SCHEME
 from sleplet.utils.harmonic_methods import ensure_f_bandlimited, rotate_earth_to_africa
-from sleplet.utils.mask_methods import create_mask
 from sleplet.utils.string_methods import convert_camel_case_to_snake_case
 
 
