+++ src/sleplet/slepian/slepian_region/slepian_polar_cap.py
@@ -13,10 +13,10 @@ from pydantic.dataclasses import dataclass
 from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.slepian.slepian_functions import SlepianFunctions
+from sleplet.utils._mask_methods import create_mask_region
 from sleplet.utils._validation import Validation
 from sleplet.utils._vars import NCPU
 from sleplet.utils.harmonic_methods import create_emm_vector
-from sleplet.utils.mask_methods import create_mask_region
 from sleplet.utils.parallel_methods import (
     attach_to_shared_memory_block,
     create_shared_memory_array,
