+++ src/sleplet/scripts/plotting_on_sphere.py
@@ -8,6 +8,7 @@ from numpy import typing as npt
 from sleplet import logger
 from sleplet.functions.coefficients import Coefficients
 from sleplet.plotting.create_plot_sphere import Plot
+from sleplet.utils._mask_methods import create_default_region
 from sleplet.utils._vars import (
     ALPHA_DEFAULT,
     ANNOTATION_COLOUR,
@@ -20,7 +21,6 @@ from sleplet.utils.harmonic_methods import (
     rotate_earth_to_africa,
     rotate_earth_to_south_america,
 )
-from sleplet.utils.mask_methods import create_default_region
 from sleplet.utils.plot_methods import calc_nearest_grid_point
 from sleplet.utils.slepian_methods import slepian_forward, slepian_inverse
 from sleplet.utils.string_methods import (
