### 2023-03-28 fe1423e
Make mesh methods private
