+++ src/sleplet/meshes/slepian_coefficients/mesh_slepian_noise_field.py
@@ -4,8 +4,8 @@ from pydantic.dataclasses import dataclass
 
 from sleplet.meshes.mesh_slepian_coefficients import MeshSlepianCoefficients
 from sleplet.meshes.slepian_coefficients.mesh_slepian_field import MeshSlepianField
+from sleplet.utils._noise import compute_snr, create_slepian_mesh_noise
 from sleplet.utils._validation import Validation
-from sleplet.utils.noise import compute_snr, create_slepian_mesh_noise
 from sleplet.utils.string_methods import filename_args
 
 
