+++ src/sleplet/utils/denoising.py
@@ -10,9 +10,7 @@ from sleplet.meshes.mesh_coefficients import MeshCoefficients
 from sleplet.meshes.slepian_coefficients.mesh_slepian_wavelets import (
     MeshSlepianWavelets,
 )
-from sleplet.utils._vars import SAMPLING_SCHEME
-from sleplet.utils.harmonic_methods import rotate_earth_to_south_america
-from sleplet.utils.noise import (
+from sleplet.utils._noise import (
     compute_sigma_j,
     compute_sigma_noise,
     compute_slepian_mesh_sigma_j,
@@ -23,6 +21,8 @@ from sleplet.utils.noise import (
     slepian_mesh_hard_thresholding,
     slepian_wavelet_hard_thresholding,
 )
+from sleplet.utils._vars import SAMPLING_SCHEME
+from sleplet.utils.harmonic_methods import rotate_earth_to_south_america
 from sleplet.utils.slepian_methods import slepian_inverse, slepian_mesh_inverse
 from sleplet.utils.wavelet_methods import (
     axisymmetric_wavelet_forward,
