+++ src/sleplet/meshes/harmonic_coefficients/mesh_noise_field.py
@@ -4,8 +4,8 @@ from pydantic.dataclasses import dataclass
 
 from sleplet.meshes.harmonic_coefficients.mesh_field import MeshField
 from sleplet.meshes.mesh_harmonic_coefficients import MeshHarmonicCoefficients
+from sleplet.utils._noise import compute_snr, create_mesh_noise
 from sleplet.utils._validation import Validation
-from sleplet.utils.noise import compute_snr, create_mesh_noise
 from sleplet.utils.string_methods import filename_args
 
 
