+++ src/sleplet/functions/flm/noise_earth.py
@@ -4,8 +4,8 @@ from pydantic.dataclasses import dataclass
 
 from sleplet.functions.f_lm import F_LM
 from sleplet.functions.flm.earth import Earth
+from sleplet.utils._noise import compute_snr, create_noise
 from sleplet.utils._validation import Validation
-from sleplet.utils.noise import compute_snr, create_noise
 from sleplet.utils.string_methods import convert_camel_case_to_snake_case, filename_args
 
 
