### 2023-03-28 cec479b
Make parallel methods private
