+++ src/sleplet/slepian/slepian_region/slepian_polar_cap.py
@@ -14,16 +14,16 @@ from sleplet import logger
 from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.slepian.slepian_functions import SlepianFunctions
 from sleplet.utils._mask_methods import create_mask_region
-from sleplet.utils._validation import Validation
-from sleplet.utils._vars import NCPU
-from sleplet.utils.harmonic_methods import create_emm_vector
-from sleplet.utils.parallel_methods import (
+from sleplet.utils._parallel_methods import (
     attach_to_shared_memory_block,
     create_shared_memory_array,
     free_shared_memory,
     release_shared_memory,
     split_arr_into_chunks,
 )
+from sleplet.utils._validation import Validation
+from sleplet.utils._vars import NCPU
+from sleplet.utils.harmonic_methods import create_emm_vector
 from sleplet.utils.region import Region
 
 L_SAVE_ALL = 16
