+++ src/sleplet/meshes/classes/mesh_slepian.py
@@ -11,15 +11,15 @@ from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.meshes.classes.mesh import Mesh
 from sleplet.utils._array_methods import fill_upper_triangle_of_hermitian_matrix
 from sleplet.utils._integration_methods import integrate_region_mesh
-from sleplet.utils._validation import Validation
-from sleplet.utils._vars import NCPU
-from sleplet.utils.parallel_methods import (
+from sleplet.utils._parallel_methods import (
     attach_to_shared_memory_block,
     create_shared_memory_array,
     free_shared_memory,
     release_shared_memory,
     split_arr_into_chunks,
 )
+from sleplet.utils._validation import Validation
+from sleplet.utils._vars import NCPU
 from sleplet.utils.slepian_arbitrary_methods import compute_mesh_shannon
 
 _data_path = Path(__file__).resolve().parents[2] / "data"
