### 2023-03-28 1d5dba3
Make plotly methods private
