### 2023-03-28 576253a
Make slepian arbitrary methods private
