### 2023-03-28 7e91e0e
Make smoothing private
