### 2023-03-28 7bb4ddd
Make class lists private
