+++ examples/arbitrary/africa/eigenvalues_africa.py
@@ -7,7 +7,7 @@ from matplotlib import pyplot as plt
 from sleplet.slepian.slepian_region.slepian_arbitrary import SlepianArbitrary
 from sleplet.utils.plot_methods import save_plot
 
-fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "figures"
+_fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "figures"
 sns.set(context="paper")
 
 L = 128
@@ -33,7 +33,7 @@ def main() -> None:
     plt.xticks(ticks, ticks)
     plt.xlabel(r"$p$")
     plt.ylabel(r"$\mu$")
-    save_plot(fig_path, f"africa_eigenvalues_L{L}")
+    save_plot(_fig_path, f"africa_eigenvalues_L{L}")
 
 
 if __name__ == "__main__":
