+++ examples/polar_cap/simons_5_3.py
@@ -9,14 +9,14 @@ from sleplet import logger
 from sleplet.slepian.slepian_region.slepian_polar_cap import SlepianPolarCap
 from sleplet.utils.plot_methods import save_plot
 
+_fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "figures"
+sns.set(context="paper")
+
 L = 16
 LEGEND_POS = (0, 0)
 RANKS = 60
 TEXT_BOX: dict[str, str | float] = {"boxstyle": "round", "color": "w"}
 THETA_RANGE = {10: (0, 0), 20: (0, 1), 30: (1, 0), 40: (1, 1)}
-
-fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "figures"
-sns.set(context="paper")
 
 
 def main() -> None:
@@ -28,7 +28,7 @@ def main() -> None:
     for theta_max, position in THETA_RANGE.items():
         _create_plot(ax, position, theta_max)
     ax[LEGEND_POS].legend(ncol=2)
-    save_plot(fig_path, "simons_5_3")
+    save_plot(_fig_path, "simons_5_3")
 
 
 def _create_plot(ax: npt.NDArray, position: tuple[int, int], theta_max: int) -> None:
