### 2023-03-28 9577a50
Remove unncessary vars
