### 2023-03-28 8879c63
Make harmonic methods private
