+++ src/sleplet/functions/flm/south_america.py
@@ -10,7 +10,7 @@ from sleplet.utils._mask_methods import create_mask
 from sleplet.utils._validation import Validation
 from sleplet.utils._vars import SAMPLING_SCHEME
 from sleplet.utils.harmonic_methods import (
-    ensure_f_bandlimited,
+    _ensure_f_bandlimited,
     rotate_earth_to_south_america,
 )
 from sleplet.utils.string_methods import convert_camel_case_to_snake_case
@@ -22,7 +22,7 @@ class SouthAmerica(F_LM):
         super().__post_init_post_parse__()
 
     def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
-        return ensure_f_bandlimited(
+        return _ensure_f_bandlimited(
             self._grid_fun, self.L, reality=self.reality, spin=self.spin
         )
 
