### 2023-03-28 42ca5bc
Make plot methods private
