+++ tests/test_plot.py
@@ -3,9 +3,9 @@ import numpy as np
 from numpy.testing import assert_equal
 
 from sleplet.utils.plot_methods import (
-    calc_nearest_grid_point,
+    _calc_nearest_grid_point,
+    _convert_colourscale,
     calc_plot_resolution,
-    convert_colourscale,
 )
 
 L = 128
@@ -28,7 +28,7 @@ def test_create_colourscale() -> None:
     """
     test creates a plotly compatible colourscale
     """
-    colourscale = convert_colourscale(cmocean.cm.ice, pl_entries=PL_ENTRIES)
+    colourscale = _convert_colourscale(cmocean.cm.ice, pl_entries=PL_ENTRIES)
     assert_equal(len(colourscale), PL_ENTRIES)
 
 
@@ -36,6 +36,6 @@ def test_find_nearest_grid_point() -> None:
     """
     test to find nearest grid point to provided angles
     """
-    alpha, beta = calc_nearest_grid_point(L, PHI_0 / np.pi, THETA_MAX / np.pi)
+    alpha, beta = _calc_nearest_grid_point(L, PHI_0 / np.pi, THETA_MAX / np.pi)
     assert_equal(alpha, 0.5154175447295755)
     assert_equal(beta, 1.055378782065321)
