### 2023-03-28 9342b9f
Make Slepian methods private
