### 2023-03-28 ee0243f
Make string methods private
