+++ src/sleplet/functions/flm/wmap.py
@@ -5,7 +5,7 @@ from pydantic.dataclasses import dataclass
 from sleplet.data.create_wmap_flm import create_flm
 from sleplet.functions.f_lm import F_LM
 from sleplet.utils._validation import Validation
-from sleplet.utils.string_methods import convert_camel_case_to_snake_case
+from sleplet.utils.string_methods import _convert_camel_case_to_snake_case
 
 
 @dataclass(config=Validation)
@@ -17,7 +17,7 @@ class Wmap(F_LM):
         return create_flm(self.L)
 
     def _create_name(self) -> str:
-        return convert_camel_case_to_snake_case(self.__class__.__name__)
+        return _convert_camel_case_to_snake_case(self.__class__.__name__)
 
     def _set_reality(self) -> bool:
         return True
