+++ src/sleplet/functions/flm/axisymmetric_wavelets.py
@@ -8,9 +8,9 @@ from sleplet import logger
 from sleplet.functions.f_lm import F_LM
 from sleplet.utils._validation import Validation
 from sleplet.utils.string_methods import (
-    convert_camel_case_to_snake_case,
+    _convert_camel_case_to_snake_case,
+    _wavelet_ending,
     filename_args,
-    wavelet_ending,
 )
 from sleplet.utils.wavelet_methods import create_axisymmetric_wavelets
 
@@ -33,10 +33,10 @@ class AxisymmetricWavelets(F_LM):
 
     def _create_name(self) -> str:
         return (
-            f"{convert_camel_case_to_snake_case(self.__class__.__name__)}"
+            f"{_convert_camel_case_to_snake_case(self.__class__.__name__)}"
             f"{filename_args(self.B, 'B')}"
             f"{filename_args(self.j_min, 'jmin')}"
-            f"{wavelet_ending(self.j_min, self.j)}"
+            f"{_wavelet_ending(self.j_min, self.j)}"
         )
 
     def _set_reality(self) -> bool:
