+++ src/sleplet/functions/fp/slepian_noise_africa.py
@@ -7,7 +7,10 @@ from sleplet.functions.fp.slepian_africa import SlepianAfrica
 from sleplet.utils._noise import compute_snr, create_slepian_noise
 from sleplet.utils._validation import Validation
 from sleplet.utils.region import Region
-from sleplet.utils.string_methods import convert_camel_case_to_snake_case, filename_args
+from sleplet.utils.string_methods import (
+    _convert_camel_case_to_snake_case,
+    filename_args,
+)
 
 
 @dataclass(config=Validation, kw_only=True)
@@ -27,7 +30,7 @@ class SlepianNoiseAfrica(F_P):
 
     def _create_name(self) -> str:
         return (
-            f"{convert_camel_case_to_snake_case(self.__class__.__name__)}"
+            f"{_convert_camel_case_to_snake_case(self.__class__.__name__)}"
             f"{filename_args(self.SNR, 'snr')}"
         )
 
