### 2023-03-28 43d4c17
Make wavelet methods private
