+++ src/sleplet/functions/flm/axisymmetric_wavelet_coefficients_south_america.py
@@ -14,8 +14,8 @@ from sleplet.utils.string_methods import (
     filename_args,
 )
 from sleplet.utils.wavelet_methods import (
+    _create_axisymmetric_wavelets,
     axisymmetric_wavelet_forward,
-    create_axisymmetric_wavelets,
 )
 
 
@@ -62,7 +62,7 @@ class AxisymmetricWaveletCoefficientsSouthAmerica(F_LM):
         """
         computes wavelet coefficients of South America
         """
-        wavelets = create_axisymmetric_wavelets(self.L, self.B, self.j_min)
+        wavelets = _create_axisymmetric_wavelets(self.L, self.B, self.j_min)
         self.south_america = SouthAmerica(self.L, smoothing=self.smoothing)
         wavelet_coefficients = axisymmetric_wavelet_forward(
             self.L, self.south_america.coefficients, wavelets
