### 2023-03-28 3a2e308
Make Slepian decomposition private
