+++ src/sleplet/functions/fp/slepian_wavelet_coefficients_africa.py
@@ -72,7 +72,7 @@ class SlepianWaveletCoefficientsAfrica(F_P):
         return wavelets, wavelet_coefficients
 
     @validator("j")
-    def check_j(cls, v, values):
+    def _check_j(cls, v, values):
         j_max = pys2let_j_max(values["B"], values["L"] ** 2, values["j_min"])
         if v is not None and v < 0:
             raise ValueError("j should be positive")
