+++ src/sleplet/functions/flm/axisymmetric_wavelet_coefficients_earth.py
@@ -70,7 +70,7 @@ class AxisymmetricWaveletCoefficientsEarth(F_LM):
         return wavelets, wavelet_coefficients
 
     @validator("j")
-    def check_j(cls, v, values):
+    def _check_j(cls, v, values):
         j_max = pys2let_j_max(values["B"], values["L"], values["j_min"])
         if v is not None and v < 0:
             raise ValueError("j should be positive")
