+++ src/sleplet/functions/flm/axisymmetric_wavelets.py
@@ -59,7 +59,7 @@ class AxisymmetricWavelets(F_LM):
         return _create_axisymmetric_wavelets(self.L, self.B, self.j_min)
 
     @validator("j")
-    def check_j(cls, v, values):
+    def _check_j(cls, v, values):
         j_max = pys2let_j_max(values["B"], values["L"], values["j_min"])
         if v is not None and v < 0:
             raise ValueError("j should be positive")
