+++ src/sleplet/meshes/slepian_coefficients/mesh_slepian_functions.py
@@ -59,7 +59,7 @@ class MeshSlepianFunctions(MeshSlepianCoefficients):
                 raise ValueError(f"rank should be less than or equal to {limit}")
 
     @validator("rank")
-    def check_rank(cls, v):
+    def _check_rank(cls, v):
         if not isinstance(v, int):
             raise TypeError("rank should be an integer")
         if v < 0:
