+++ src/sleplet/utils/slepian_methods.py
@@ -3,8 +3,8 @@ import pyssht as ssht
 from numpy import typing as npt
 
 from sleplet import logger
+from sleplet.meshes.classes._mesh_slepian_decomposition import MeshSlepianDecomposition
 from sleplet.meshes.classes.mesh_slepian import MeshSlepian
-from sleplet.meshes.classes.mesh_slepian_decomposition import MeshSlepianDecomposition
 from sleplet.slepian._slepian_decomposition import SlepianDecomposition
 from sleplet.slepian.slepian_functions import SlepianFunctions
 from sleplet.slepian.slepian_region.slepian_arbitrary import SlepianArbitrary
