### 2023-03-29 6a0e37d
Move amplitude plots
