### 2023-03-29 c970386
Make scripts private
