+++ src/sleplet/functions/fp/slepian_noise_africa.py
@@ -2,12 +2,12 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+from sleplet._noise import compute_snr, create_slepian_noise
+from sleplet._validation import Validation
 from sleplet.functions.f_p import F_P
 from sleplet.functions.fp.slepian_africa import SlepianAfrica
-from sleplet.utils._noise import compute_snr, create_slepian_noise
-from sleplet.utils._validation import Validation
-from sleplet.utils.region import Region
-from sleplet.utils.string_methods import (
+from sleplet.region import Region
+from sleplet.string_methods import (
     _convert_camel_case_to_snake_case,
     filename_args,
 )
