+++ src/sleplet/functions/flm/gaussian.py
@@ -3,9 +3,9 @@ import pyssht as ssht
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+from sleplet._validation import Validation
 from sleplet.functions.f_lm import F_LM
-from sleplet.utils._validation import Validation
-from sleplet.utils.string_methods import (
+from sleplet.string_methods import (
     _convert_camel_case_to_snake_case,
     filename_args,
 )
