+++ tests/test_translation.py
@@ -4,10 +4,10 @@ from hypothesis import given, seed, settings
 from hypothesis.strategies import SearchStrategy, floats
 from numpy.testing import assert_allclose, assert_equal, assert_raises
 
+from sleplet._vars import RANDOM_SEED, SAMPLING_SCHEME
 from sleplet.functions.flm.dirac_delta import DiracDelta
-from sleplet.utils._vars import RANDOM_SEED, SAMPLING_SCHEME
-from sleplet.utils.plot_methods import _calc_nearest_grid_point
-from sleplet.utils.slepian_methods import slepian_inverse
+from sleplet.plot_methods import _calc_nearest_grid_point
+from sleplet.slepian_methods import slepian_inverse
 
 L = 128
 THETA_MAX = np.pi / 3
