+++ examples/arbitrary/africa/sparsity_wavelet_slepian_comparison_africa.py
@@ -11,8 +11,8 @@ from sleplet.functions.flm.axisymmetric_wavelet_coefficients_africa import (
 from sleplet.functions.fp.slepian_wavelet_coefficients_africa import (
     SlepianWaveletCoefficientsAfrica,
 )
-from sleplet.utils.plot_methods import save_plot
-from sleplet.utils.region import Region
+from sleplet.plot_methods import save_plot
+from sleplet.region import Region
 
 _fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "figures"
 sns.set(context="paper")
