+++ src/sleplet/harmonic_methods.py
@@ -6,11 +6,11 @@ import pyssht as ssht
 from numpy import typing as npt
 from numpy.random import Generator
 
-from sleplet.meshes.classes.mesh import Mesh
-from sleplet.utils._integration_methods import integrate_whole_mesh
-from sleplet.utils._vars import (
+from sleplet._integration_methods import integrate_whole_mesh
+from sleplet._vars import (
     SAMPLING_SCHEME,
 )
+from sleplet.meshes.classes.mesh import Mesh
 
 AFRICA_ALPHA = np.deg2rad(44)
 AFRICA_BETA = np.deg2rad(87)
