### 2023-03-29 970267d
Remove string method private
