+++ src/sleplet/functions/fp/slepian_africa.py
@@ -2,12 +2,12 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+from sleplet._string_methods import _convert_camel_case_to_snake_case
 from sleplet._validation import Validation
 from sleplet.functions.f_p import F_P
 from sleplet.functions.flm.africa import Africa
 from sleplet.region import Region
 from sleplet.slepian_methods import slepian_forward
-from sleplet.string_methods import _convert_camel_case_to_snake_case
 
 
 @dataclass(config=Validation)
