+++ src/sleplet/functions/flm/earth.py
@@ -2,10 +2,10 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+from sleplet._string_methods import _convert_camel_case_to_snake_case
 from sleplet._validation import Validation
 from sleplet.data.create_earth_flm import create_flm
 from sleplet.functions.f_lm import F_LM
-from sleplet.string_methods import _convert_camel_case_to_snake_case
 
 
 @dataclass(config=Validation)
