+++ src/sleplet/functions/flm/noise_earth.py
@@ -3,13 +3,13 @@ from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 from sleplet._noise import compute_snr, create_noise
-from sleplet._validation import Validation
-from sleplet.functions.f_lm import F_LM
-from sleplet.functions.flm.earth import Earth
-from sleplet.string_methods import (
+from sleplet._string_methods import (
     _convert_camel_case_to_snake_case,
     filename_args,
 )
+from sleplet._validation import Validation
+from sleplet.functions.f_lm import F_LM
+from sleplet.functions.flm.earth import Earth
 
 
 @dataclass(config=Validation, kw_only=True)
