+++ src/sleplet/functions/flm/axisymmetric_wavelet_coefficients_africa.py
@@ -5,14 +5,14 @@ from pydantic.dataclasses import dataclass
 from pys2let import pys2let_j_max
 
 from sleplet import logger
-from sleplet._validation import Validation
-from sleplet.functions.f_lm import F_LM
-from sleplet.functions.flm.africa import Africa
-from sleplet.string_methods import (
+from sleplet._string_methods import (
     _convert_camel_case_to_snake_case,
     _wavelet_ending,
     filename_args,
 )
+from sleplet._validation import Validation
+from sleplet.functions.f_lm import F_LM
+from sleplet.functions.flm.africa import Africa
 from sleplet.wavelet_methods import (
     _create_axisymmetric_wavelets,
     axisymmetric_wavelet_forward,
