+++ src/sleplet/_scripts/plotting_on_mesh.py
@@ -3,6 +3,7 @@ from argparse import ArgumentParser, Namespace
 
 from sleplet import logger
 from sleplet._class_lists import MESH_COEFFICIENTS, MESHES
+from sleplet._string_methods import _convert_classes_list_to_snake_case
 from sleplet.meshes.classes.mesh import Mesh
 from sleplet.meshes.mesh_coefficients import MeshCoefficients
 from sleplet.meshes.mesh_slepian_coefficients import MeshSlepianCoefficients
@@ -11,7 +12,6 @@ from sleplet.plot_methods import (
     compute_amplitude_for_noisy_mesh_plots,
 )
 from sleplet.plotting.create_plot_mesh import Plot
-from sleplet.string_methods import _convert_classes_list_to_snake_case
 
 
 def valid_meshes(mesh_name: str) -> str:
