+++ src/sleplet/functions/flm/elongated_gaussian.py
@@ -2,14 +2,14 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+from sleplet._string_methods import (
+    _convert_camel_case_to_snake_case,
+    filename_args,
+)
 from sleplet._validation import Validation
 from sleplet._vars import PHI_0, THETA_0
 from sleplet.functions.f_lm import F_LM
 from sleplet.harmonic_methods import _ensure_f_bandlimited
-from sleplet.string_methods import (
-    _convert_camel_case_to_snake_case,
-    filename_args,
-)
 
 
 @dataclass(config=Validation, kw_only=True)
