+++ src/sleplet/functions/flm/directional_spin_wavelets.py
@@ -8,8 +8,8 @@ from pys2let import pys2let_j_max, wavelet_tiling
 from sleplet import logger
 from sleplet._string_methods import (
     _convert_camel_case_to_snake_case,
-    _wavelet_ending,
     filename_args,
+    wavelet_ending,
 )
 from sleplet._validation import Validation
 from sleplet.functions.f_lm import F_LM
@@ -40,7 +40,7 @@ class DirectionalSpinWavelets(F_LM):
             f"{filename_args(self.j_min, 'jmin')}"
             f"{filename_args(self.spin, 'spin')}"
             f"{filename_args(self.N, 'N')}"
-            f"{_wavelet_ending(self.j_min, self.j)}"
+            f"{wavelet_ending(self.j_min, self.j)}"
         )
 
     def _set_reality(self) -> bool:
