+++ src/sleplet/functions/fp/slepian_wavelet_coefficients_africa.py
@@ -7,8 +7,8 @@ from pys2let import pys2let_j_max
 from sleplet import logger
 from sleplet._string_methods import (
     _convert_camel_case_to_snake_case,
-    _wavelet_ending,
     filename_args,
+    wavelet_ending,
 )
 from sleplet._validation import Validation
 from sleplet.functions.f_p import F_P
@@ -41,7 +41,7 @@ class SlepianWaveletCoefficientsAfrica(F_P):
             f"{_convert_camel_case_to_snake_case(self.__class__.__name__)}"
             f"{filename_args(self.B, 'B')}"
             f"{filename_args(self.j_min, 'jmin')}"
-            f"{_wavelet_ending(self.j_min, self.j)}"
+            f"{wavelet_ending(self.j_min, self.j)}"
         )
 
     def _set_reality(self) -> bool:
