+++ src/sleplet/functions/flm/wmap.py
@@ -2,9 +2,9 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+from sleplet._data.create_wmap_flm import create_flm
 from sleplet._string_methods import _convert_camel_case_to_snake_case
 from sleplet._validation import Validation
-from sleplet.data.create_wmap_flm import create_flm
 from sleplet.functions.f_lm import F_LM
 
 
