+++ src/sleplet/meshes/classes/mesh_slepian.py
@@ -8,6 +8,7 @@ from pydantic.dataclasses import dataclass
 
 from sleplet import logger
 from sleplet._array_methods import fill_upper_triangle_of_hermitian_matrix
+from sleplet._data.setup_pooch import find_on_pooch_then_local
 from sleplet._integration_methods import integrate_region_mesh
 from sleplet._parallel_methods import (
     attach_to_shared_memory_block,
@@ -19,10 +20,9 @@ from sleplet._parallel_methods import (
 from sleplet._slepian_arbitrary_methods import compute_mesh_shannon
 from sleplet._validation import Validation
 from sleplet._vars import NCPU
-from sleplet.data.setup_pooch import find_on_pooch_then_local
 from sleplet.meshes.classes.mesh import Mesh
 
-_data_path = Path(__file__).resolve().parents[2] / "data"
+_data_path = Path(__file__).resolve().parents[2] / "_data"
 
 
 @dataclass(config=Validation)
