+++ src/sleplet/_data/setup_pooch.py
@@ -10,7 +10,6 @@ _data_path = Path(__file__).resolve().parent
 ZENODO_DATA_DOI = "10.5281/zenodo.7767698"
 POOCH = pooch.create(
     path=pooch.os_cache("sleplet"),
-    # Use the figshare DOI
     base_url=f"doi:{ZENODO_DATA_DOI}/",
     registry=None,
 )
