+++ src/sleplet/_data/create_earth_flm.py
@@ -3,8 +3,8 @@ import pyssht as ssht
 from numpy import typing as npt
 from scipy import io as sio
 
+from sleplet._data.setup_pooch import find_on_pooch_then_local
 from sleplet._smoothing import apply_gaussian_smoothing
-from sleplet.data.setup_pooch import find_on_pooch_then_local
 
 
 def create_flm(L: int, *, smoothing: int | None = None) -> npt.NDArray[np.complex_]:
