### 2023-03-29 b835780
Make Slepian functions private
