+++ src/sleplet/slepian_methods.py
@@ -5,11 +5,11 @@ from numpy import typing as npt
 from sleplet import logger
 from sleplet._vars import SAMPLING_SCHEME
 from sleplet.harmonic_methods import _boost_coefficient_resolution, mesh_inverse
-from sleplet.meshes.classes._mesh_slepian_decomposition import MeshSlepianDecomposition
-from sleplet.meshes.classes.mesh_slepian import MeshSlepian
+from sleplet.meshes._mesh_slepian_decomposition import MeshSlepianDecomposition
+from sleplet.meshes.mesh_slepian import MeshSlepian
 from sleplet.region import Region
 from sleplet.slepian._slepian_decomposition import SlepianDecomposition
-from sleplet.slepian.slepian_functions import SlepianFunctions
+from sleplet.slepian._slepian_functions import SlepianFunctions
 from sleplet.slepian.slepian_region.slepian_arbitrary import SlepianArbitrary
 from sleplet.slepian.slepian_region.slepian_limit_lat_lon import SlepianLimitLatLon
 from sleplet.slepian.slepian_region.slepian_polar_cap import SlepianPolarCap
