### 2023-03-29 2a138fa
Remove slepian region folder
