+++ examples/arbitrary/eigenvalues_combined.py
@@ -5,7 +5,7 @@ import seaborn as sns
 from matplotlib import pyplot as plt
 
 from sleplet.plot_methods import save_plot
-from sleplet.slepian.slepian_region.slepian_arbitrary import SlepianArbitrary
+from sleplet.slepian.slepian_arbitrary import SlepianArbitrary
 
 _fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "figures"
 sns.set(context="paper")
