+++ examples/polar_cap/slepian_coefficients.py
@@ -8,7 +8,7 @@ from numpy import typing as npt
 from sleplet.functions.flm.earth import Earth
 from sleplet.plot_methods import save_plot
 from sleplet.region import Region
-from sleplet.slepian.slepian_region.slepian_polar_cap import SlepianPolarCap
+from sleplet.slepian.slepian_polar_cap import SlepianPolarCap
 from sleplet.slepian_methods import choose_slepian_method, slepian_forward
 
 _fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "figures"
