+++ src/sleplet/slepian/slepian_arbitrary.py
@@ -33,7 +33,7 @@ from sleplet.harmonic_methods import (
 from sleplet.region import Region
 from sleplet.slepian._slepian_functions import SlepianFunctions
 
-_data_path = Path(__file__).resolve().parents[2] / "_data"
+_data_path = Path(__file__).resolve().parents[1] / "_data"
 
 SAMPLES = 2
 
