### 2023-03-29 8202d02
Undo Slepian functions
