+++ src/sleplet/slepian_methods.py
@@ -9,8 +9,8 @@ from sleplet.meshes._mesh_slepian_decomposition import MeshSlepianDecomposition
 from sleplet.meshes.mesh_slepian import MeshSlepian
 from sleplet.region import Region
 from sleplet.slepian._slepian_decomposition import SlepianDecomposition
-from sleplet.slepian._slepian_functions import SlepianFunctions
 from sleplet.slepian.slepian_arbitrary import SlepianArbitrary
+from sleplet.slepian.slepian_functions import SlepianFunctions
 from sleplet.slepian.slepian_limit_lat_lon import SlepianLimitLatLon
 from sleplet.slepian.slepian_polar_cap import SlepianPolarCap
 
