+++ src/sleplet/_noise.py
@@ -7,7 +7,7 @@ from sleplet import logger
 from sleplet._vars import RANDOM_SEED, SAMPLING_SCHEME
 from sleplet.harmonic_methods import mesh_forward
 from sleplet.meshes.mesh_slepian import MeshSlepian
-from sleplet.slepian._slepian_functions import SlepianFunctions
+from sleplet.slepian.slepian_functions import SlepianFunctions
 from sleplet.slepian_methods import (
     _compute_mesh_s_p_pixel,
     _compute_s_p_omega,
