### 2023-03-29 5bd97da
Make figure directory private
