+++ examples/arbitrary/africa/tiling_africa.py
@@ -9,7 +9,7 @@ from sleplet.plot_methods import save_plot
 from sleplet.slepian.slepian_arbitrary import SlepianArbitrary
 from sleplet.wavelet_methods import create_kappas
 
-_fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "figures"
+_fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "_figures"
 sns.set(context="paper")
 
 B = 3
