+++ examples/polar_cap/simons_5_3.py
@@ -9,7 +9,7 @@ from sleplet import logger
 from sleplet.plot_methods import save_plot
 from sleplet.slepian.slepian_polar_cap import SlepianPolarCap
 
-_fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "figures"
+_fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "_figures"
 sns.set(context="paper")
 
 L = 16
