+++ examples/mesh/mesh_slepian_eigenvalues.py
@@ -9,7 +9,7 @@ from sleplet.meshes.mesh import Mesh
 from sleplet.meshes.mesh_slepian import MeshSlepian
 from sleplet.plot_methods import save_plot
 
-_fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "figures"
+_fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "_figures"
 sns.set(context="paper")
 
 MESHES = [
