+++ examples/mesh/mesh_tiling.py
@@ -11,7 +11,7 @@ from sleplet.meshes.mesh_slepian import MeshSlepian
 from sleplet.plot_methods import save_plot
 from sleplet.wavelet_methods import create_kappas
 
-_fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "figures"
+_fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "_figures"
 sns.set(context="paper")
 
 B = 3
