+++ src/sleplet/plotting/__init__.py
@@ -1,0 +1,3 @@
+"""
+classes to create plots on the sphere or mesh
+"""
