+++ src/sleplet/harmonic_methods.py
@@ -1,3 +1,6 @@
+"""
+methods to perform operations in Fourier space of the sphere or mesh
+"""
 from collections.abc import Callable
 from typing import Any
 
