+++ src/sleplet/functions/__init__.py
@@ -1,0 +1,3 @@
+"""
+classes to create and handle spherical data
+"""
