+++ src/sleplet/plot_methods.py
@@ -1,3 +1,6 @@
+"""
+methods to help in creating plots
+"""
 from pathlib import Path
 
 import numpy as np
