+++ src/sleplet/slepian/__init__.py
@@ -1,0 +1,3 @@
+"""
+classes to create the Slepian regions on the sphere
+"""
