+++ src/sleplet/slepian_methods.py
@@ -1,3 +1,6 @@
+"""
+methods to work with Slepian coefficients
+"""
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
