+++ src/sleplet/meshes/__init__.py
@@ -1,0 +1,3 @@
+"""
+classes to create and handle mesh (manifold) data
+"""
