### 2023-03-29 c35e34c
Make denoising private and noise public
