+++ src/sleplet/functions/fp/__init__.py
@@ -1,0 +1,3 @@
+"""
+classes of functions on the sphere created in Slepian space
+"""
