+++ src/sleplet/functions/coefficients.py
@@ -1,3 +1,6 @@
+"""
+parent class to deal with harmonic/Slepian coefficients on the sphere
+"""
 from abc import abstractmethod
 from dataclasses import KW_ONLY
 
