+++ src/sleplet/functions/flm/__init__.py
@@ -1,0 +1,3 @@
+"""
+classes of functions on the sphere created in harmonic space
+"""
