+++ src/sleplet/functions/f_p.py
@@ -1,3 +1,6 @@
+"""
+parent class to deal with Slepian coefficients on the sphere
+"""
 from abc import abstractmethod
 
 import numpy as np
