+++ src/sleplet/meshes/harmonic_coefficients/__init__.py
@@ -1,0 +1,3 @@
+"""
+classes of functions on the mesh created in Fourier space
+"""
