+++ src/sleplet/meshes/mesh_harmonic_coefficients.py
@@ -1,3 +1,6 @@
+"""
+parent class to handle Fourier coefficients on the mesh
+"""
 from abc import abstractmethod
 
 import numpy as np
