+++ src/sleplet/slepian/slepian_arbitrary.py
@@ -11,7 +11,7 @@ from numpy import linalg as LA  # noqa: N812
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
-from sleplet import logger
+from sleplet import NCPU, logger
 from sleplet._array_methods import fill_upper_triangle_of_hermitian_matrix
 from sleplet._data.setup_pooch import find_on_pooch_then_local
 from sleplet._integration_methods import (
@@ -28,7 +28,6 @@ from sleplet._parallel_methods import (
 )
 from sleplet._slepian_arbitrary_methods import clean_evals_and_evecs
 from sleplet._validation import Validation
-from sleplet._vars import NCPU
 from sleplet.harmonic_methods import (
     _create_spherical_harmonic,
     invert_flm_boosted,
