### 2023-03-29 9e7cc18
Move Slepian coefficients to init file
