+++ examples/arbitrary/south_america/denoising_slepian_south_america.py
@@ -3,8 +3,7 @@ from argparse import ArgumentParser
 from pathlib import Path
 
 from sleplet import logger
-from sleplet.functions.fp.slepian_south_america import SlepianSouthAmerica
-from sleplet.functions.fp.slepian_wavelets import SlepianWavelets
+from sleplet.functions.fp import SlepianSouthAmerica, SlepianWavelets
 from sleplet.plot_methods import find_max_amplitude
 from sleplet.plotting.create_plot_sphere import Plot
 from sleplet.region import Region
