+++ examples/arbitrary/africa/progressive_wavelet_reconstruction_africa.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from sleplet import logger
-from sleplet.functions.fp.slepian_wavelet_coefficients_africa import (
+from sleplet.functions.fp import (
     SlepianWaveletCoefficientsAfrica,
 )
 from sleplet.plotting.create_plot_sphere import Plot
