+++ examples/arbitrary/africa/sparsity_wavelet_slepian_comparison_africa.py
@@ -5,10 +5,10 @@ import numpy as np
 import seaborn as sns
 from matplotlib import pyplot as plt
 
-from sleplet.functions.flm.axisymmetric_wavelet_coefficients_africa import (
+from sleplet.functions.flm import (
     AxisymmetricWaveletCoefficientsAfrica,
 )
-from sleplet.functions.fp.slepian_wavelet_coefficients_africa import (
+from sleplet.functions.fp import (
     SlepianWaveletCoefficientsAfrica,
 )
 from sleplet.plot_methods import save_plot
