+++ examples/arbitrary/south_america/wavelet_reconstruction_south_america.py
@@ -1,4 +1,4 @@
-from sleplet.functions.fp.slepian_wavelet_coefficients_south_america import (
+from sleplet.functions.fp import (
     SlepianWaveletCoefficientsSouthAmerica,
 )
 from sleplet.plotting.create_plot_sphere import Plot
