+++ examples/arbitrary/africa/denoising_slepian_africa.py
@@ -3,8 +3,7 @@ from argparse import ArgumentParser
 from pathlib import Path
 
 from sleplet import logger
-from sleplet.functions.fp.slepian_africa import SlepianAfrica
-from sleplet.functions.fp.slepian_wavelets import SlepianWavelets
+from sleplet.functions.fp import SlepianAfrica, SlepianWavelets
 from sleplet.plot_methods import find_max_amplitude
 from sleplet.plotting.create_plot_sphere import Plot
 from sleplet.region import Region
