### 2023-03-29 b1cec33
Mesh function init
