+++ src/sleplet/meshes/slepian_coefficients/__init__.py
@@ -1,3 +1,9 @@
 """
 classes of functions on the sphere created in Fourier space
 """
+
+from .mesh_slepian_field import MeshSlepianField
+from .mesh_slepian_functions import MeshSlepianFunctions
+from .mesh_slepian_noise_field import MeshSlepianNoiseField
+from .mesh_slepian_wavelet_coefficients import MeshSlepianWaveletCoefficients
+from .mesh_slepian_wavelets import MeshSlepianWavelets
