+++ examples/mesh/denoising_slepian_mesh.py
@@ -6,8 +6,8 @@ from numpy import typing as npt
 from sleplet import logger
 from sleplet.meshes.mesh import Mesh
 from sleplet.meshes.mesh_coefficients import MeshCoefficients
-from sleplet.meshes.slepian_coefficients.mesh_slepian_field import MeshSlepianField
-from sleplet.meshes.slepian_coefficients.mesh_slepian_wavelets import (
+from sleplet.meshes.slepian_coefficients import (
+    MeshSlepianField,
     MeshSlepianWavelets,
 )
 from sleplet.noise import (
