+++ README.md
@@ -459,7 +459,7 @@ done
 ```python
 from sleplet.harmonic_methods import mesh_inverse
 from sleplet.meshes.mesh import Mesh
-from sleplet.meshes.harmonic_coefficients.mesh_basis_functions import MeshBasisFunctions
+from sleplet.meshes.harmonic_coefficients import MeshBasisFunctions
 from sleplet.plotting.create_plot_mesh import Plot
 
 mesh = Mesh("homer")
@@ -491,9 +491,7 @@ done
 
 ```python
 from sleplet.meshes.mesh import Mesh
-from sleplet.meshes.slepian_coefficients.mesh_slepian_functions import (
-    MeshSlepianFunctions,
-)
+from sleplet.meshes.slepian_coefficients import MeshSlepianFunctions
 from sleplet.plotting.create_plot_mesh import Plot
 from sleplet.slepian_methods import slepian_mesh_inverse
 
@@ -523,9 +521,7 @@ done
 
 ```python
 from sleplet.meshes.mesh import Mesh
-from sleplet.meshes.slepian_coefficients.mesh_slepian_wavelets import (
-    MeshSlepianWavelets,
-)
+from sleplet.meshes.slepian_coefficients import MeshSlepianWavelets
 from sleplet.plotting.create_plot_mesh import Plot
 from sleplet.slepian_methods import slepian_mesh_inverse
 
@@ -545,7 +541,7 @@ mesh homer -m field -u
 ```python
 from sleplet.harmonic_methods import mesh_inverse
 from sleplet.meshes.mesh import Mesh
-from sleplet.meshes.harmonic_coefficients.mesh_field import MeshField
+from sleplet.meshes.harmonic_coefficients import MeshField
 from sleplet.plotting.create_plot_mesh import Plot
 
 mesh = Mesh("homer")
@@ -567,9 +563,7 @@ done
 
 ```python
 from sleplet.meshes.mesh import Mesh
-from sleplet.meshes.slepian_coefficients.mesh_slepian_wavelet_coefficients import (
-    MeshSlepianWaveletCoefficients,
-)
+from sleplet.meshes.slepian_coefficients import MeshSlepianWaveletCoefficients
 from sleplet.plotting.create_plot_mesh import Plot
 from sleplet.slepian_methods import slepian_mesh_inverse
 
@@ -593,9 +587,7 @@ python -m examples.mesh.denoising_slepian_mesh homer -n -5 -s 2
 
 ```python
 from sleplet.meshes.mesh import Mesh
-from sleplet.meshes.slepian_coefficients.mesh_slepian_field import (
-    MeshSlepianField,
-)
+from sleplet.meshes.slepian_coefficients import MeshSlepianField
 from sleplet.plot_methods import compute_amplitude_for_noisy_mesh_plots
 from sleplet.plotting.create_plot_mesh import Plot
 from sleplet.slepian_methods import slepian_mesh_inverse
