### 2023-03-29 ebb9ef6
Move Slepian classes
