+++ examples/polar_cap/eigenfunctions.py
@@ -9,7 +9,7 @@ from numpy import typing as npt
 from sleplet import logger
 from sleplet.harmonic_methods import invert_flm_boosted
 from sleplet.plot_methods import calc_plot_resolution, save_plot
-from sleplet.slepian.slepian_polar_cap import SlepianPolarCap
+from sleplet.slepian import SlepianPolarCap
 
 _fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "_figures"
 sns.set(context="paper")
