+++ src/sleplet/slepian/__init__.py
@@ -1,3 +1,7 @@
 """
 classes to create the Slepian regions on the sphere
 """
+
+from .slepian_arbitrary import SlepianArbitrary
+from .slepian_limit_lat_lon import SlepianLimitLatLon
+from .slepian_polar_cap import SlepianPolarCap
