+++ src/sleplet/meshes/__init__.py
@@ -1,3 +1,6 @@
 """
 classes to create and handle mesh (manifold) data
 """
+
+from .mesh import Mesh
+from .mesh_slepian import MeshSlepian
