+++ src/sleplet/plot_methods.py
@@ -9,18 +9,9 @@ from matplotlib import colors
 from matplotlib import pyplot as plt
 from numpy import typing as npt
 
-from sleplet import logger
-from sleplet._mask_methods import create_mask_region
-from sleplet._vars import (
-    SAMPLING_SCHEME,
-    SPHERE_UNSEEN,
-)
-from sleplet.functions.coefficients import Coefficients
-from sleplet.harmonic_methods import invert_flm_boosted, mesh_inverse
-from sleplet.meshes.mesh_coefficients import MeshCoefficients
-from sleplet.meshes.mesh_slepian_coefficients import MeshSlepianCoefficients
-from sleplet.region import Region
-from sleplet.slepian_methods import slepian_inverse, slepian_mesh_inverse
+import sleplet
+import sleplet.functions
+import sleplet.meshes
 
 
 def calc_plot_resolution(L: int) -> int:
@@ -64,11 +55,11 @@ def _calc_nearest_grid_point(
     values - the translation needs to be at the same position
     as the rotation such that the difference error is small
     """
-    thetas, phis = ssht.sample_positions(L, Method=SAMPLING_SCHEME)
+    thetas, phis = ssht.sample_positions(L, Method=sleplet._vars.SAMPLING_SCHEME)
     pix_j = np.abs(phis - alpha_pi_fraction * np.pi).argmin()
     pix_i = np.abs(thetas - beta_pi_fraction * np.pi).argmin()
     alpha, beta = phis[pix_j], thetas[pix_i]
-    logger.info(f"grid point: (alpha, beta)=({alpha:e}, {beta:e})")
+    sleplet.logger.info(f"grid point: (alpha, beta)=({alpha:e}, {beta:e})")
     return alpha, beta
 
 
@@ -79,7 +70,7 @@ def save_plot(path: Path, name: str) -> None:
     plt.tight_layout()
     for file_type in {"png", "pdf"}:
         filename = path / file_type / f"{name}.{file_type}"
-        logger.info(f"saving {filename}")
+        sleplet.logger.info(f"saving {filename}")
         plt.savefig(filename, bbox_inches="tight")
     plt.show(block=False)
     plt.pause(3)
@@ -87,7 +78,10 @@ def save_plot(path: Path, name: str) -> None:
 
 
 def find_max_amplitude(
-    function: Coefficients, *, plot_type: str = "real", upsample: bool = True
+    function: sleplet.functions.Coefficients,
+    *,
+    plot_type: str = "real",
+    upsample: bool = True,
 ) -> float:
     """
     for a given set of coefficients it finds the largest absolute value for a
@@ -95,9 +89,13 @@ def find_max_amplitude(
     """
     # compute inverse transform
     if hasattr(function, "slepian"):
-        field = slepian_inverse(function.coefficients, function.L, function.slepian)
+        field = sleplet.slepian_methods.slepian_inverse(
+            function.coefficients, function.L, function.slepian
+        )
     else:
-        field = ssht.inverse(function.coefficients, function.L, Method=SAMPLING_SCHEME)
+        field = ssht.inverse(
+            function.coefficients, function.L, Method=sleplet._vars.SAMPLING_SCHEME
+        )
 
     # find resolution of final plot for boosting if necessary
     resolution = calc_plot_resolution(function.L) if upsample else function.L
@@ -122,7 +120,7 @@ def _create_plot_type(
     """
     gets the given plot type of the field
     """
-    logger.info(f"plotting type: '{plot_type}'")
+    sleplet.logger.info(f"plotting type: '{plot_type}'")
     plot_dict = {
[…]
