+++ src/sleplet/_mask_methods.py
@@ -4,26 +4,9 @@ import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
-from sleplet import (
-    PHI_MAX,
-    PHI_MIN,
-    POLAR_GAP,
-    SLEPIAN_MASK,
-    THETA_MAX,
-    THETA_MIN,
-    logger,
-)
-from sleplet._data.create_earth_flm import create_flm
-from sleplet._data.setup_pooch import find_on_pooch_then_local
-from sleplet._vars import SAMPLING_SCHEME
-from sleplet.harmonic_methods import (
-    mesh_forward,
-    mesh_inverse,
-    rotate_earth_to_africa,
-    rotate_earth_to_south_america,
-)
-from sleplet.meshes.mesh import Mesh
-from sleplet.region import Region
+import sleplet
+import sleplet._data
+import sleplet.meshes
 
 _data_path = Path(__file__).resolve().parent / "_data"
 
@@ -31,7 +14,7 @@ AFRICA_RANGE = np.deg2rad(41)
 SOUTH_AMERICA_RANGE = np.deg2rad(40)
 
 
-def create_mask_region(L: int, region: Region) -> npt.NDArray[np.float_]:
+def create_mask_region(L: int, region: sleplet.region.Region) -> npt.NDArray[np.float_]:
     """
     creates a mask of a region of interested, the output will be based
     on the value of the provided L. The mask could be either:
@@ -40,11 +23,13 @@ def create_mask_region(L: int, region: Region) -> npt.NDArray[np.float_]:
                                    phi_min or phi_max is provided
     * arbitrary - just checks the shape of the input mask
     """
-    thetas, phis = ssht.sample_positions(L, Grid=True, Method=SAMPLING_SCHEME)
+    thetas, phis = ssht.sample_positions(
+        L, Grid=True, Method=sleplet._vars.SAMPLING_SCHEME
+    )
 
     match region.region_type:
         case "arbitrary":
-            logger.info("loading and checking shape of provided mask")
+            sleplet.logger.logger.info("loading and checking shape of provided mask")
             name = f"{region.mask_name}_L{L}.npy"
             mask = _load_mask(L, name)
             assert mask.shape == thetas.shape, (  # noqa: S101
@@ -53,7 +38,7 @@ def create_mask_region(L: int, region: Region) -> npt.NDArray[np.float_]:
             )
 
         case "lim_lat_lon":
-            logger.info("creating limited latitude longitude mask")
+            sleplet.logger.info("creating limited latitude longitude mask")
             mask = (
                 (thetas >= region.theta_min)
                 & (thetas <= region.theta_max)
@@ -62,10 +47,10 @@ def create_mask_region(L: int, region: Region) -> npt.NDArray[np.float_]:
             )
 
         case "polar":
-            logger.info("creating polar cap mask")
+            sleplet.logger.logger.info("creating polar cap mask")
             mask = thetas <= region.theta_max
             if region.gap:
-                logger.info("creating polar gap mask")
+                sleplet.logger.logger.info("creating polar gap mask")
                 mask += thetas >= np.pi - region.theta_max
     return mask
 
@@ -74,33 +59,42 @@ def _load_mask(L: int, mask_name: str) -> npt.NDArray[np.float_]:
     """
     attempts to read the mask from the config file
     """
[…]
