+++ src/sleplet/_smoothing.py
@@ -2,7 +2,7 @@ import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
-from sleplet import logger
+import sleplet
 
 
 def apply_gaussian_smoothing(
@@ -17,5 +17,5 @@ def apply_gaussian_smoothing(
     """
     sigma = np.pi / (smoothing_factor * L)
     fwhm = 2 * np.sqrt(np.log(2)) * sigma
-    logger.info(f"FWHM = {np.rad2deg(fwhm):.2f}degrees")
+    sleplet.logger.info(f"FWHM = {np.rad2deg(fwhm):.2f}degrees")
     return ssht.gaussian_smoothing(flm, L, sigma)

[…]
