+++ src/sleplet/functions/flm/dirac_delta.py
@@ -3,13 +3,13 @@ import pyssht as ssht
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
-from sleplet._string_methods import _convert_camel_case_to_snake_case
-from sleplet._validation import Validation
-from sleplet.functions.f_lm import F_LM
+import sleplet
+import sleplet._validation
+import sleplet.functions.f_lm
 
 
-@dataclass(config=Validation)
-class DiracDelta(F_LM):
+@dataclass(config=sleplet._validation.Validation)
+class DiracDelta(sleplet.functions.f_lm.F_LM):
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
 
@@ -21,7 +21,9 @@ class DiracDelta(F_LM):
         return flm
 
     def _create_name(self) -> str:
-        return _convert_camel_case_to_snake_case(self.__class__.__name__)
+        return sleplet._string_methods._convert_camel_case_to_snake_case(
+            self.__class__.__name__
+        )
 
     def _set_reality(self) -> bool:
         return True
