+++ src/sleplet/harmonic_methods.py
@@ -10,7 +10,8 @@ from numpy import typing as npt
 from numpy.random import Generator
 
 import sleplet
-import sleplet.meshes
+import sleplet._data.create_earth_flm
+import sleplet.meshes.mesh
 
 AFRICA_ALPHA = np.deg2rad(44)
 AFRICA_BETA = np.deg2rad(87)
@@ -109,7 +110,7 @@ def compute_random_signal(
 
 
 def mesh_forward(
-    mesh: sleplet.meshes.Mesh, u: npt.NDArray[np.complex_ | np.float_]
+    mesh: sleplet.meshes.mesh.Mesh, u: npt.NDArray[np.complex_ | np.float_]
 ) -> npt.NDArray[np.float_]:
     """
     computes the mesh forward transform from real space to harmonic space
@@ -123,7 +124,7 @@ def mesh_forward(
 
 
 def mesh_inverse(
-    mesh: sleplet.meshes.Mesh, u_i: npt.NDArray[np.complex_ | np.float_]
+    mesh: sleplet.meshes.mesh.Mesh, u_i: npt.NDArray[np.complex_ | np.float_]
 ) -> npt.NDArray[np.complex_ | np.float_]:
     """
     computes the mesh inverse transform from harmonic space to real space
