+++ src/sleplet/functions/fp/slepian_wavelet_coefficients_africa.py
@@ -4,44 +4,38 @@ from pydantic import validator
 from pydantic.dataclasses import dataclass
 from pys2let import pys2let_j_max
 
-from sleplet import logger
-from sleplet._string_methods import (
-    _convert_camel_case_to_snake_case,
-    filename_args,
-    wavelet_ending,
-)
-from sleplet._validation import Validation
-from sleplet.functions.f_p import F_P
-from sleplet.functions.fp.slepian_africa import SlepianAfrica
-from sleplet.functions.fp.slepian_wavelets import SlepianWavelets
-from sleplet.region import Region
-from sleplet.wavelet_methods import slepian_wavelet_forward
+import sleplet
+import sleplet._validation
+import sleplet.functions.f_p
 
 
-@dataclass(config=Validation, kw_only=True)
-class SlepianWaveletCoefficientsAfrica(F_P):
+@dataclass(config=sleplet._validation.Validation, kw_only=True)
+class SlepianWaveletCoefficientsAfrica(sleplet.functions.f_p.F_P):
     B: int = 3
     j_min: int = 2
     j: int | None = None
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
-        if isinstance(self.region, Region) and self.region.name_ending != "africa":
+        if (
+            isinstance(self.region, sleplet.region.Region)
+            and self.region.name_ending != "africa"
+        ):
             raise RuntimeError("Slepian region selected must be 'africa'")
 
     def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
-        logger.info("start computing wavelet coefficients")
+        sleplet.logger.info("start computing wavelet coefficients")
         self.wavelets, self.wavelet_coefficients = self._create_wavelet_coefficients()
-        logger.info("finish computing wavelet coefficients")
+        sleplet.logger.info("finish computing wavelet coefficients")
         jth = 0 if self.j is None else self.j + 1
         return self.wavelet_coefficients[jth]
 
     def _create_name(self) -> str:
         return (
-            f"{_convert_camel_case_to_snake_case(self.__class__.__name__)}"
-            f"{filename_args(self.B, 'B')}"
-            f"{filename_args(self.j_min, 'jmin')}"
-            f"{wavelet_ending(self.j_min, self.j)}"
+            f"{sleplet._string_methods._convert_camel_case_to_snake_case(self.__class__.__name__)}"
+            f"{sleplet._string_methods.filename_args(self.B, 'B')}"
+            f"{sleplet._string_methods.filename_args(self.j_min, 'jmin')}"
+            f"{sleplet._string_methods.wavelet_ending(self.j_min, self.j)}"
         )
 
     def _set_reality(self) -> bool:
@@ -63,10 +57,14 @@ class SlepianWaveletCoefficientsAfrica(F_P):
         """
         computes wavelet coefficients in Slepian space
         """
-        sw = SlepianWavelets(self.L, B=self.B, j_min=self.j_min, region=self.region)
-        sa = SlepianAfrica(self.L, region=self.region, smoothing=self.smoothing)
+        sw = sleplet.functions.fp.SlepianWavelets(
+            self.L, B=self.B, j_min=self.j_min, region=self.region
+        )
+        sa = sleplet.functions.fp.SlepianAfrica(
+            self.L, region=self.region, smoothing=self.smoothing
+        )
         wavelets = sw.wavelets
-        wavelet_coefficients = slepian_wavelet_forward(
+        wavelet_coefficients = sleplet.wavelet_methods.slepian_wavelet_forward(
             sa.coefficients, wavelets, self.slepian.N
         )
         return wavelets, wavelet_coefficients
