+++ src/sleplet/functions/flm/elongated_gaussian.py
@@ -2,9 +2,11 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
-import sleplet
+import sleplet._string_methods
 import sleplet._validation
+import sleplet._vars
 import sleplet.functions.f_lm
+import sleplet.harmonic_methods
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
