+++ src/sleplet/meshes/mesh_slepian_coefficients.py
@@ -7,8 +7,8 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
-import sleplet
 import sleplet._validation
+import sleplet.meshes.mesh_coefficients
 import sleplet.meshes.mesh_slepian
 import sleplet.noise
 
