+++ src/sleplet/meshes/_mesh_slepian_decomposition.py
@@ -5,7 +5,9 @@ from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 import sleplet
+import sleplet._integration_methods
 import sleplet._validation
+import sleplet.harmonic_methods
 import sleplet.meshes.mesh_slepian
 
 
