+++ src/sleplet/slepian/slepian_arbitrary.py
@@ -12,9 +12,15 @@ from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
 import sleplet
+import sleplet._array_methods
+import sleplet._data.setup_pooch
+import sleplet._integration_methods
 import sleplet._mask_methods
+import sleplet._parallel_methods
 import sleplet._slepian_arbitrary_methods
 import sleplet._validation
+import sleplet.harmonic_methods
+import sleplet.region
 import sleplet.slepian.slepian_functions
 
 _data_path = Path(__file__).resolve().parents[1] / "_data"
