+++ src/sleplet/functions/__init__.py
@@ -1,0 +1,3 @@
+"""
+set of classes to create functions on the sphere
+"""
