+++ src/sleplet/_class_lists.py
@@ -30,9 +30,7 @@ import sleplet.functions.fp.slepian_south_america
 import sleplet.functions.fp.slepian_wavelet_coefficients_africa
 import sleplet.functions.fp.slepian_wavelet_coefficients_south_america
 import sleplet.functions.fp.slepian_wavelets
-import sleplet.meshes.harmonic_coefficients.mesh_basis_functions
-import sleplet.meshes.harmonic_coefficients.mesh_field
-import sleplet.meshes.harmonic_coefficients.mesh_noise_field
+import sleplet.meshes.harmonic_coefficients
 import sleplet.meshes.mesh_coefficients
 import sleplet.meshes.slepian_coefficients.mesh_slepian_field
 import sleplet.meshes.slepian_coefficients.mesh_slepian_functions
@@ -90,9 +88,9 @@ MAPS_LM: list[type[sleplet.functions.coefficients.Coefficients]] = [
 ]
 
 MESH_HARMONIC: list[type[sleplet.meshes.mesh_coefficients.MeshCoefficients]] = [
-    sleplet.meshes.harmonic_coefficients.mesh_basis_functions.MeshBasisFunctions,
-    sleplet.meshes.harmonic_coefficients.mesh_field.MeshField,
-    sleplet.meshes.harmonic_coefficients.mesh_noise_field.MeshNoiseField,
+    sleplet.meshes.harmonic_coefficients.MeshBasisFunctions,
+    sleplet.meshes.harmonic_coefficients.MeshField,
+    sleplet.meshes.harmonic_coefficients.MeshNoiseField,
 ]
 
 MESH_SLEPIAN: list[type[sleplet.meshes.mesh_coefficients.MeshCoefficients]] = [
