+++ examples/polar_cap/fried_egg.py
@@ -1,8 +1,8 @@
 import numpy as np
 
 from sleplet import logger
-from sleplet.functions.fp.slepian import Slepian
-from sleplet.plotting.create_plot_sphere import Plot
+from sleplet.functions.fp import Slepian
+from sleplet.plotting import PlotSphere
 from sleplet.region import Region
 from sleplet.slepian_methods import slepian_inverse
 
@@ -21,7 +21,7 @@ def main() -> None:
         logger.info(f"plotting rank={r}")
         slepian_function = Slepian(L, rank=r, region=region)
         f = slepian_inverse(slepian_function.coefficients, L, slepian_function.slepian)
-        Plot(
+        PlotSphere(
             f,
             L,
             slepian_function.name,
