+++ examples/arbitrary/africa/progressive_wavelet_reconstruction_africa.py
@@ -1,10 +1,8 @@
 import numpy as np
 
 from sleplet import logger
-from sleplet.functions.fp.slepian_wavelet_coefficients_africa import (
-    SlepianWaveletCoefficientsAfrica,
-)
-from sleplet.plotting.create_plot_sphere import Plot
+from sleplet.functions.fp import SlepianWaveletCoefficientsAfrica
+from sleplet.plotting import PlotSphere
 from sleplet.region import Region
 from sleplet.slepian_methods import slepian_inverse
 from sleplet.wavelet_methods import slepian_wavelet_inverse
@@ -32,7 +30,7 @@ def main() -> None:
         f_p += slepian_wavelet_inverse(coeff, swc.wavelets, swc.slepian.N)
         f = slepian_inverse(f_p, L, swc.slepian)
         name = f"africa_wavelet_reconstruction_progressive_{p}_L{L}"
-        Plot(f, L, name, normalise=NORMALISE, region=swc.region).execute()
+        PlotSphere(f, L, name, normalise=NORMALISE, region=swc.region).execute()
 
 
 if __name__ == "__main__":

[…]
