+++ src/sleplet/plotting/__init__.py
@@ -1,3 +1,11 @@
 """
 classes to create plots on the sphere or mesh
 """
+
+from .create_plot_mesh import PlotMesh
+from .create_plot_sphere import PlotSphere
+
+__all__ = [
+    "PlotSphere",
+    "PlotMesh",
+]
