+++ src/sleplet/meshes/slepian_coefficients/__init__.py
@@ -1,3 +1,17 @@
 """
 classes of functions on the sphere created in Fourier space
 """
+
+from ._mesh_slepian_field import MeshSlepianField
+from ._mesh_slepian_functions import MeshSlepianFunctions
+from ._mesh_slepian_noise_field import MeshSlepianNoiseField
+from ._mesh_slepian_wavelet_coefficients import MeshSlepianWaveletCoefficients
+from ._mesh_slepian_wavelets import MeshSlepianWavelets
+
+__all__ = [
+    "MeshSlepianField",
+    "MeshSlepianFunctions",
+    "MeshSlepianNoiseField",
+    "MeshSlepianWaveletCoefficients",
+    "MeshSlepianWavelets",
+]
