+++ src/sleplet/meshes/slepian_coefficients/_mesh_slepian_wavelet_coefficients.py
@@ -7,13 +7,14 @@ from pys2let import pys2let_j_max
 import sleplet
 import sleplet._string_methods
 import sleplet._validation
-import sleplet.meshes.mesh_slepian_coefficients
+import sleplet.meshes._mesh_slepian_coefficients
+import sleplet.meshes.mesh_slepian
 import sleplet.wavelet_methods
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class MeshSlepianWavelets(
-    sleplet.meshes.mesh_slepian_coefficients.MeshSlepianCoefficients
+class MeshSlepianWaveletCoefficients(
+    sleplet.meshes._mesh_slepian_coefficients.MeshSlepianCoefficients
 ):
     B: int = 3
     j_min: int = 2
@@ -23,15 +24,15 @@ class MeshSlepianWavelets(
         super().__post_init_post_parse__()
 
     def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
-        sleplet.logger.info("start computing wavelets")
-        self.wavelets = self._create_wavelets()
-        sleplet.logger.info("finish computing wavelets")
+        sleplet.logger.info("start computing wavelet coefficients")
+        self.wavelets, self.wavelet_coefficients = self._create_wavelet_coefficients()
+        sleplet.logger.info("finish computing wavelet coefficients")
         jth = 0 if self.j is None else self.j + 1
-        return self.wavelets[jth]
+        return self.wavelet_coefficients[jth]
 
     def _create_name(self) -> str:
         return (
-            f"slepian_wavelets_{self.mesh.name}"
+            f"slepian_wavelet_coefficients_{self.mesh.name}"
             f"{sleplet._string_methods.filename_args(self.B, 'B')}"
             f"{sleplet._string_methods.filename_args(self.j_min, 'jmin')}"
             f"{sleplet._string_methods.wavelet_ending(self.j_min, self.j)}"
@@ -44,16 +45,26 @@ class MeshSlepianWavelets(
                 raise ValueError(f"The number of extra arguments should be {num_args}")
             self.B, self.j_min, self.j = self.extra_args
 
-    def _create_wavelets(self) -> npt.NDArray[np.float_]:
+    def _create_wavelet_coefficients(
+        self,
+    ) -> tuple[npt.NDArray[np.float_], npt.NDArray[np.complex_ | np.float_]]:
         """
-        creates the Slepian wavelets of the mesh
+        computes wavelet coefficients in Slepian space
         """
-        return sleplet.wavelet_methods.create_kappas(
-            self.mesh.mesh_eigenvalues.shape[0], self.B, self.j_min
+        smw = sleplet.meshes.mesh_slepian.MeshSlepianWavelets(
+            self.mesh, B=self.B, j_min=self.j_min
         )
+        smf = sleplet.meshes.mesh_slepian.MeshSlepianField(self.mesh)
+        wavelets = smw.wavelets
+        wavelet_coefficients = sleplet.wavelet_methods.slepian_wavelet_forward(
+            smf.coefficients,
+            wavelets,
+            self.mesh_slepian.N,
+        )
+        return wavelets, wavelet_coefficients
 
     @validator("j")
-    def _check_j(cls, v, values) -> int | None:
+    def _check_j(cls, v, values):
         j_max = pys2let_j_max(
             values["B"], values["mesh"].mesh_eigenvalues.shape[0], values["j_min"]
         )
