+++ src/sleplet/meshes/harmonic_coefficients/_mesh_basis_functions.py
@@ -6,12 +6,12 @@ from pydantic.dataclasses import dataclass
 import sleplet
 import sleplet._validation
 import sleplet.harmonic_methods
-import sleplet.meshes.mesh_harmonic_coefficients
+import sleplet.meshes._mesh_harmonic_coefficients
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class MeshBasisFunctions(
-    sleplet.meshes.mesh_harmonic_coefficients.MeshHarmonicCoefficients
+    sleplet.meshes._mesh_harmonic_coefficients.MeshHarmonicCoefficients
 ):
     rank: int = 0
 
