+++ src/sleplet/meshes/harmonic_coefficients/__init__.py
@@ -2,9 +2,9 @@
 classes of functions on the mesh created in Fourier space
 """
 
-from ._mesh_basis_functions import MeshBasisFunctions
-from ._mesh_field import MeshField
-from ._mesh_noise_field import MeshNoiseField
+from .mesh_basis_functions import MeshBasisFunctions
+from .mesh_field import MeshField
+from .mesh_noise_field import MeshNoiseField
 
 __all__ = [
     "MeshBasisFunctions",
