+++ src/sleplet/slepian_methods.py
@@ -15,7 +15,6 @@ import sleplet.slepian.slepian_arbitrary
 import sleplet.slepian.slepian_functions
 import sleplet.slepian.slepian_limit_lat_lon
 import sleplet.slepian.slepian_polar_cap
-from sleplet.meshes import MeshSlepian
 
 
 def choose_slepian_method(
@@ -119,7 +118,7 @@ def _compute_s_p_omega_prime(
 
 
 def slepian_mesh_forward(
-    mesh_slepian: MeshSlepian,
+    mesh_slepian: sleplet.meshes.MeshSlepian,
     *,
     u: npt.NDArray[np.complex_ | np.float_] | None = None,
     u_i: npt.NDArray[np.complex_ | np.float_] | None = None,
