+++ src/sleplet/meshes/slepian_coefficients/__init__.py
@@ -2,11 +2,11 @@
 classes of functions on the sphere created in Fourier space
 """
 
-from ._mesh_slepian_field import MeshSlepianField
-from ._mesh_slepian_functions import MeshSlepianFunctions
-from ._mesh_slepian_noise_field import MeshSlepianNoiseField
-from ._mesh_slepian_wavelet_coefficients import MeshSlepianWaveletCoefficients
-from ._mesh_slepian_wavelets import MeshSlepianWavelets
+from .mesh_slepian_field import MeshSlepianField
+from .mesh_slepian_functions import MeshSlepianFunctions
+from .mesh_slepian_noise_field import MeshSlepianNoiseField
+from .mesh_slepian_wavelet_coefficients import MeshSlepianWaveletCoefficients
+from .mesh_slepian_wavelets import MeshSlepianWavelets
 
 __all__ = [
     "MeshSlepianField",
