+++ examples/misc/_denoising_axisym.py
@@ -2,8 +2,7 @@ import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
-from sleplet.functions import Coefficients
-from sleplet.functions.flm import AxisymmetricWavelets
+from sleplet.functions.flm import Africa, AxisymmetricWavelets, Earth, SouthAmerica
 from sleplet.harmonic_methods import rotate_earth_to_south_america
 from sleplet.noise import (
     _compute_sigma_j,
@@ -19,8 +18,8 @@ SAMPLING_SCHEME = "MWSS"
 
 
 def denoising_axisym(
-    signal: Coefficients,
-    noised_signal: Coefficients,
+    signal: Africa | Earth | SouthAmerica,
+    noised_signal: Africa | Earth | SouthAmerica,
     axisymmetric_wavelets: AxisymmetricWavelets,
     snr_in: float,
     n_sigma: int,

[…]
