+++ src/sleplet/noise.py
@@ -12,6 +12,7 @@ import sleplet.harmonic_methods
 import sleplet.meshes.mesh_slepian
 import sleplet.slepian.slepian_functions
 import sleplet.slepian_methods
+from sleplet.slepian.slepian_functions import SlepianFunctions
 
 
 def _signal_power(signal: npt.NDArray[np.complex_ | np.float_]) -> float:
@@ -82,7 +83,7 @@ def _create_noise(
 def _create_slepian_noise(
     L: int,
     slepian_signal: npt.NDArray[np.complex_ | np.float_],
-    slepian: sleplet.slepian.slepian_functions.SlepianFunctions,
+    slepian: SlepianFunctions,
     snr_in: float,
 ) -> npt.NDArray[np.complex_]:
     """
@@ -134,7 +135,7 @@ def slepian_wavelet_hard_thresholding(
     wav_coeffs: npt.NDArray[np.complex_ | np.float_],
     sigma_j: npt.NDArray[np.float_],
     n_sigma: int,
-    slepian: sleplet.slepian.slepian_functions.SlepianFunctions,
+    slepian: SlepianFunctions,
 ) -> npt.NDArray[np.complex_ | np.float_]:
     """
     perform thresholding in Slepian wavelet space
@@ -155,7 +156,7 @@ def slepian_function_hard_thresholding(
     coefficients: npt.NDArray[np.complex_ | np.float_],
     sigma: float,
     n_sigma: int,
-    slepian: sleplet.slepian.slepian_functions.SlepianFunctions,
+    slepian: SlepianFunctions,
 ) -> npt.NDArray[np.complex_]:
     """
     perform thresholding in Slepian function space
@@ -184,7 +185,7 @@ def _compute_slepian_sigma_j(
     signal: npt.NDArray[np.complex_ | np.float_],
     psi_j: npt.NDArray[np.float_],
     snr_in: float,
-    slepian: sleplet.slepian.slepian_functions.SlepianFunctions,
+    slepian: SlepianFunctions,
 ) -> npt.NDArray[np.float_]:
     """
     compute sigma_j for wavelets used in denoising the signal
