+++ src/sleplet/slepian/slepian_limit_lat_lon.py
@@ -15,14 +15,15 @@ import sleplet._data.setup_pooch
 import sleplet._mask_methods
 import sleplet._validation
 import sleplet._vars
-import sleplet.region
 import sleplet.slepian.slepian_functions
+from sleplet.slepian.region import Region
+from sleplet.slepian.slepian_functions import SlepianFunctions
 
 _data_path = Path(__file__).resolve().parents[1] / "_data"
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class SlepianLimitLatLon(sleplet.slepian.slepian_functions.SlepianFunctions):
+class SlepianLimitLatLon(SlepianFunctions):
     phi_max: float = sleplet._vars.PHI_MAX_DEFAULT
     phi_min: float = sleplet._vars.PHI_MIN_DEFAULT
     theta_max: float = sleplet._vars.THETA_MAX_DEFAULT
@@ -34,8 +35,8 @@ class SlepianLimitLatLon(sleplet.slepian.slepian_functions.SlepianFunctions):
     def _create_fn_name(self) -> str:
         return f"slepian_{self.region.name_ending}"
 
-    def _create_region(self) -> sleplet.region.Region:
-        return sleplet.region.Region(
+    def _create_region(self) -> Region:
+        return Region(
             theta_min=self.theta_min,
             theta_max=self.theta_max,
             phi_min=self.phi_min,
