+++ examples/arbitrary/south_america/slepian_reconstruction_south_america.py
@@ -1,6 +1,6 @@
 from sleplet.functions.fp import SlepianSouthAmerica
 from sleplet.plotting import PlotSphere
-from sleplet.slepian.region import Region
+from sleplet.slepian import Region
 from sleplet.slepian_methods import choose_slepian_method, slepian_inverse
 
 L = 128
