+++ examples/polar_cap/slepian_error.py
@@ -8,8 +8,7 @@ from numpy import typing as npt
 
 from sleplet.functions.flm import Earth
 from sleplet.plot_methods import save_plot
-from sleplet.slepian import SlepianPolarCap
-from sleplet.slepian.region import Region
+from sleplet.slepian import Region, SlepianPolarCap
 from sleplet.slepian_methods import choose_slepian_method, slepian_forward
 
 _fig_path = Path(__file__).resolve().parents[2] / "src" / "sleplet" / "_figures"
