### 2023-03-30 361c922
Add docstring templates for all functions
