+++ src/sleplet/meshes/slepian_coefficients/mesh_slepian_wavelets.py
@@ -16,8 +16,11 @@ class MeshSlepianWavelets(
     sleplet.meshes.mesh_slepian_coefficients.MeshSlepianCoefficients
 ):
     B: int = 3
+    """TODO"""
     j_min: int = 2
+    """TODO"""
     j: int | None = None
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
