+++ src/sleplet/functions/fp/slepian_noise_africa.py
@@ -13,6 +13,7 @@ from sleplet.slepian.region import Region
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class SlepianNoiseAfrica(sleplet.functions.f_p.F_P):
     SNR: float = -10
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
