+++ src/sleplet/meshes/harmonic_coefficients/mesh_basis_functions.py
@@ -14,6 +14,7 @@ class MeshBasisFunctions(
     sleplet.meshes.mesh_harmonic_coefficients.MeshHarmonicCoefficients
 ):
     rank: int = 0
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         self._validate_rank()
