+++ src/sleplet/functions/flm/directional_spin_wavelets.py
@@ -14,10 +14,15 @@ import sleplet.functions.f_lm
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class DirectionalSpinWavelets(sleplet.functions.f_lm.F_LM):
     B: int = 3
+    """TODO"""
     j_min: int = 2
+    """TODO"""
     j: int | None = None
+    """TODO"""
     N: int = 2
+    """TODO"""
     spin: int = 0
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
