+++ src/sleplet/meshes/slepian_coefficients/mesh_slepian_wavelet_coefficients.py
@@ -16,6 +16,8 @@ import sleplet.wavelet_methods
 class MeshSlepianWaveletCoefficients(
     sleplet.meshes.mesh_slepian_coefficients.MeshSlepianCoefficients
 ):
+    """TODO"""
+
     B: int = 3
     """TODO"""
     j_min: int = 2
