+++ src/sleplet/functions/fp/slepian_wavelets.py
@@ -13,6 +13,8 @@ import sleplet.wavelet_methods
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class SlepianWavelets(sleplet.functions.f_p.F_P):
+    """TODO"""
+
     B: int = 3
     """TODO"""
     j_min: int = 2
