+++ src/sleplet/meshes/harmonic_coefficients/mesh_basis_functions.py
@@ -13,6 +13,8 @@ import sleplet.meshes.mesh_harmonic_coefficients
 class MeshBasisFunctions(
     sleplet.meshes.mesh_harmonic_coefficients.MeshHarmonicCoefficients
 ):
+    """TODO"""
+
     rank: int = 0
     """TODO"""
 
