+++ src/sleplet/functions/flm/gaussian.py
@@ -10,7 +10,10 @@ import sleplet.functions.f_lm
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class Gaussian(sleplet.functions.f_lm.F_LM):
+    """TODO"""
+
     sigma: float = 10
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
