+++ src/sleplet/functions/flm/noise_earth.py
@@ -11,7 +11,10 @@ import sleplet.noise
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class NoiseEarth(sleplet.functions.f_lm.F_LM):
+    """TODO"""
+
     SNR: float = 10
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
