+++ src/sleplet/functions/fp/slepian_noise_south_america.py
@@ -12,7 +12,10 @@ from sleplet.slepian.region import Region
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class SlepianNoiseSouthAmerica(sleplet.functions.f_p.F_P):
+    """TODO"""
+
     SNR: float = -10
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
