+++ src/sleplet/functions/flm/harmonic_gaussian.py
@@ -10,8 +10,12 @@ import sleplet.functions.f_lm
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class HarmonicGaussian(sleplet.functions.f_lm.F_LM):
+    """TODO"""
+
     l_sigma: float = 10
+    """TODO"""
     m_sigma: float = 10
+    """TODO"""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
