+++ src/sleplet/slepian/_slepian_decomposition.py
@@ -9,7 +9,6 @@ import sleplet
 import sleplet._integration_methods
 import sleplet._validation
 import sleplet._vars
-import sleplet.slepian.slepian_functions
 from sleplet.slepian.slepian_functions import SlepianFunctions
 
 
