+++ src/sleplet/wavelet_methods.py
@@ -7,7 +7,6 @@ from numpy import typing as npt
 from pys2let import axisym_wav_l
 
 import sleplet._convolution_methods
-import sleplet.slepian.slepian_functions
 import sleplet.slepian_methods
 
 
