+++ src/sleplet/slepian/slepian_functions.py
@@ -4,6 +4,7 @@ import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
+import sleplet
 import sleplet._validation
 from sleplet.slepian import Region
 
