+++ src/sleplet/functions/fp/slepian_wavelet_coefficients_south_america.py
@@ -10,8 +10,8 @@ import sleplet._validation
 import sleplet.functions.f_p
 import sleplet.functions.fp.slepian_south_america
 import sleplet.functions.fp.slepian_wavelets
+import sleplet.slepian.region
 import sleplet.wavelet_methods
-from sleplet.slepian import Region
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
@@ -28,7 +28,7 @@ class SlepianWaveletCoefficientsSouthAmerica(sleplet.functions.f_p.F_P):
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
         if (
-            isinstance(self.region, Region)
+            isinstance(self.region, sleplet.slepian.region.Region)
             and self.region.name_ending != "south_america"
         ):
             raise RuntimeError("Slepian region selected must be 'south_america'")
