+++ src/sleplet/functions/fp/slepian_south_america.py
@@ -6,7 +6,7 @@ import sleplet._string_methods
 import sleplet._validation
 import sleplet.functions.f_p
 import sleplet.functions.flm.south_america
-from sleplet.slepian import Region
+import sleplet.slepian.region
 
 
 @dataclass(config=sleplet._validation.Validation)
@@ -16,7 +16,7 @@ class SlepianSouthAmerica(sleplet.functions.f_p.F_P):
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
         if (
-            isinstance(self.region, Region)
+            isinstance(self.region, sleplet.slepian.region.Region)
             and self.region.name_ending != "south_america"
         ):
             raise RuntimeError("Slepian region selected must be 'south_america'")
