+++ src/sleplet/__init__.py
@@ -6,7 +6,7 @@ the rest of the documentation.
 import logging
 import os
 
-from ._version import __version__
+from ._version import __version__  # noqa: F401
 
 logging.basicConfig(
     format="[%(asctime)s] [%(levelname)s] --- %(message)s (%(filename)s:%(lineno)s)",
