+++ pyproject.toml
@@ -106,9 +106,7 @@ ignore = [
     "N803",
     "N806",
 ]
-per-file-ignores = {"__init__.py" = [
-    "F401",
-], "test_*" = [
+per-file-ignores = {"test_*" = [
     "S101",
 ]}
 select = [
