+++ README.md
@@ -82,7 +82,7 @@ isn't provided as it is contained within the
 
 [![Sifting Convolution on the Sphere](https://img.shields.io/badge/DOI-10.1109/LSP.2021.3050961-pink.svg)](https://dx.doi.org/10.1109/LSP.2021.3050961)
 
-#### Fig. 1
+#### Sifting Convolution on the Sphere: Fig. 1
 
 ```sh
 for ell in $(seq 2 -1 1); do
@@ -104,7 +104,7 @@ for ell in range(2, 0, -1):
     PlotSphere(f_sphere, f.L, f"fig_1_ell_{ell}", annotations=[]).execute()
 ```
 
-#### Fig. 2
+#### Sifting Convolution on the Sphere: Fig. 2
 
 ```sh
 sphere earth -L 128
@@ -123,7 +123,7 @@ f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
 PlotSphere(f_sphere, f.L, "fig_2").execute()
 ```
 
-#### Fig. 3
+#### Sifting Convolution on the Sphere: Fig. 3
 
 ```sh
 for ell in $(seq 2 -1 1); do
@@ -151,13 +151,13 @@ for ell in range(2, 0, -1):
 
 [![Slepian Scale-Discretised Wavelets on the Sphere](https://img.shields.io/badge/DOI-10.1109/TSP.2022.3233309-pink.svg)](https://dx.doi.org/10.1109/TSP.2022.3233309)
 
-#### Fig. 2
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 2
 
 ```sh
 python -m examples.arbitrary.south_america.tiling_south_america
 ```
 
-#### Fig. 3
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 3
 
 ```sh
 export SLEPIAN_MASK="south_america"
@@ -189,7 +189,7 @@ g_sphere = slepian_inverse(g.coefficients, g.L, g.slepian)
 PlotSphere(g_sphere, g.L, "fig_3_b", normalise=False, region=g.region).execute()
 ```
 
-#### Fig. 4
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 4
 
 ```sh
 export SLEPIAN_MASK="south_america"
@@ -213,13 +213,13 @@ for p in [0, 9, 24, 49, 99, 199]:
     ).execute()
 ```
 
-#### Fig. 5
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 5
 
 ```sh
 python -m examples.arbitrary.south_america.eigenvalues_south_america
 ```
 
-#### Fig. 6
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 6
 
 ```sh
 export SLEPIAN_MASK="south_america"
@@ -246,7 +246,7 @@ for j in [None, *list(range(5))]:
     ).execute()
 ```
 
-#### Fig. 7
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 7
 
 ```sh
 export SLEPIAN_MASK="south_america"
@@ -275,7 +275,7 @@ for j in [None, *list(range(5))]:
     ).execute()
 ```
 
-#### Fig. 8
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 8
 
 ```sh
 export SLEPIAN_MASK="south_america"
@@ -304,7 +304,7 @@ PlotSphere(
 ).execute()
 ```
 
-#### Fig. 9
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 9
 
 ```sh
 export SLEPIAN_MASK="africa"
[…]
