### 2023-03-31 741126b
Add sifting convolution examples
