### 2023-03-31 9fd2082
Add figure 2.1
