+++ .github/workflows/examples.yml
@@ -17,6 +17,13 @@ jobs:
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
 
+      - name: Cache pooch
+        uses: actions/cache@v3
+        with:
+          path: |
+            ~/.cache/sleplet
+          key: tox-${{ hashFiles('pyproject.toml') }}
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -37,6 +44,13 @@ jobs:
           python-version: "3.10"
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
+
+      - name: Cache pooch
+        uses: actions/cache@v3
+        with:
+          path: |
+            ~/.cache/sleplet
+          key: tox-${{ hashFiles('pyproject.toml') }}
 
       - name: Install dependencies
         run: |
