### 2023-03-31 1ac56f5
Improve API documentation (#114)
Making unrequired things private
