+++ examples/arbitrary/africa/sparsity_wavelet_slepian_comparison_africa.py
@@ -5,16 +5,12 @@ import numpy as np
 import seaborn as sns
 from matplotlib import pyplot as plt
 
-from sleplet.functions.flm.axisymmetric_wavelet_coefficients_africa import (
-    AxisymmetricWaveletCoefficientsAfrica,
-)
-from sleplet.functions.fp.slepian_wavelet_coefficients_africa import (
-    SlepianWaveletCoefficientsAfrica,
-)
-from sleplet.utils.plot_methods import save_plot
-from sleplet.utils.region import Region
+from sleplet.functions.flm import AxisymmetricWaveletCoefficientsAfrica
+from sleplet.functions.fp import SlepianWaveletCoefficientsAfrica
+from sleplet.plot_methods import save_plot
+from sleplet.slepian import Region
 
-fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "figures"
+_fig_path = Path(__file__).resolve().parents[3] / "src" / "sleplet" / "_figures"
 sns.set(context="paper")
 
 B = 3
@@ -75,7 +71,7 @@ def main(*, limit: bool) -> None:
         plt.xlim(right=300)
         plt.ylim(top=50)
         filename += "_limit"
-    save_plot(fig_path, filename)
+    save_plot(_fig_path, filename)
 
 
 if __name__ == "__main__":
