### 2023-03-31 77c0d0a
Apply new ruff rules to code
