+++ src/sleplet/harmonic_methods.py
@@ -1,6 +1,4 @@
-"""
-methods to perform operations in Fourier space of the sphere or mesh
-"""
+"""methods to perform operations in Fourier space of the sphere or mesh."""
 from collections.abc import Callable
 from typing import Any
 
@@ -23,13 +21,15 @@ SOUTH_AMERICA_GAMMA = np.deg2rad(63)
 
 
 def _create_spherical_harmonic(L: int, ind: int) -> npt.NDArray[np.complex_]:
-    """TODO create a spherical harmonic in harmonic space for the given index
+    """TODO create a spherical harmonic in harmonic space for the given index.
 
     Args:
+    ----
         L: _description_
         ind: _description_
 
     Returns:
+    -------
         _description_
     """
     flm = np.zeros(L**2, dtype=np.complex_)
@@ -38,11 +38,10 @@ def _create_spherical_harmonic(L: int, ind: int) -> npt.NDArray[np.complex_]:
 
 
 def _boost_coefficient_resolution(
-    flm: npt.NDArray[Any], boost: int
+    flm: npt.NDArray[Any],
+    boost: int,
 ) -> npt.NDArray[Any]:
-    """
-    calculates a boost in resolution for given flm
-    """
+    """calculates a boost in resolution for given flm."""
     return np.pad(flm, (0, boost), "constant")
 
 
@@ -54,9 +53,10 @@ def invert_flm_boosted(
     reality: bool = False,
     spin: int = 0,
 ) -> npt.NDArray[np.complex_ | np.float_]:
-    """TODO performs the inverse harmonic transform
+    """TODO performs the inverse harmonic transform.
 
     Args:
+    ----
         flm: _description_
         L: _description_
         resolution: _description_
@@ -64,6 +64,7 @@ def invert_flm_boosted(
         spin: _description_
 
     Returns:
+    -------
         _description_
     """
     boost = resolution**2 - L**2
@@ -79,30 +80,34 @@ def invert_flm_boosted(
 
 def _ensure_f_bandlimited(
     grid_fun: Callable[
-        [npt.NDArray[np.float_], npt.NDArray[np.float_]], npt.NDArray[np.float_]
+        [npt.NDArray[np.float_], npt.NDArray[np.float_]],
+        npt.NDArray[np.float_],
     ],
     L: int,
     *,
     reality: bool,
     spin: int,
 ) -> npt.NDArray[np.complex_]:
-    """
-    if the function created is created in pixel space rather than harmonic
-    space then need to transform it into harmonic space first before using it
+    """if the function created is created in pixel space rather than harmonic
+    space then need to transform it into harmonic space first before using it.
     """
     thetas, phis = ssht.sample_positions(
-        L, Grid=True, Method=sleplet._vars.SAMPLING_SCHEME
+        L,
+        Grid=True,
+        Method=sleplet._vars.SAMPLING_SCHEME,
     )
     f = grid_fun(thetas, phis)
     return ssht.forward(
-        f, L, Reality=reality, Spin=spin, Method=sleplet._vars.SAMPLING_SCHEME
+        f,
+        L,
+        Reality=reality,
+        Spin=spin,
+        Method=sleplet._vars.SAMPLING_SCHEME,
     )
 
 
 def _create_emm_vector(L: int) -> npt.NDArray[np.float_]:
-    """
-    create vector of m values for a given L
-    """
+    """create vector of m values for a given L."""
     emm = np.zeros(2 * L * 2 * L)
     k = 0
 
@@ -114,17 +119,22 @@ def _create_emm_vector(L: int) -> npt.NDArray[np.float_]:
 
 
 def compute_random_signal(
[…]
