+++ examples/arbitrary/south_america/sparsity_wavelet_slepian_comparison_south_america.py
@@ -20,9 +20,7 @@ STEP = 0.01
 
 
 def _plot_slepian_coefficients() -> int:
-    """
-    plot the Slepian wavelet coefficients for the South America region
-    """
+    """plot the Slepian wavelet coefficients for the South America region."""
     # initialise wavelet coefficients
     region = Region(mask_name="south_america")
     swc = SlepianWaveletCoefficientsSouthAmerica(L, B=B, j_min=J_MIN, region=region)
@@ -40,9 +38,7 @@ def _plot_slepian_coefficients() -> int:
 
 
 def _plot_axisymmetric_coefficients(shannon: int) -> None:
-    """
-    plot the axisymmetric wavelet coefficients for the South America region
-    """
+    """plot the axisymmetric wavelet coefficients for the South America region."""
     # initialise wavelet coefficients
     awc = AxisymmetricWaveletCoefficientsSouthAmerica(L, B=B, j_min=J_MIN)
 
@@ -57,8 +53,7 @@ def _plot_axisymmetric_coefficients(shannon: int) -> None:
 
 
 def main(*, limit: bool) -> None:
-    """
-    Plot a comparison of the absolute values of the wavelet coefficients
+    """Plot a comparison of the absolute values of the wavelet coefficients
     compared to the Slepian coefficients. Expect the Slepian coefficients to
     decay faster than the wavelet coefficients.
     """
