### 2023-03-31 bef0fbe
Apply new ruff rules to code (#117)
