+++ examples/polar_cap/simons_5_3.py
@@ -20,9 +20,7 @@ THETA_RANGE = {10: (0, 0), 20: (0, 1), 30: (1, 0), 40: (1, 1)}
 
 
 def main() -> None:
-    """
-    creates a plot of Slepian eigenvalues against rank
-    """
+    """creates a plot of Slepian eigenvalues against rank."""
     x = len(THETA_RANGE) // 2
     _, ax = plt.subplots(x, x, sharex="col", sharey="row")
     for theta_max, position in THETA_RANGE.items():
@@ -32,9 +30,7 @@ def main() -> None:
 
 
 def _create_plot(ax: npt.NDArray, position: tuple[int, int], theta_max: int) -> None:
-    """
-    helper method which actually makes the plot
-    """
+    """helper method which actually makes the plot."""
     logger.info(f"theta_max={theta_max}")
     slepian = SlepianPolarCap(L, np.deg2rad(theta_max))
     axs = ax[position]
