+++ src/sleplet/_plotly_methods.py
@@ -23,9 +23,7 @@ def create_camera(
     y_center: float = 0,
     z_center: float = 0,
 ) -> Camera:
-    """
-    creates default camera view with a zoom factor
-    """
+    """creates default camera view with a zoom factor."""
     return Camera(
         eye=Eye(x=x_eye / zoom, y=y_eye / zoom, z=z_eye / zoom),
         center=Center(x=x_center, y=y_center, z=z_center),
@@ -33,9 +31,7 @@ def create_camera(
 
 
 def create_layout(camera: Camera, *, annotations: list[dict] | None = None) -> Layout:
-    """
-    default plotly layout
-    """
+    """default plotly layout."""
     return Layout(
         scene=Scene(
             dragmode="orbit",
@@ -52,11 +48,12 @@ def create_layout(camera: Camera, *, annotations: list[dict] | None = None) -> L
 
 
 def create_tick_mark(
-    fmin: float, fmax: float, *, amplitude: float | None = None
+    fmin: float,
+    fmax: float,
+    *,
+    amplitude: float | None = None,
 ) -> float:
-    """
-    creates tick mark to use when using a non-normalised plot
-    """
+    """creates tick mark to use when using a non-normalised plot."""
     return amplitude if amplitude is not None else max(abs(fmin), abs(fmax))
 
 
@@ -68,9 +65,7 @@ def create_colour_bar(
     bar_pos: float = 0.97,
     font_size: int = 38,
 ) -> dict:
-    """
-    default plotly colour bar
-    """
+    """default plotly colour bar."""
     return {
         "x": bar_pos,
         "len": bar_len,
