+++ src/sleplet/functions/flm/spherical_harmonic.py
@@ -12,7 +12,7 @@ import sleplet.harmonic_methods
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class SphericalHarmonic(sleplet.functions.f_lm.F_LM):
-    """TODO"""
+    """TODO."""
 
     ell: int = 0
     """TODO"""

[…]
