+++ pyproject.toml
@@ -111,27 +111,46 @@ per-file-ignores = {"test_*" = [
 ]}
 select = [
     "A",
+    # "ANN", TODO:
+    "ARG",
     "B",
+    "BLE",
     "C",
+    "COM",
+    # "D", TODO:
+    "DJ",
     "DTZ",
     "E",
     "EM",
+    "ERA",
+    "EXE",
     "F",
     "FBT",
+    # "G", TODO:
     "I",
     "ICN",
+    # "INP", TODO:
     "ISC",
     "N",
-    "PLC",
-    "PLE",
-    "PLR",
-    "PLW",
+    "NPY",
+    "PD",
+    "PGH",
+    "PIE",
+    "PL",
+    "PT",
+    "PTH",
+    "PYI",
     "Q",
+    "RET",
+    "RSE",
     "RUF",
     "S",
     "SIM",
+    # "SLF", TODO:
     "T",
+    "TCH",
     "TID",
+    # "TRY", TODO:
     "UP",
     "W",
     "YTT",
