+++ src/sleplet/functions/fp/slepian_africa.py
@@ -12,7 +12,7 @@ import sleplet.slepian_methods
 
 @dataclass(config=sleplet._validation.Validation)
 class SlepianAfrica(sleplet.functions.f_p.F_P):
-    """TODO"""
+    """TODO."""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
@@ -25,12 +25,14 @@ class SlepianAfrica(sleplet.functions.f_p.F_P):
     def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
         a = sleplet.functions.flm.africa.Africa(self.L, smoothing=self.smoothing)
         return sleplet.slepian_methods.slepian_forward(
-            self.L, self.slepian, flm=a.coefficients
+            self.L,
+            self.slepian,
+            flm=a.coefficients,
         )
 
     def _create_name(self) -> str:
         return sleplet._string_methods._convert_camel_case_to_snake_case(
-            self.__class__.__name__
+            self.__class__.__name__,
         )
 
     def _set_reality(self) -> bool:
@@ -42,5 +44,5 @@ class SlepianAfrica(sleplet.functions.f_p.F_P):
     def _setup_args(self) -> None:
         if isinstance(self.extra_args, list):
             raise AttributeError(
-                f"{self.__class__.__name__} does not support extra arguments"
+                f"{self.__class__.__name__} does not support extra arguments",
             )
