+++ src/sleplet/functions/flm/axisymmetric_wavelet_coefficients_earth.py
@@ -14,7 +14,7 @@ import sleplet.wavelet_methods
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class AxisymmetricWaveletCoefficientsEarth(sleplet.functions.f_lm.F_LM):
-    """TODO"""
+    """TODO."""
 
     B: int = 3
     """TODO"""
@@ -57,15 +57,17 @@ class AxisymmetricWaveletCoefficientsEarth(sleplet.functions.f_lm.F_LM):
     def _create_wavelet_coefficients(
         self,
     ) -> tuple[npt.NDArray[np.complex_], npt.NDArray[np.complex_]]:
-        """
-        computes wavelet coefficients of the Earth
-        """
+        """computes wavelet coefficients of the Earth."""
         wavelets = sleplet.wavelet_methods._create_axisymmetric_wavelets(
-            self.L, self.B, self.j_min
+            self.L,
+            self.B,
+            self.j_min,
         )
         self.earth = sleplet.functions.flm.earth.Earth(self.L, smoothing=self.smoothing)
         wavelet_coefficients = sleplet.wavelet_methods.axisymmetric_wavelet_forward(
-            self.L, self.earth.coefficients, wavelets
+            self.L,
+            self.earth.coefficients,
+            wavelets,
         )
         return wavelets, wavelet_coefficients
 
@@ -76,6 +78,6 @@ class AxisymmetricWaveletCoefficientsEarth(sleplet.functions.f_lm.F_LM):
             raise ValueError("j should be positive")
         if v is not None and v > j_max - values["j_min"]:
             raise ValueError(
-                f"j should be less than j_max - j_min: {j_max - values['j_min'] + 1}"
+                f"j should be less than j_max - j_min: {j_max - values['j_min'] + 1}",
             )
         return v
