+++ .github/workflows/deploy.yml
@@ -15,7 +15,7 @@ jobs:
       - name: Set up python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.10"
+          python-version: "3.x"
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
 
