+++ README.md
@@ -71,10 +71,596 @@ The default region is the `south_america` arbitrary region.
 - `SLEPIAN_MASK`
   - for an arbitrary Slepian region, currently `africa`/`south_america` supported
 
+## Paper Figures
+
+To recreate the figures from the below papers, one may use the CLI or the API.
+For those which don't use the `mesh` or `sphere` CLIs, the relevant API code
+isn't provided as it is contained within the
+[examples folder](https://github.com/astro-informatics/sleplet/tree/main/examples).
+
+### Sifting Convolution on the Sphere
+
+[![Sifting Convolution on the Sphere](https://img.shields.io/badge/DOI-10.1109/LSP.2021.3050961-pink.svg)](https://dx.doi.org/10.1109/LSP.2021.3050961)
+
+#### Sifting Convolution on the Sphere: Fig. 1
+
+```sh
+for ell in $(seq 2 -1 1); do
+    sphere harmonic_gaussian -a 0.75 -b 0.125 -e ${ell} 1 -L 128 -m translate -o
+done
+```
+
+```python
+import numpy as np
+import pyssht as ssht
+
+from sleplet.functions.flm import HarmonicGaussian
+from sleplet.plotting import PlotSphere
+
+for ell in range(2, 0, -1):
+    f = HarmonicGaussian(L=128, l_sigma=10**ell, m_sigma=10)
+    flm = f.translate(alpha=0.75 * np.pi, beta=0.125 * np.pi)
+    f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
+    PlotSphere(f_sphere, f.L, f"fig_1_ell_{ell}", annotations=[]).execute()
+```
+
+#### Sifting Convolution on the Sphere: Fig. 2
+
+```sh
+sphere earth -L 128
+```
+
+```python
+import pyssht as ssht
+
+from sleplet.functions.flm import Earth
+from sleplet.harmonic_methods import rotate_earth_to_south_america
+from sleplet.plotting import PlotSphere
+
+f = Earth(L=128)
+flm = rotate_earth_to_south_america(f.coefficients, f.L)
+f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
+PlotSphere(f_sphere, f.L, "fig_2").execute()
+```
+
+#### Sifting Convolution on the Sphere: Fig. 3
+
+```sh
+for ell in $(seq 2 -1 1); do
+    sphere harmonic_gaussian -c earth -e ${ell} 1 -L 128
+done
+```
+
+```python
+import pyssht as ssht
+
+from sleplet.functions.flm import Earth, HarmonicGaussian
+from sleplet.harmonic_methods import rotate_earth_to_south_america
+from sleplet.plotting import PlotSphere
+
+for ell in range(2, 0, -1):
+    f = HarmonicGaussian(L=128, l_sigma=10**ell, m_sigma=10)
+    g = Earth(L=128)
+    flm = f.convolve(f.coefficients, g.coefficients.conj())
+    flm_rot = rotate_earth_to_south_america(flm, f.L)
+    f_sphere = ssht.inverse(flm_rot, f.L, Method="MWSS")
+    PlotSphere(f_sphere, f.L, f"fig_3_ell_{ell}").execute()
+```
+
+### Slepian Scale-Discretised Wavelets on the Sphere
+
+[![Slepian Scale-Discretised Wavelets on the Sphere](https://img.shields.io/badge/DOI-10.1109/TSP.2022.3233309-pink.svg)](https://dx.doi.org/10.1109/TSP.2022.3233309)
+
+#### Slepian Scale-Discretised Wavelets on the Sphere: Fig. 2
+
[…]
