### 2023-04-01 4b2bac8
Remove harder 6 rules
