### 2023-04-01 3c39611
Update ruff config (#118)
