+++ src/sleplet/_scripts/plotting_on_mesh.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 from argparse import ArgumentParser, Namespace
 
 import sleplet
@@ -14,8 +13,7 @@ def valid_meshes(mesh_name: str) -> str:
     """check if valid mesh name."""
     if mesh_name in sleplet._class_lists.MESHES:
         return mesh_name
-    else:
-        raise ValueError(f"'{mesh_name}' is not a valid mesh name to plot")
+    raise ValueError(f"'{mesh_name}' is not a valid mesh name to plot")
 
 
 def valid_methods(method_name: str) -> str:
@@ -25,8 +23,7 @@ def valid_methods(method_name: str) -> str:
         word_to_remove="Mesh",
     ):
         return method_name
-    else:
-        raise ValueError(f"'{method_name}' is not a valid method to plot")
+    raise ValueError(f"'{method_name}' is not a valid method to plot")
 
 
 def read_args() -> Namespace:
