+++ src/sleplet/_data/setup_pooch.py
@@ -21,9 +21,8 @@ def find_on_pooch_then_local(filename: str) -> os.PathLike | None:
     if filename in POOCH.registry:
         sleplet.logger.info(f"Found {filename} at https://doi.org/{ZENODO_DATA_DOI}")
         return POOCH.fetch(filename, progressbar=True)
-    elif (_data_path / filename).exists():
+    if (_data_path / filename).exists():
         sleplet.logger.info(f"Found {filename} at {_data_path / filename}")
         return _data_path / filename
-    else:
-        sleplet.logger.info(f"No {filename} found, calculating...")
-        return None
+    sleplet.logger.info(f"No {filename} found, calculating...")
+    return None
