+++ .github/workflows/test.yml
@@ -37,4 +37,4 @@ jobs:
           python -m pip install tox tox-gh-actions
 
       - name: Test with tox
-        run: tox -e py310
+        run: tox
