+++ pyproject.toml
@@ -111,26 +111,39 @@ per-file-ignores = {"test_*" = [
 ]}
 select = [
     "A",
+    "ARG",
     "B",
+    "BLE",
     "C",
+    "COM",
+    "DJ",
     "DTZ",
     "E",
     "EM",
+    "ERA",
+    "EXE",
     "F",
     "FBT",
     "I",
     "ICN",
     "ISC",
     "N",
-    "PLC",
-    "PLE",
-    "PLR",
-    "PLW",
+    "NPY",
+    "PD",
+    "PGH",
+    "PIE",
+    "PL",
+    "PT",
+    "PTH",
+    "PYI",
     "Q",
+    "RET",
+    "RSE",
     "RUF",
     "S",
     "SIM",
     "T",
+    "TCH",
     "TID",
     "UP",
     "W",
