+++ src/sleplet/_integration_methods.py
@@ -9,7 +9,7 @@ import sleplet._vars
 
 
 def calc_integration_weight(L: int) -> npt.NDArray[np.float_]:
-    """computes the spherical Jacobian for the integration."""
+    """Computes the spherical Jacobian for the integration."""
     thetas, phis = ssht.sample_positions(
         L,
         Grid=True,
@@ -24,7 +24,7 @@ def integrate_whole_sphere(
     weight: npt.NDArray[np.float_],
     *functions: npt.NDArray[np.complex_],
 ) -> complex:
-    """computes the integration for the whole sphere."""
+    """Computes the integration for the whole sphere."""
     multiplied_inputs = _multiply_args(*functions)
     return (multiplied_inputs * weight).sum()
 
@@ -34,7 +34,7 @@ def integrate_region_sphere(
     weight: npt.NDArray[np.float_],
     *functions: npt.NDArray[np.complex_ | np.float_],
 ) -> complex:
-    """computes the integration for a region of the sphere."""
+    """Computes the integration for a region of the sphere."""
     multiplied_inputs = _multiply_args(*functions)
     return (multiplied_inputs * weight * mask).sum()
 
@@ -44,7 +44,7 @@ def integrate_whole_mesh(
     faces: npt.NDArray[np.int_],  # noqa: ARG001
     *functions: npt.NDArray[np.complex_ | np.float_],
 ) -> float:
-    """computes the integral of functions on the vertices."""
+    """Computes the integral of functions on the vertices."""
     multiplied_inputs = _multiply_args(*functions)
     return multiplied_inputs.sum()
 
@@ -55,11 +55,11 @@ def integrate_region_mesh(
     faces: npt.NDArray[np.int_],  # noqa: ARG001
     *functions: npt.NDArray[np.complex_ | np.float_],
 ) -> float:
-    """computes the integral of a region of functions on the vertices."""
+    """Computes the integral of a region of functions on the vertices."""
     multiplied_inputs = _multiply_args(*functions)
     return (multiplied_inputs * mask).sum()
 
 
 def _multiply_args(*args: npt.NDArray[Any]) -> npt.NDArray[Any]:
-    """method to multiply an unknown number of arguments."""
+    """Method to multiply an unknown number of arguments."""
     return reduce((lambda x, y: x * y), args)
