+++ tests/test_plot.py
@@ -15,7 +15,7 @@ THETA_MAX = np.pi / 3
 
 
 def test_resolution_values() -> None:
-    """verifies the correct resolution is chosen for a given bandlimit."""
+    """Verifies the correct resolution is chosen for a given bandlimit."""
     arguments = [1, 10, 100, 1000]
     output = [64, 80, 800, 2000]
     for c, arg in enumerate(arguments):
@@ -23,13 +23,13 @@ def test_resolution_values() -> None:
 
 
 def test_create_colourscale() -> None:
-    """test creates a plotly compatible colourscale."""
+    """Test creates a plotly compatible colourscale."""
     colourscale = _convert_colourscale(cmocean.cm.ice, pl_entries=PL_ENTRIES)
     assert_equal(len(colourscale), PL_ENTRIES)
 
 
 def test_find_nearest_grid_point() -> None:
-    """test to find nearest grid point to provided angles."""
+    """Test to find nearest grid point to provided angles."""
     alpha, beta = _calc_nearest_grid_point(L, PHI_0 / np.pi, THETA_MAX / np.pi)
     assert_equal(alpha, 0.5154175447295755)
     assert_equal(beta, 1.055378782065321)
