+++ src/sleplet/functions/flm/south_america.py
@@ -50,7 +50,7 @@ class SouthAmerica(sleplet.functions.f_lm.F_LM):
         theta: npt.NDArray[np.float_],  # noqa: ARG002
         phi: npt.NDArray[np.float_],  # noqa: ARG002
     ) -> npt.NDArray[np.float_]:
-        """function on the grid."""
+        """Function on the grid."""
         earth_flm = sleplet._data.create_earth_flm.create_flm(
             self.L,
             smoothing=self.smoothing,

[…]
