+++ src/sleplet/_class_lists.py
@@ -30,15 +30,15 @@ import sleplet.functions.fp.slepian_south_america
 import sleplet.functions.fp.slepian_wavelet_coefficients_africa
 import sleplet.functions.fp.slepian_wavelet_coefficients_south_america
 import sleplet.functions.fp.slepian_wavelets
-import sleplet.meshes.harmonic_coefficients.mesh_basis_functions
-import sleplet.meshes.harmonic_coefficients.mesh_field
-import sleplet.meshes.harmonic_coefficients.mesh_noise_field
+import sleplet.meshes.mesh_basis_functions
 import sleplet.meshes.mesh_coefficients
-import sleplet.meshes.slepian_coefficients.mesh_slepian_field
-import sleplet.meshes.slepian_coefficients.mesh_slepian_functions
-import sleplet.meshes.slepian_coefficients.mesh_slepian_noise_field
-import sleplet.meshes.slepian_coefficients.mesh_slepian_wavelet_coefficients
-import sleplet.meshes.slepian_coefficients.mesh_slepian_wavelets
+import sleplet.meshes.mesh_field
+import sleplet.meshes.mesh_noise_field
+import sleplet.meshes.mesh_slepian_field
+import sleplet.meshes.mesh_slepian_functions
+import sleplet.meshes.mesh_slepian_noise_field
+import sleplet.meshes.mesh_slepian_wavelet_coefficients
+import sleplet.meshes.mesh_slepian_wavelets
 
 _data_path = Path(__file__).resolve().parent / "_data"
 
@@ -90,17 +90,17 @@ MAPS_LM: list[type[sleplet.functions.coefficients.Coefficients]] = [
 ]
 
 MESH_HARMONIC: list[type[sleplet.meshes.mesh_coefficients.MeshCoefficients]] = [
-    sleplet.meshes.harmonic_coefficients.mesh_basis_functions.MeshBasisFunctions,
-    sleplet.meshes.harmonic_coefficients.mesh_field.MeshField,
-    sleplet.meshes.harmonic_coefficients.mesh_noise_field.MeshNoiseField,
+    sleplet.meshes.mesh_basis_functions.MeshBasisFunctions,
+    sleplet.meshes.mesh_field.MeshField,
+    sleplet.meshes.mesh_noise_field.MeshNoiseField,
 ]
 
 MESH_SLEPIAN: list[type[sleplet.meshes.mesh_coefficients.MeshCoefficients]] = [
-    sleplet.meshes.slepian_coefficients.mesh_slepian_field.MeshSlepianField,
-    sleplet.meshes.slepian_coefficients.mesh_slepian_functions.MeshSlepianFunctions,
-    sleplet.meshes.slepian_coefficients.mesh_slepian_noise_field.MeshSlepianNoiseField,
-    sleplet.meshes.slepian_coefficients.mesh_slepian_wavelet_coefficients.MeshSlepianWaveletCoefficients,
-    sleplet.meshes.slepian_coefficients.mesh_slepian_wavelets.MeshSlepianWavelets,
+    sleplet.meshes.mesh_slepian_field.MeshSlepianField,
+    sleplet.meshes.mesh_slepian_functions.MeshSlepianFunctions,
+    sleplet.meshes.mesh_slepian_noise_field.MeshSlepianNoiseField,
+    sleplet.meshes.mesh_slepian_wavelet_coefficients.MeshSlepianWaveletCoefficients,
+    sleplet.meshes.mesh_slepian_wavelets.MeshSlepianWavelets,
 ]
 
 MESH_COEFFICIENTS: list[type[sleplet.meshes.mesh_coefficients.MeshCoefficients]] = (
