+++ src/sleplet/meshes/__init__.py
@@ -1,9 +1,25 @@
 """Classes to create and handle mesh (manifold) data."""
 
 from .mesh import Mesh
+from .mesh_basis_functions import MeshBasisFunctions
+from .mesh_field import MeshField
+from .mesh_noise_field import MeshNoiseField
 from .mesh_slepian import MeshSlepian
+from .mesh_slepian_field import MeshSlepianField
+from .mesh_slepian_functions import MeshSlepianFunctions
+from .mesh_slepian_noise_field import MeshSlepianNoiseField
+from .mesh_slepian_wavelet_coefficients import MeshSlepianWaveletCoefficients
+from .mesh_slepian_wavelets import MeshSlepianWavelets
 
 __all__ = [
     "Mesh",
+    "MeshBasisFunctions",
+    "MeshField",
+    "MeshNoiseField",
     "MeshSlepian",
+    "MeshSlepianField",
+    "MeshSlepianFunctions",
+    "MeshSlepianNoiseField",
+    "MeshSlepianWaveletCoefficients",
+    "MeshSlepianWavelets",
 ]
