+++ src/sleplet/meshes/harmonic_coefficients/__init__.py
@@ -1,11 +1,0 @@
-"""Classes of functions on the mesh created in Fourier space."""
-
-from .mesh_basis_functions import MeshBasisFunctions
-from .mesh_field import MeshField
-from .mesh_noise_field import MeshNoiseField
-
-__all__ = [
-    "MeshBasisFunctions",
-    "MeshField",
-    "MeshNoiseField",
-]
