+++ src/sleplet/slepian/slepian_arbitrary.py
@@ -18,7 +18,7 @@ import sleplet._slepian_arbitrary_methods
 import sleplet._validation
 import sleplet.harmonic_methods
 import sleplet.slepian.region
-from sleplet.slepian.slepian_functions import SlepianFunctions
+import sleplet.slepian.slepian_functions
 
 _data_path = Path(__file__).resolve().parents[1] / "_data"
 
@@ -26,7 +26,7 @@ SAMPLES = 2
 
 
 @dataclass(config=sleplet._validation.Validation)
-class SlepianArbitrary(SlepianFunctions):
+class SlepianArbitrary(sleplet.slepian.slepian_functions.SlepianFunctions):
     """Class to create an arbitrary Slepian region on the sphere."""
 
     mask_name: str
