### 2023-04-01 f62d93b
Remove slepian sub-folders
