+++ examples/arbitrary/south_america/denoising_slepian_south_america.py
@@ -3,7 +3,7 @@ from argparse import ArgumentParser
 from pathlib import Path
 
 from sleplet import logger
-from sleplet.functions.fp import SlepianSouthAmerica, SlepianWavelets
+from sleplet.functions import SlepianSouthAmerica, SlepianWavelets
 from sleplet.plot_methods import find_max_amplitude
 from sleplet.plotting import PlotSphere
 from sleplet.slepian import Region
