+++ src/sleplet/functions/axisymmetric_wavelet_coefficients_south_america.py
@@ -7,13 +7,12 @@ from pys2let import pys2let_j_max
 import sleplet
 import sleplet._string_methods
 import sleplet._validation
-import sleplet.functions.f_lm
-import sleplet.functions.flm.earth
+import sleplet.functions.south_america
 import sleplet.wavelet_methods
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class AxisymmetricWaveletCoefficientsEarth(sleplet.functions.f_lm.F_LM):
+class AxisymmetricWaveletCoefficientsSouthAmerica(sleplet.functions.f_lm.F_LM):
     """TODO."""
 
     B: int = 3
@@ -57,16 +56,19 @@ class AxisymmetricWaveletCoefficientsEarth(sleplet.functions.f_lm.F_LM):
     def _create_wavelet_coefficients(
         self,
     ) -> tuple[npt.NDArray[np.complex_], npt.NDArray[np.complex_]]:
-        """Computes wavelet coefficients of the Earth."""
+        """Computes wavelet coefficients of South America."""
         wavelets = sleplet.wavelet_methods._create_axisymmetric_wavelets(
             self.L,
             self.B,
             self.j_min,
         )
-        self.earth = sleplet.functions.flm.earth.Earth(self.L, smoothing=self.smoothing)
+        self.south_america = sleplet.functions.south_america.SouthAmerica(
+            self.L,
+            smoothing=self.smoothing,
+        )
         wavelet_coefficients = sleplet.wavelet_methods.axisymmetric_wavelet_forward(
             self.L,
-            self.earth.coefficients,
+            self.south_america.coefficients,
             wavelets,
         )
         return wavelets, wavelet_coefficients
