+++ examples/arbitrary/africa/wavelet_reconstruction_africa.py
@@ -1,4 +1,4 @@
-from sleplet.functions.fp import SlepianWaveletCoefficientsAfrica
+from sleplet.functions import SlepianWaveletCoefficientsAfrica
 from sleplet.plotting import PlotSphere
 from sleplet.slepian import Region
 from sleplet.slepian_methods import slepian_inverse
