+++ src/sleplet/functions/elongated_gaussian.py
@@ -5,12 +5,12 @@ from pydantic.dataclasses import dataclass
 import sleplet._string_methods
 import sleplet._validation
 import sleplet._vars
-import sleplet.functions.f_lm
+import sleplet.functions.flm
 import sleplet.harmonic_methods
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class ElongatedGaussian(sleplet.functions.f_lm.F_LM):
+class ElongatedGaussian(sleplet.functions.flm.F_LM):
     """TODO."""
 
     p_sigma: float = 0.1
