+++ src/sleplet/functions/directional_spin_wavelets.py
@@ -8,11 +8,11 @@ from pys2let import pys2let_j_max, wavelet_tiling
 import sleplet
 import sleplet._string_methods
 import sleplet._validation
-import sleplet.functions.f_lm
+import sleplet.functions.flm
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class DirectionalSpinWavelets(sleplet.functions.f_lm.F_LM):
+class DirectionalSpinWavelets(sleplet.functions.flm.F_LM):
     """TODO."""
 
     B: int = 3
