+++ src/sleplet/functions/slepian_identity.py
@@ -4,11 +4,11 @@ from pydantic.dataclasses import dataclass
 
 import sleplet._string_methods
 import sleplet._validation
-import sleplet.functions.f_p
+import sleplet.functions.fp
 
 
 @dataclass(config=sleplet._validation.Validation)
-class SlepianIdentity(sleplet.functions.f_p.F_P):
+class SlepianIdentity(sleplet.functions.fp.F_P):
     """TODO."""
 
     def __post_init_post_parse__(self) -> None:
