+++ src/sleplet/functions/africa.py
@@ -14,7 +14,7 @@ import sleplet.harmonic_methods
 
 
 @dataclass(config=sleplet._validation.Validation)
-class Africa(sleplet.functions.flm.F_LM):
+class Africa(sleplet.functions.flm.FLM):
     """TODO."""
 
     def __post_init_post_parse__(self) -> None:
