+++ src/sleplet/functions/fp.py
@@ -13,7 +13,7 @@ import sleplet.slepian_methods
 
 
 @dataclass(config=sleplet._validation.Validation)
-class F_P(sleplet.functions.coefficients.Coefficients):  # noqa: N801
+class FP(sleplet.functions.coefficients.Coefficients):
     """Abstract parent class to handle Slepian coefficients on the sphere."""
 
     def __post_init_post_parse__(self) -> None:
