+++ src/sleplet/functions/noise_earth.py
@@ -10,7 +10,7 @@ import sleplet.noise
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class NoiseEarth(sleplet.functions.flm.F_LM):
+class NoiseEarth(sleplet.functions.flm.FLM):
     """TODO."""
 
     SNR: float = 10
