+++ src/sleplet/functions/axisymmetric_wavelets.py
@@ -12,7 +12,7 @@ import sleplet.wavelet_methods
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class AxisymmetricWavelets(sleplet.functions.flm.F_LM):
+class AxisymmetricWavelets(sleplet.functions.flm.FLM):
     """TODO."""
 
     B: int = 3
