+++ src/sleplet/slepian/slepian_limit_lat_lon.py
@@ -12,14 +12,14 @@ import sleplet._data.setup_pooch
 import sleplet._mask_methods
 import sleplet._validation
 import sleplet._vars
-import sleplet.slepian.region
-import sleplet.slepian.slepian_functions
+import sleplet.slepian
+from sleplet.slepian.slepian_functions import SlepianFunctions
 
 _data_path = Path(__file__).resolve().parents[1] / "_data"
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class SlepianLimitLatLon(sleplet.slepian.slepian_functions.SlepianFunctions):
+class SlepianLimitLatLon(SlepianFunctions):
     """Class to create a limited latitude longitude Slepian region on the sphere."""
 
     phi_max: float = sleplet._vars.PHI_MAX_DEFAULT
@@ -37,8 +37,8 @@ class SlepianLimitLatLon(sleplet.slepian.slepian_functions.SlepianFunctions):
     def _create_fn_name(self) -> str:
         return f"slepian_{self.region.name_ending}"
 
-    def _create_region(self) -> "sleplet.slepian.region.Region":
-        return sleplet.slepian.region.Region(
+    def _create_region(self) -> "sleplet.slepian.Region":
+        return sleplet.slepian.Region(
             theta_min=self.theta_min,
             theta_max=self.theta_max,
             phi_min=self.phi_min,
