+++ src/sleplet/_mask_methods.py
@@ -9,8 +9,8 @@ import sleplet._data.create_earth_flm
 import sleplet._data.setup_pooch
 import sleplet._vars
 import sleplet.harmonic_methods
-import sleplet.meshes.mesh
-import sleplet.slepian.region
+import sleplet.meshes
+import sleplet.slepian
 
 _data_path = Path(__file__).resolve().parent / "_data"
 
@@ -20,7 +20,7 @@ SOUTH_AMERICA_RANGE = np.deg2rad(40)
 
 def create_mask_region(
     L: int,
-    region: sleplet.slepian.region.Region,
+    region: sleplet.slepian.Region,
 ) -> npt.NDArray[np.float_]:
     """Creates a mask of a region of interested, the output will be based
     on the value of the provided L. The mask could be either:
@@ -74,7 +74,7 @@ def _load_mask(L: int, mask_name: str) -> npt.NDArray[np.float_]:
 def ensure_masked_flm_bandlimited(
     flm: npt.NDArray[np.complex_],
     L: int,
-    region: sleplet.slepian.region.Region,
+    region: sleplet.slepian.Region,
     *,
     reality: bool,
     spin: int,
@@ -98,9 +98,9 @@ def ensure_masked_flm_bandlimited(
     )
 
 
-def create_default_region() -> "sleplet.slepian.region.Region":
+def create_default_region() -> sleplet.slepian.Region:
     """Creates default region."""
-    return sleplet.slepian.region.Region(
+    return sleplet.slepian.Region(
         gap=sleplet.POLAR_GAP,
         mask_name=sleplet.SLEPIAN_MASK,
         phi_max=np.deg2rad(sleplet.PHI_MAX),
@@ -126,7 +126,7 @@ def create_mesh_region(
 
 
 def ensure_masked_bandlimit_mesh_signal(
-    mesh: sleplet.meshes.mesh.Mesh,
+    mesh: "sleplet.meshes.Mesh",
     u_i: npt.NDArray[np.complex_ | np.float_],
 ) -> npt.NDArray[np.float_]:
     """Ensures that signal in pixel space is bandlimited."""
@@ -136,7 +136,7 @@ def ensure_masked_bandlimit_mesh_signal(
 
 
 def convert_region_on_vertices_to_faces(
-    mesh: sleplet.meshes.mesh.Mesh,
+    mesh: "sleplet.meshes.Mesh",
 ) -> npt.NDArray[np.float_]:
     """Converts the region on vertices to faces."""
     region_reshape = np.argwhere(mesh.region).reshape(-1)
