+++ src/sleplet/meshes/mesh_slepian_wavelet_coefficients.py
@@ -7,9 +7,8 @@ from pys2let import pys2let_j_max
 import sleplet
 import sleplet._string_methods
 import sleplet._validation
+import sleplet.meshes
 import sleplet.meshes.mesh_slepian_coefficients
-import sleplet.meshes.mesh_slepian_field
-import sleplet.meshes.mesh_slepian_wavelets
 import sleplet.wavelet_methods
 
 
@@ -55,12 +54,12 @@ class MeshSlepianWaveletCoefficients(
         self,
     ) -> tuple[npt.NDArray[np.float_], npt.NDArray[np.complex_ | np.float_]]:
         """Computes wavelet coefficients in Slepian space."""
-        smw = sleplet.meshes.mesh_slepian_wavelets.MeshSlepianWavelets(
+        smw = sleplet.meshes.MeshSlepianWavelets(
             self.mesh,
             B=self.B,
             j_min=self.j_min,
         )
-        smf = sleplet.meshes.mesh_slepian_field.MeshSlepianField(
+        smf = sleplet.meshes.MeshSlepianField(
             self.mesh,
         )
         wavelets = smw.wavelets
