+++ src/sleplet/meshes/mesh_slepian_functions.py
@@ -5,13 +5,13 @@ from pydantic.dataclasses import dataclass
 
 import sleplet
 import sleplet._validation
-import sleplet.meshes
 import sleplet.slepian_methods
+from sleplet.meshes.mesh_slepian_coefficients import MeshSlepianCoefficients
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class MeshSlepianFunctions(
-    sleplet.meshes.mesh_slepian_coefficients.MeshSlepianCoefficients,
+    MeshSlepianCoefficients,
 ):
     """TODO."""
 
