+++ src/sleplet/meshes/mesh_basis_functions.py
@@ -6,12 +6,12 @@ from pydantic.dataclasses import dataclass
 import sleplet
 import sleplet._validation
 import sleplet.harmonic_methods
-import sleplet.meshes.mesh_harmonic_coefficients
+from sleplet.meshes.mesh_harmonic_coefficients import MeshHarmonicCoefficients
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class MeshBasisFunctions(
-    sleplet.meshes.mesh_harmonic_coefficients.MeshHarmonicCoefficients,
+    MeshHarmonicCoefficients,
 ):
     """TODO."""
 
