+++ src/sleplet/functions/slepian_noise_south_america.py
@@ -6,7 +6,7 @@ import sleplet._string_methods
 import sleplet._validation
 import sleplet.functions
 import sleplet.noise
-import sleplet.slepian
+import sleplet.slepian.region
 from sleplet.functions.fp import FP
 
 
@@ -20,7 +20,7 @@ class SlepianNoiseSouthAmerica(FP):
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
         if (
-            isinstance(self.region, sleplet.slepian.Region)
+            isinstance(self.region, sleplet.slepian.region.Region)
             and self.region.name_ending != "south_america"
         ):
             raise RuntimeError("Slepian region selected must be 'south_america'")
