+++ src/sleplet/functions/slepian_africa.py
@@ -5,7 +5,7 @@ from pydantic.dataclasses import dataclass
 import sleplet._string_methods
 import sleplet._validation
 import sleplet.functions
-import sleplet.slepian
+import sleplet.slepian.region
 import sleplet.slepian_methods
 from sleplet.functions.fp import FP
 
@@ -17,7 +17,7 @@ class SlepianAfrica(FP):
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
         if (
-            isinstance(self.region, sleplet.slepian.Region)
+            isinstance(self.region, sleplet.slepian.region.Region)
             and self.region.name_ending != "africa"
         ):
             raise RuntimeError("Slepian region selected must be 'africa'")
