+++ src/sleplet/functions/slepian_wavelet_coefficients_south_america.py
@@ -8,7 +8,7 @@ import sleplet
 import sleplet._string_methods
 import sleplet._validation
 import sleplet.functions
-import sleplet.slepian
+import sleplet.slepian.region
 import sleplet.wavelet_methods
 from sleplet.functions.fp import FP
 
@@ -27,7 +27,7 @@ class SlepianWaveletCoefficientsSouthAmerica(FP):
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
         if (
-            isinstance(self.region, sleplet.slepian.Region)
+            isinstance(self.region, sleplet.slepian.region.Region)
             and self.region.name_ending != "south_america"
         ):
             raise RuntimeError("Slepian region selected must be 'south_america'")
