+++ src/sleplet/functions/gaussian.py
@@ -5,11 +5,11 @@ from pydantic.dataclasses import dataclass
 
 import sleplet._string_methods
 import sleplet._validation
-from sleplet.functions.flm import FLM
+from sleplet.functions.flm import Flm
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class Gaussian(FLM):
+class Gaussian(Flm):
     """TODO."""
 
     sigma: float = 10
