+++ src/sleplet/functions/axisymmetric_wavelet_coefficients_africa.py
@@ -9,11 +9,11 @@ import sleplet._string_methods
 import sleplet._validation
 import sleplet.functions
 import sleplet.wavelet_methods
-from sleplet.functions.flm import FLM
+from sleplet.functions.flm import Flm
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
-class AxisymmetricWaveletCoefficientsAfrica(FLM):
+class AxisymmetricWaveletCoefficientsAfrica(Flm):
     """TODO."""
 
     B: int = 3
