+++ src/sleplet/harmonic_methods.py
@@ -10,7 +10,7 @@ from numpy.random import Generator
 import sleplet._data.create_earth_flm
 import sleplet._integration_methods
 import sleplet._vars
-import sleplet.meshes
+import sleplet.meshes.mesh
 
 AFRICA_ALPHA = np.deg2rad(44)
 AFRICA_BETA = np.deg2rad(87)
@@ -143,7 +143,7 @@ def compute_random_signal(
 
 
 def mesh_forward(
-    mesh: sleplet.meshes.Mesh,
+    mesh: sleplet.meshes.mesh.Mesh,
     u: npt.NDArray[np.complex_ | np.float_],
 ) -> npt.NDArray[np.float_]:
     """TODO computes the mesh forward transform from real space to harmonic space.
@@ -169,7 +169,7 @@ def mesh_forward(
 
 
 def mesh_inverse(
-    mesh: sleplet.meshes.Mesh,
+    mesh: sleplet.meshes.mesh.Mesh,
     u_i: npt.NDArray[np.complex_ | np.float_],
 ) -> npt.NDArray[np.complex_ | np.float_]:
     """TODO computes the mesh inverse transform from harmonic space to real space.
