+++ src/sleplet/_slepian_arbitrary_methods.py
@@ -4,7 +4,7 @@ import numpy as np
 from numpy import typing as npt
 
 if TYPE_CHECKING:
-    import sleplet.meshes
+    import sleplet.meshes.mesh
 
 MACHINE_EPSILON = 1e-14
 
@@ -34,7 +34,7 @@ def clean_evals_and_evecs(
     return eigenvalues, eigenvectors
 
 
-def compute_mesh_shannon(mesh: "sleplet.meshes.Mesh") -> int:
+def compute_mesh_shannon(mesh: "sleplet.meshes.mesh.Mesh") -> int:
     """Computes the effective Shannon number for a region of a mesh."""
     num_basis_fun = mesh.mesh_eigenvalues.shape[0]
     region_vertices = mesh.region.sum()
