+++ src/sleplet/slepian/slepian_functions.py
@@ -6,8 +6,7 @@ from pydantic.dataclasses import dataclass
 
 import sleplet
 import sleplet._validation
-
-from . import Region
+import sleplet.slepian.region
 
 
 @dataclass(config=sleplet._validation.Validation)
@@ -35,7 +34,7 @@ class SlepianFunctions:
         raise NotImplementedError
 
     @abstractmethod
-    def _create_region(self) -> Region:
+    def _create_region(self) -> "sleplet.slepian.region.Region":
         """Creates a region object for area of interest."""
         raise NotImplementedError
 
