+++ src/sleplet/functions/fp/slepian_noise_africa.py
@@ -1,60 +1,0 @@
-import numpy as np
-from numpy import typing as npt
-from pydantic.dataclasses import dataclass
-
-import sleplet._string_methods
-import sleplet._validation
-import sleplet.functions.f_p
-import sleplet.functions.fp.slepian_africa
-import sleplet.noise
-import sleplet.slepian.region
-
-
-@dataclass(config=sleplet._validation.Validation, kw_only=True)
-class SlepianNoiseAfrica(sleplet.functions.f_p.F_P):
-    """TODO."""
-
-    SNR: float = -10
-    """TODO"""
-
-    def __post_init_post_parse__(self) -> None:
-        super().__post_init_post_parse__()
-        if (
-            isinstance(self.region, sleplet.slepian.region.Region)
-            and self.region.name_ending != "africa"
-        ):
-            raise RuntimeError("Slepian region selected must be 'africa'")
-
-    def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
-        sa = sleplet.functions.fp.slepian_africa.SlepianAfrica(
-            self.L,
-            region=self.region,
-            smoothing=self.smoothing,
-        )
-        noise = sleplet.noise._create_slepian_noise(
-            self.L,
-            sa.coefficients,
-            self.slepian,
-            self.SNR,
-        )
-        sleplet.noise.compute_snr(sa.coefficients, noise, "Slepian")
-        return noise
-
-    def _create_name(self) -> str:
-        return (
-            f"{sleplet._string_methods._convert_camel_case_to_snake_case(self.__class__.__name__)}"
-            f"{sleplet._string_methods.filename_args(self.SNR, 'snr')}"
-        )
-
-    def _set_reality(self) -> bool:
-        return False
-
-    def _set_spin(self) -> int:
-        return 0
-
-    def _setup_args(self) -> None:
-        if isinstance(self.extra_args, list):
-            num_args = 1
-            if len(self.extra_args) != num_args:
-                raise ValueError(f"The number of extra arguments should be {num_args}")
-            self.SNR = self.extra_args[0]
