+++ examples/misc/denoising_axisym_africa.py
@@ -1,7 +1,7 @@
 import sys
 from pathlib import Path
 
-from sleplet.functions.flm import Africa, AxisymmetricWavelets
+from sleplet.functions import Africa, AxisymmetricWavelets
 from sleplet.plot_methods import find_max_amplitude
 from sleplet.plotting import PlotSphere
 
@@ -18,7 +18,7 @@ SNR_IN = 10
 
 
 def main() -> None:
-    """contrast denosiing with an Earth map versus Africa map."""
+    """Contrast denosiing with an Earth map versus Africa map."""
     # create map & noised map
     fun = Africa(L)
     fun_noised = Africa(L, noise=SNR_IN)
