+++ examples/wavelets/axisymmetric_wavelet_covariance.py
@@ -27,7 +27,8 @@ def _compute_wavelet_covariance(
 
 
 def _is_ergodic(j_min: int, *, j: int = 0) -> bool:
-    """Computes whether the function follows ergodicity.
+    """
+    Computes whether the function follows ergodicity.
 
     ergodicity fails for J_min = 0, because the scaling function will only
     cover f00. Hence <flm flm*> will be 0 in that case and the scaling
@@ -46,7 +47,8 @@ def axisymmetric_wavelet_covariance(
     runs: int = 100,
     var_flm: float = 1,
 ) -> None:
-    """Compute theoretical covariance of wavelet coefficients.
+    """
+    Compute theoretical covariance of wavelet coefficients.
 
     the covariance <Wj(omega)Wj*(omega)> is given by the following expression:
     sigma^2 Sum(l,0) |Psi^j_l0|^2
