+++ examples/polar_cap/eigenfunctions.py
@@ -29,7 +29,8 @@ ZENODO_DATA_DOI = "10.5281/zenodo.7767698"
 
 
 def main() -> None:
-    """Create fig 5.1 from Spatiospectral Concentration on a Sphere
+    """
+    Create fig 5.1 from Spatiospectral Concentration on a Sphere
     by Simons et al 2006.
     """
     x = np.linspace(THETA_MIN_DEFAULT, np.rad2deg(THETA_MAX_DEFAULT), RESOLUTION + 1)
