+++ src/sleplet/slepian/slepian_limit_lat_lon.py
@@ -23,13 +23,13 @@ class SlepianLimitLatLon(SlepianFunctions):
     """Class to create a limited latitude longitude Slepian region on the sphere."""
 
     phi_max: float = sleplet._vars.PHI_MAX_DEFAULT
-    """TODO"""
+    """For a limited latitude longitude region."""
     phi_min: float = sleplet._vars.PHI_MIN_DEFAULT
-    """TODO"""
+    """For a limited latitude longitude region."""
     theta_max: float = sleplet._vars.THETA_MAX_DEFAULT
-    """TODO"""
+    """For a limited latitude longitude region."""
     theta_min: float = sleplet._vars.THETA_MIN_DEFAULT
-    """TODO"""
+    """For a limited latitude longitude region."""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
