+++ src/sleplet/functions/slepian_south_america.py
@@ -11,7 +11,10 @@ from sleplet.functions.fp import Fp
 
 @dataclass(config=sleplet._validation.Validation)
 class SlepianSouthAmerica(Fp):
-    """TODO."""
+    """
+    Creates a Slepian region on the topographic map of the Earth of the
+    South America continent.
+    """
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
