+++ src/sleplet/functions/south_america.py
@@ -15,7 +15,7 @@ from sleplet.functions.flm import Flm
 
 @dataclass(config=sleplet._validation.Validation)
 class SouthAmerica(Flm):
-    """TODO."""
+    """Creates the South America region of the topographic map of the Earth."""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
