+++ src/sleplet/functions/axisymmetric_wavelet_coefficients_south_america.py
@@ -14,14 +14,18 @@ from sleplet.functions.flm import Flm
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class AxisymmetricWaveletCoefficientsSouthAmerica(Flm):
-    """TODO."""
+    """
+    Creates the axisymmetric wavelet coefficients of the South America
+    region.
+    """
 
     B: int = 3
-    """TODO"""
+    r"""The wavelet parameter. Represented as \(\lambda\) in the papers."""
     j_min: int = 2
-    """TODO"""
+    """The minimum wavelet scale."""
     j: int | None = None
-    """TODO"""
+    """Option to select a given wavelet. `None` indicates the scaling function,
+    whereas `0` would correspond to the selected `j_min`."""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
