+++ src/sleplet/functions/axisymmetric_wavelets.py
@@ -13,14 +13,18 @@ from sleplet.functions.flm import Flm
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class AxisymmetricWavelets(Flm):
-    """TODO."""
+    """
+    Creates scale-discretised axisymmetric wavelets on the sphere. As
+    seen in <https://doi.org/10.1051/0004-6361/201220729>.
+    """
 
     B: int = 3
-    """TODO"""
+    r"""The wavelet parameter. Represented as \(\lambda\) in the papers."""
     j_min: int = 2
-    """TODO"""
+    """The minimum wavelet scale."""
     j: int | None = None
-    """TODO"""
+    """Option to select a given wavelet. `None` indicates the scaling function,
+    whereas `0` would correspond to the selected `j_min`."""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
