### 2023-04-02 ea42e77
Correct some docstrings
