+++ src/sleplet/wavelet_methods.py
@@ -14,7 +14,7 @@ def slepian_wavelet_forward(
     shannon: int,
 ) -> npt.NDArray[np.complex_ | np.float_]:
     """
-    TODO computes the coefficients of the given tiling function in Slepian space.
+    Computes the coefficients of the given tiling function in Slepian space.
 
     Args:
     ----
@@ -42,7 +42,7 @@ def slepian_wavelet_inverse(
     shannon: int,
 ) -> npt.NDArray[np.complex_ | np.float_]:
     """
-    TODO computes the inverse wavelet transform in Slepian space.
+    Computes the inverse wavelet transform in Slepian space.
 
     Args:
     ----
@@ -70,7 +70,7 @@ def axisymmetric_wavelet_forward(
     wavelets: npt.NDArray[np.complex_],
 ) -> npt.NDArray[np.complex_]:
     """
-    TODO computes the coefficients of the axisymmetric wavelets.
+    Computes the coefficients of the axisymmetric wavelets.
 
     Args:
     ----
@@ -98,7 +98,7 @@ def axisymmetric_wavelet_inverse(
     wavelets: npt.NDArray[np.complex_],
 ) -> npt.NDArray[np.complex_]:
     """
-    TODO computes the inverse axisymmetric wavelet transform.
+    Computes the inverse axisymmetric wavelet transform.
 
     Args:
     ----
@@ -137,7 +137,7 @@ def _create_axisymmetric_wavelets(
 
 def create_kappas(xlim: int, B: int, j_min: int) -> npt.NDArray[np.float_]:
     """
-    TODO computes the Slepian wavelets.
+    Computes the Slepian wavelets.
 
     Args:
     ----
@@ -159,7 +159,7 @@ def find_non_zero_wavelet_coefficients(
     axis: int | tuple[int, ...],
 ) -> npt.NDArray[np.complex_ | np.float_]:
     """
-    TODO finds the coefficients within the shannon number to speed up computations.
+    Finds the coefficients within the shannon number to speed up computations.
 
     Args:
     ----
