### 2023-04-02 74fe865
Add docstring to all public methods
