+++ src/sleplet/functions/slepian_noise_africa.py
@@ -1,3 +1,4 @@
+"""Contains the `SlepianNoiseAfrica` class."""
 import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
