+++ src/sleplet/functions/south_america.py
@@ -1,3 +1,4 @@
+"""Contains the `SouthAmerica` class."""
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
