+++ src/sleplet/slepian/slepian_functions.py
@@ -1,3 +1,4 @@
+"""Contains the abstract `SlepianFunctions` class."""
 from abc import abstractmethod
 
 import numpy as np
