+++ src/sleplet/meshes/mesh_slepian_wavelet_coefficients.py
@@ -1,3 +1,4 @@
+"""Contains the `MeshSlepianWaveletCoefficients` class."""
 import numpy as np
 from numpy import typing as npt
 from pydantic import validator
