+++ src/sleplet/functions/directional_spin_wavelets.py
@@ -1,3 +1,4 @@
+"""Contains the `DirectionalSpinWavelets` class."""
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
