+++ src/sleplet/functions/axisymmetric_wavelet_coefficients_earth.py
@@ -1,3 +1,4 @@
+"""Contains the `AxisymmetricWaveletCoefficientsEarth` class."""
 import numpy as np
 from numpy import typing as npt
 from pydantic import validator
