### 2023-04-02 1d6b3a4
Only specify figure path once
