+++ src/sleplet/functions/slepian_noise_africa.py
@@ -19,7 +19,7 @@ class SlepianNoiseAfrica(Fp):
     """
 
     SNR: float = -10
-    """A parameter which controls the level of signal to noise in the noised
+    """A parameter which controls the level of signal-to-noise in the noised
     data."""
 
     def __post_init_post_parse__(self) -> None:
