+++ src/sleplet/meshes/mesh_noise_field.py
@@ -15,7 +15,7 @@ class MeshNoiseField(MeshHarmonicCoefficients):
     """Creates a noised per-vertex normals field on a given mesh."""
 
     SNR: float = 10
-    """A parameter which controls the level of signal to noise in the noised
+    """A parameter which controls the level of signal-to-noise in the noised
     data."""
 
     def __post_init_post_parse__(self) -> None:
