### 2023-04-02 41af835
Fix some docstrings
