+++ src/sleplet/meshes/mesh_slepian_wavelets.py
@@ -14,12 +14,12 @@ from sleplet.meshes.mesh_slepian_coefficients import MeshSlepianCoefficients
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class MeshSlepianWavelets(MeshSlepianCoefficients):
-    """Creates the Slepian wavelets of a given mesh."""
+    """Creates Slepian wavelets of a given mesh."""
 
     B: int = 3
     r"""The wavelet parameter. Represented as \(\lambda\) in the papers."""
     j_min: int = 2
-    """The minimum wavelet scale."""
+    r"""The minimum wavelet scale. Represented as \(J_{0}\) in the papers."""
     j: int | None = None
     """Option to select a given wavelet. `None` indicates the scaling function,
     whereas `0` would correspond to the selected `j_min`."""
@@ -50,7 +50,7 @@ class MeshSlepianWavelets(MeshSlepianCoefficients):
             self.B, self.j_min, self.j = self.extra_args
 
     def _create_wavelets(self) -> npt.NDArray[np.float_]:
-        """Creates the Slepian wavelets of the mesh."""
+        """Creates Slepian wavelets of the mesh."""
         return sleplet.wavelet_methods.create_kappas(
             self.mesh.mesh_eigenvalues.shape[0],
             self.B,
