+++ src/sleplet/functions/axisymmetric_wavelets.py
@@ -22,7 +22,7 @@ class AxisymmetricWavelets(Flm):
     B: int = 3
     r"""The wavelet parameter. Represented as \(\lambda\) in the papers."""
     j_min: int = 2
-    """The minimum wavelet scale."""
+    r"""The minimum wavelet scale. Represented as \(J_{0}\) in the papers."""
     j: int | None = None
     """Option to select a given wavelet. `None` indicates the scaling function,
     whereas `0` would correspond to the selected `j_min`."""
