+++ src/sleplet/functions/slepian_identity.py
@@ -10,7 +10,7 @@ from sleplet.functions.fp import Fp
 
 @dataclass(config=sleplet._validation.Validation)
 class SlepianIdentity(Fp):
-    """Creates the identify function on ther sphere in the Slepian region."""
+    """Creates an identify function on the sphere in the Slepian region."""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
