+++ src/sleplet/functions/gaussian.py
@@ -11,10 +11,7 @@ from sleplet.functions.flm import Flm
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class Gaussian(Flm):
-    r"""
-    Creates a Gaussian on the sphere
-    \(\exp(-\frac{{\ell}^{2}}{2\sigma^{2}})\).
-    """
+    r"""Creates a Gaussian \(\exp(-\frac{{\ell}^{2}}{2\sigma^{2}})\)."""
 
     sigma: float = 10
     r"""Sets the \(\sigma\) value."""
