+++ src/sleplet/functions/slepian_dirac_delta.py
@@ -14,7 +14,7 @@ from sleplet.functions.fp import Fp
 
 @dataclass(config=sleplet._validation.Validation)
 class SlepianDiracDelta(Fp):
-    """Createa a Dirac delta on the sphere of the Slepian coefficients."""
+    """Createa a Dirac delta of the Slepian coefficients."""
 
     def __post_init_post_parse__(self) -> None:
         super().__post_init_post_parse__()
