+++ src/sleplet/functions/slepian_wavelets.py
@@ -14,7 +14,7 @@ from sleplet.functions.fp import Fp
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
 class SlepianWavelets(Fp):
-    """Creats the Slepian wavelets on the sphere."""
+    """Creats the Slepian wavelets."""
 
     B: int = 3
     r"""The wavelet parameter. Represented as \(\lambda\) in the papers."""
