+++ src/sleplet/plotting/_create_plot_mesh.py
@@ -50,7 +50,7 @@ class PlotMesh:
             self.filename += "_norm"
 
     def execute(self) -> None:
-        """Creates 3d plotly mesh plot."""
+        """Performs the plot."""
         vmin, vmax = self.f.min(), self.f.max()
         f = self._prepare_field(self.f)
 
