### 2023-04-02 #5221 @paddyroddy
I've added argument/function level description of the API https://astro-informatics.github.io/sleplet/index.html
