+++ CONTRIBUTING.md
@@ -24,11 +24,6 @@ forward to your contributions. 🎉
 - [I Want To Contribute](#i-want-to-contribute)
   - [Reporting Bugs](#reporting-bugs)
   - [Suggesting Enhancements](#suggesting-enhancements)
-  - [Your First Code Contribution](#your-first-code-contribution)
-  - [Improving The Documentation](#improving-the-documentation)
-- [Styleguides](#styleguides)
-  - [Commit Messages](#commit-messages)
-- [Join The Project Team](#join-the-project-team)
 
 ## I Have a Question
 
@@ -116,11 +111,11 @@ Once it's filed:
   the `needs-repro` tag will not be addressed until they are reproduced.
 - If the team is able to reproduce the issue, it will be marked `needs-fix`, as
   well as possibly other tags (such as `critical`), and the issue will be left to
-  be [implemented by someone](#your-first-code-contribution).
+  be implemented by someone.
 
 ### Suggesting Enhancements
 
-This section guides you through submitting an enhancement suggestion for SLEPLET,
+This section guides you through submitting an enhancement suggestion for `SLEPLET`,
 **including completely new features and minor improvements to existing**
 **functionality**. Following these guidelines will help maintainers and the
 community to understand your suggestion and find related suggestions.
@@ -140,8 +135,6 @@ community to understand your suggestion and find related suggestions.
   to the majority of our users and not just a small subset. If you're just
   targeting a minority of users, consider writing an add-on/plugin library.
 
-<!-- omit in toc -->
-
 #### How Do I Submit a Good Enhancement Suggestion?
 
 Enhancement suggestions are tracked as
@@ -150,9 +143,9 @@ Enhancement suggestions are tracked as
 - Use a **clear and descriptive title** for the issue to identify the suggestion.
 - Provide a **step-by-step description of the suggested enhancement** in as many
   details as possible.
-- **Describe the current behavior** and **explain which behavior you expected**
+- **Describe the current behaviour** and **explain which behaviour you expected**
   **to see instead** and why. At this point you can also tell which alternatives
   do not work for you.
-- **Explain why this enhancement would be useful** to most SLEPLET users. You
+- **Explain why this enhancement would be useful** to most `SLEPLET` users. You
   may also want to point out the other projects that solved it better and which
   could serve as inspiration.
