+++ CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 First off, thanks for taking the time to contribute! ⭐
 
-All types of contributions are encouraged and valued. See the
+All types of contributions are encouraged and valued. See the rest of this
 [Table of Contents](#table-of-contents) for different ways to help and details
 about how this project handles them. Please make sure to read the relevant
 section before making your contribution. It will make it a lot easier for us
@@ -28,7 +28,7 @@ forward to your contributions. 🎉
 ## I Have a Question
 
 > If you want to ask a question, we assume that you have read the available
-> [Documentation](https://astro-informatics.github.io/sleplet/).
+> [Documentation](https://astro-informatics.github.io/sleplet).
 
 Before you ask a question, it is best to search for existing
 [Issues](https://github.com/astro-informatics/sleplet/issues) that might help
@@ -66,7 +66,7 @@ steps in advance to help us fix any potential bug as fast as possible.
 - Make sure that you are using the latest version.
 - Determine if your bug is really a bug and not an error on your side e.g. using
   incompatible environment components/versions (Make sure that you have read the
-  [documentation](https://astro-informatics.github.io/sleplet/). If you are
+  [documentation](https://astro-informatics.github.io/sleplet). If you are
   looking for support, you might want to check [this section](#i-have-a-question)).
 - To see if other users have experienced (and potentially already solved) the
   same issue you are having, check if there is not already a bug report existing
@@ -123,7 +123,7 @@ community to understand your suggestion and find related suggestions.
 #### Before Submitting an Enhancement
 
 - Make sure that you are using the latest version.
-- Read the [documentation](https://astro-informatics.github.io/sleplet/)
+- Read the [documentation](https://astro-informatics.github.io/sleplet)
   carefully and find out if the functionality is already covered, maybe by an
   individual configuration.
 - Perform a [search](https://github.com/astro-informatics/sleplet/issues) to see
