+++ CONTRIBUTING.md
@@ -1,0 +1,151 @@
+# Contributing to SLEPLET
+
+First off, thanks for taking the time to contribute! ⭐
+
+All types of contributions are encouraged and valued. See the rest of this
+[Table of Contents](#table-of-contents) for different ways to help and details
+about how this project handles them. Please make sure to read the relevant
+section before making your contribution. It will make it a lot easier for us
+maintainers and smooth out the experience for all involved. The community looks
+forward to your contributions. 🎉
+
+> And if you like the project, but just don't have time to contribute, that's
+> fine. There are other easy ways to support the project and show your
+> appreciation, which we would also be very happy about:
+>
+> - Star the project
+> - Tweet about it
+> - Refer this project in your project's readme
+> - Mention the project at local meetups and tell your friends/colleagues
+
+## Table of Contents
+
+- [I Have a Question](#i-have-a-question)
+- [I Want To Contribute](#i-want-to-contribute)
+  - [Reporting Bugs](#reporting-bugs)
+  - [Suggesting Enhancements](#suggesting-enhancements)
+
+## I Have a Question
+
+> If you want to ask a question, we assume that you have read the available
+> [Documentation](https://astro-informatics.github.io/sleplet).
+
+Before you ask a question, it is best to search for existing
+[Issues](https://github.com/astro-informatics/sleplet/issues) that might help
+you. In case you have found a suitable issue and still need clarification, you
+can write your question in this issue. It is also advisable to search the
+internet for answers first.
+
+If you then still feel the need to ask a question and need clarification, we
+recommend the following:
+
+- Open an [Issue](https://github.com/astro-informatics/sleplet/issues/new).
+- Provide as much context as you can about what you're running into.
+- Provide project and platform versions (python etc.), depending on what seems
+  relevant.
+
+We will then take care of the issue as soon as possible.
+
+## I Want To Contribute
+
+> ### Legal Notice
+>
+> When contributing to this project, you must agree that you have authored 100%
+> of the content, that you have the necessary rights to the content and that the
+> content you contribute may be provided under the project licence.
+
+### Reporting Bugs
+
+#### Before Submitting a Bug Report
+
+A good bug report shouldn't leave others needing to chase you up for more
+information. Therefore, we ask you to investigate carefully, collect information
+and describe the issue in detail in your report. Please complete the following
+steps in advance to help us fix any potential bug as fast as possible.
+
+- Make sure that you are using the latest version.
+- Determine if your bug is really a bug and not an error on your side e.g. using
+  incompatible environment components/versions (Make sure that you have read the
+  [documentation](https://astro-informatics.github.io/sleplet). If you are
+  looking for support, you might want to check [this section](#i-have-a-question)).
+- To see if other users have experienced (and potentially already solved) the
+  same issue you are having, check if there is not already a bug report existing
+  for your bug or error in the
+  [bug tracker](https://github.com/astro-informatics/sleplet/issues?q=label%3Abug).
+- Also make sure to search the internet (including Stack Overflow) to see if
+  users outside the GitHub community have discussed the issue.
+- Collect information about the bug:
+  - Stack trace (Traceback)
+  - OS, Platform and Version (Windows, Linux, macOS, x86, ARM)
+  - Version of the interpreter, compiler, SDK, runtime environment, package
+    manager, depending on what seems relevant.
+  - Possibly your input and the output
+  - Can you reliably reproduce the issue? And can you also reproduce it with
+    older versions?
+
+#### How Do I Submit a Good Bug Report?
+
+> You must never report security related issues, vulnerabilities or bugs
+> including sensitive information to the issue tracker, or elsewhere in public.
+> Instead, sensitive bugs must be sent by email to <patrickjamesroddy@gmail.com>.
+
[…]
