+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,21 @@
+---
+name: Feature Request
+about: Suggest a Way to Improve This Project
+title: "[FEATURE]"
+labels: enhancement
+assignees: paddyroddy
+---
+
+**Is Your Feature Request Related to a Problem? Please Describe.**
+A clear and concise description of what the problem is. Ex. I'm always
+frustrated when [...]
+
+**Describe the Solution You'd Like**
+A clear and concise description of what you want to happen.
+
+**Describe Alternatives You've Considered**
+A clear and concise description of any alternative solutions or features you've
+considered.
+
+**Additional Context**
+Add any other context or screenshots about the feature request here.
