### 2023-04-02 9806552
Add comparisons to other codes
