+++ paper/paper.bib
@@ -1,3 +1,13 @@
+@article{Bates2017,
+  author  = {Bates, Alice P. and Khalid, Zubair and Kennedy, Rodney A.},
+  title   = {{Slepian Spatial-Spectral Concentration Problem on the Sphere: Analytical Formulation for Limited Colatitude-Longitude Spatial Region}},
+  year    = 2017,
+  journal = {IEEE Transactions on Signal Processing},
+  volume  = 65,
+  number  = 6,
+  pages   = {1527--1537},
+  doi     = {10.1109/TSP.2016.2646668}
+}
 @article{Landau1961,
   title   = {{Prolate spheroidal wave functions, fourier analysis and uncertainty -- II}},
   author  = {Landau, H. J. and Pollak, H. O.},
@@ -46,11 +56,6 @@
   pages   = {5876--5887},
   doi     = {10.1109/TSP.2011.2166394}
 }
-@software{RoddySLEPLET,
-  title  = {{SLEPLET}},
-  author = {Roddy, Patrick J.},
-  doi    = {10.5281/zenodo.7268074},
-}
 @article{Roddy2021,
   title   = {{Sifting Convolution on the Sphere}},
   author  = {Roddy, Patrick J. and McEwen, Jason D.},
@@ -82,6 +87,21 @@
   year      = 2023,
   publisher = {arXiv},
   url       = {https://arxiv.org/abs/2302.06006}
+}
+@software{RoddySLEPLET,
+  title  = {{SLEPLET}},
+  author = {Roddy, Patrick J.},
+  doi    = {10.5281/zenodo.7268074}
+}
+@article{Simons2006,
+  title   = {{Spatiospectral Concentration on a Sphere}},
+  author  = {Simons, Frederik J. and Dahlen, F. A. and Wieczorek, Mark A.},
+  year    = 2006,
+  journal = {SIAM Review},
+  volume  = 48,
+  number  = 3,
+  pages   = {504--536},
+  doi     = {10.1137/S0036144504445765}
 }
 @software{Simons2020,
   author = {Frederik J Simons and Chris Harig and Alain Plattner and Max von Hippel and Albert},
