### 2023-04-02 #5221 @paddyroddy
> > On closer inspection `ShTools` only computes the Slepian functions of a polar cap https://shtools.github.io/SHTOOLS/pysphericalcapcoef.html which is much easier to compute. Whereas `SLEPLET` handles any arbitrary mask. Further, it works for mesh data rather than just the sphere.
> > `slepian_alpha` is a huge set of MATLAB code with no documentation or anything. I believe it's largely used to create the various plots on this paper, https://arxiv.org/pdf/math/0408424.pdf. Again, it is defined only for the sphere.
> > In an ideal world, `SLEPLET` would have relied on another library to generate the Slepian functions and the build the wavelets from them. However, this proved non-trivial to use and, as such, it was necessary to compute them myself. My wording should have been clearer.
>
> Thank you for clarifying. I guess it would be good, if some of these details could also be added to the paper in order to make the differences to the `sleplet` package clear.

I've added some comments to explain this.
