+++ README.md
@@ -15,6 +15,28 @@ installation, `SLEPLET` comes with two command line interfaces - `sphere` and
 `mesh` - which allows one to easily generate plots on the sphere and a set of
 meshes using `plotly`.
 
+## Citing
+
+If you use `SLEPLET` in your research, please cite the paper.
+
+```bibtex
+@article{SLEPLET2023,
+  author    = {Roddy, Patrick J.},
+  title     = {{SLEPLET: Slepian scale-discretised wavelets in Python}},
+  journal   = {Journal of Open Source Software},
+  volume    = ,
+  number    = ,
+  pages     = ,
+  year      = 2023,
+  doi       = {},
+}
+```
+
+Please also cite [S2LET](https://doi.org/10.1051/0004-6361/201220729) upon which
+`SLEPLET` is built, along with [SSHT](https://doi.org/10.1109/TSP.2011.2166394) in
+the spherical setting or [libigl](https://doi.org/10.1145/3134472.3134497) in
+mesh setting.
+
 ## Installation
 
 The recommended way to install `SLEPLET` is via
@@ -71,28 +93,6 @@ The default region is the `south_america` arbitrary region.
   - for a Slepian `limited latitude longitude region`
 - `SLEPIAN_MASK`
   - for an arbitrary Slepian region, currently `africa`/`south_america` supported
-
-## Citing
-
-If you use `SLEPLET` in your research, please cite the paper.
-
-```bibtex
-@article{SLEPLET2023,
-  author    = {Roddy, Patrick J.},
-  title     = {{SLEPLET: Slepian scale-discretised wavelets in Python}},
-  journal   = {Journal of Open Source Software},
-  volume    = ,
-  number    = ,
-  pages     = ,
-  year      = 2023,
-  doi       = {},
-}
-```
-
-Please also cite [S2LET](https://doi.org/10.1051/0004-6361/201220729) upon which
-`SLEPLET` is built, along with [SSHT](https://doi.org/10.1109/TSP.2011.2166394) in
-the spherical setting or [libigl](https://doi.org/10.1145/3134472.3134497) in
-mesh setting.
 
 ## Paper Figures
 
