### 2023-04-03 9b7138a
Add citation after installation
