### 2023-04-03 b014495
Move citation before figures
