### 2023-04-03 7805b87
Check env var at run time
