### 2023-04-03 aed9300
Fix ncpu hack
