+++ src/sleplet/slepian/slepian_arbitrary.py
@@ -120,7 +120,8 @@ class SlepianArbitrary(SlepianFunctions):
             sleplet._parallel_methods.free_shared_memory(shm_r_int, shm_i_int)
 
         # split up L range to maximise effiency
-        ncpu = sleplet.logger.info(f"Number of CPU={int(os.getenv('NCPU', '4'))}")
+        ncpu = int(os.getenv("NCPU", "4"))
+        sleplet.logger.info(f"Number of CPU={ncpu}")
         chunks = sleplet._parallel_methods.split_arr_into_chunks(
             self.L**2,
             ncpu,
