+++ src/sleplet/meshes/mesh_slepian.py
@@ -95,7 +95,8 @@ class MeshSlepian:
             sleplet._parallel_methods.free_shared_memory(shm_int)
 
         # split up L range to maximise effiency
-        ncpu = sleplet.logger.info(f"Number of CPU={int(os.getenv('NCPU', '4'))}")
+        ncpu = int(os.getenv("NCPU", "4"))
+        sleplet.logger.info(f"Number of CPU={ncpu}")
         chunks = sleplet._parallel_methods.split_arr_into_chunks(
             self.mesh.mesh_eigenvalues.shape[0],
             ncpu,
