### 2023-04-03 62af177
Add version to CLI
