+++ src/sleplet/slepian/slepian_arbitrary.py
@@ -1,5 +1,4 @@
 """Contains the `SlepianArbitrary` class."""
-import os
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import KW_ONLY
 from pathlib import Path
@@ -18,6 +17,7 @@ import sleplet._mask_methods
 import sleplet._parallel_methods
 import sleplet._slepian_arbitrary_methods
 import sleplet._validation
+import sleplet._vars
 import sleplet.harmonic_methods
 import sleplet.slepian.region
 from sleplet.slepian.slepian_functions import SlepianFunctions
@@ -120,15 +120,14 @@ class SlepianArbitrary(SlepianFunctions):
             sleplet._parallel_methods.free_shared_memory(shm_r_int, shm_i_int)
 
         # split up L range to maximise effiency
-        ncpu = int(os.getenv("NCPU", "4"))
-        sleplet.logger.info(f"Number of CPU={ncpu}")
+        sleplet.logger.info(f"Number of CPU={sleplet._vars.NCPU}")
         chunks = sleplet._parallel_methods.split_arr_into_chunks(
             self.L**2,
-            ncpu,
+            sleplet._vars.NCPU,
         )
 
         # initialise pool and apply function
-        with ThreadPoolExecutor(max_workers=ncpu) as e:
+        with ThreadPoolExecutor(max_workers=sleplet._vars.NCPU) as e:
             e.map(func, chunks)
 
         # retrieve from parallel function
