+++ src/sleplet/slepian/_slepian_decomposition.py
@@ -9,13 +9,13 @@ import sleplet
 import sleplet._integration_methods
 import sleplet._validation
 import sleplet._vars
-import sleplet.slepian.slepian_functions
+from sleplet.slepian.slepian_functions import SlepianFunctions
 
 
 @dataclass(config=sleplet._validation.Validation)
 class SlepianDecomposition:
     L: int
-    slepian: sleplet.slepian.slepian_functions.SlepianFunctions
+    slepian: SlepianFunctions
     _: KW_ONLY
     f: npt.NDArray[np.complex_] | None = None
     flm: npt.NDArray[np.complex_ | np.float_] | None = None
