+++ tests/test_arbitrary_region.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from numpy.testing import assert_allclose
 
-from sleplet.slepian import SlepianArbitrary
+import sleplet.slepian
 
 
 @pytest.mark.slow()
@@ -12,7 +12,7 @@ def test_equality_to_polar_cap_method(slepian_polar_cap) -> None:
     in comparison to the smarter Slepian polar cap method.
     """
     mask_name = slepian_polar_cap.region.name_ending
-    slepian = SlepianArbitrary(slepian_polar_cap.L, mask_name)
+    slepian = sleplet.slepian.SlepianArbitrary(slepian_polar_cap.L, mask_name)
     assert_allclose(np.abs(slepian_polar_cap.N - slepian.N), 0, atol=3)
     assert_allclose(
         np.abs(
@@ -39,7 +39,7 @@ def test_equality_to_lim_lat_lon_method(slepian_lim_lat_lon) -> None:
     in comparison to the smarter Slepian lim lat lon method.
     """
     mask_name = slepian_lim_lat_lon.region.name_ending
-    slepian = SlepianArbitrary(slepian_lim_lat_lon.L, mask_name)
+    slepian = sleplet.slepian.SlepianArbitrary(slepian_lim_lat_lon.L, mask_name)
     assert_allclose(np.abs(slepian_lim_lat_lon.N - slepian.N), 0, atol=1)
     assert_allclose(
         np.abs(
