+++ tests/conftest.py
@@ -3,11 +3,11 @@ import pytest
 from numpy import typing as npt
 from numpy.random import default_rng
 
+import sleplet.functions
+import sleplet.meshes
+import sleplet.slepian
 from sleplet._vars import RANDOM_SEED
-from sleplet.functions import Earth, SlepianDiracDelta, SlepianWavelets, SouthAmerica
 from sleplet.harmonic_methods import compute_random_signal
-from sleplet.meshes import Mesh, MeshField, MeshSlepian, MeshSlepianWavelets
-from sleplet.slepian import SlepianArbitrary, SlepianLimitLatLon, SlepianPolarCap
 
 ARRAY_DIM = 10
 L = 16
@@ -20,15 +20,15 @@ THETA_MAX = 2 * np.pi / 9
 
 
 @pytest.fixture(scope="session")
-def slepian_polar_cap() -> SlepianPolarCap:
+def slepian_polar_cap() -> sleplet.slepian.SlepianPolarCap:
     """Create a Slepian polar cap class."""
-    return SlepianPolarCap(L, THETA_MAX)
+    return sleplet.slepian.SlepianPolarCap(L, THETA_MAX)
 
 
 @pytest.fixture(scope="session")
-def slepian_lim_lat_lon() -> SlepianLimitLatLon:
+def slepian_lim_lat_lon() -> sleplet.slepian.SlepianLimitLatLon:
     """Create a Slepian limited latitude longitude class."""
-    return SlepianLimitLatLon(
+    return sleplet.slepian.SlepianLimitLatLon(
         L,
         theta_min=THETA_0,
         theta_max=THETA_1,
@@ -38,51 +38,72 @@ def slepian_lim_lat_lon() -> SlepianLimitLatLon:
 
 
 @pytest.fixture(scope="session")
-def slepian_arbitrary() -> SlepianArbitrary:
+def slepian_arbitrary() -> sleplet.slepian.SlepianArbitrary:
     """Create a Slepian arbitrary class."""
-    return SlepianArbitrary(L, MASK)
+    return sleplet.slepian.SlepianArbitrary(L, MASK)
 
 
 @pytest.fixture(scope="session")
-def earth_polar_cap(slepian_polar_cap) -> Earth:
+def earth_polar_cap(slepian_polar_cap) -> sleplet.functions.Earth:
     """Earth with polar cap region."""
-    return Earth(slepian_polar_cap.L, region=slepian_polar_cap.region)
+    return sleplet.functions.Earth(slepian_polar_cap.L, region=slepian_polar_cap.region)
 
 
 @pytest.fixture(scope="session")
-def earth_lim_lat_lon(slepian_lim_lat_lon) -> Earth:
+def earth_lim_lat_lon(slepian_lim_lat_lon) -> sleplet.functions.Earth:
     """Earth with limited latitude longitude region."""
-    return Earth(slepian_lim_lat_lon.L, region=slepian_lim_lat_lon.region)
+    return sleplet.functions.Earth(
+        slepian_lim_lat_lon.L,
+        region=slepian_lim_lat_lon.region,
+    )
 
 
 @pytest.fixture(scope="session")
-def south_america_arbitrary(slepian_arbitrary) -> SouthAmerica:
+def south_america_arbitrary(slepian_arbitrary) -> sleplet.functions.SouthAmerica:
     """South America already has region."""
-    return SouthAmerica(slepian_arbitrary.L)
+    return sleplet.functions.SouthAmerica(slepian_arbitrary.L)
 
 
 @pytest.fixture(scope="session")
-def slepian_dirac_delta_polar_cap(slepian_polar_cap) -> SlepianDiracDelta:
[…]
