+++ tests/test_wavelets.py
@@ -2,10 +2,7 @@ import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal, assert_equal
 from pys2let import pys2let_j_max
 
-from sleplet.functions import (
-    AxisymmetricWaveletCoefficientsEarth,
-    AxisymmetricWaveletCoefficientsSouthAmerica,
-)
+import sleplet.functions
 from sleplet.slepian_methods import slepian_forward
 from sleplet.wavelet_methods import (
     axisymmetric_wavelet_inverse,
@@ -64,14 +61,22 @@ def test_synthesis_lim_lat_lon(slepian_wavelets_lim_lat_lon, earth_lim_lat_lon) 
 
 def test_axisymmetric_synthesis_earth() -> None:
     """Tests that the axisymmetric wavelet synthesis recoveres the coefficients."""
-    awc = AxisymmetricWaveletCoefficientsEarth(L_SMALL, B=B, j_min=J_MIN)
+    awc = sleplet.functions.AxisymmetricWaveletCoefficientsEarth(
+        L_SMALL,
+        B=B,
+        j_min=J_MIN,
+    )
     flm = axisymmetric_wavelet_inverse(L_SMALL, awc.wavelet_coefficients, awc.wavelets)
     assert_allclose(np.abs(flm - awc.earth.coefficients).mean(), 0, atol=1e-13)
 
 
 def test_axisymmetric_synthesis_south_america() -> None:
     """Tests that the axisymmetric wavelet synthesis recoveres the coefficients."""
-    awc = AxisymmetricWaveletCoefficientsSouthAmerica(L_SMALL, B=B, j_min=J_MIN)
+    awc = sleplet.functions.AxisymmetricWaveletCoefficientsSouthAmerica(
+        L_SMALL,
+        B=B,
+        j_min=J_MIN,
+    )
     flm = axisymmetric_wavelet_inverse(L_SMALL, awc.wavelet_coefficients, awc.wavelets)
     assert_allclose(np.abs(flm - awc.south_america.coefficients).mean(), 0, atol=1e-14)
 
