+++ src/sleplet/functions/slepian_wavelet_coefficients_south_america.py
@@ -1,11 +1,12 @@
 """Contains the `SlepianWaveletCoefficientsSouthAmerica` class."""
+import logging
+
 import numpy as np
 from numpy import typing as npt
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 from pys2let import pys2let_j_max
 
-import sleplet
 import sleplet._string_methods
 import sleplet._validation
 import sleplet.functions.slepian_south_america
@@ -13,6 +14,8 @@ import sleplet.functions.slepian_wavelets
 import sleplet.slepian.region
 import sleplet.wavelet_methods
 from sleplet.functions.fp import Fp
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
@@ -36,9 +39,9 @@ class SlepianWaveletCoefficientsSouthAmerica(Fp):
             raise RuntimeError("Slepian region selected must be 'south_america'")
 
     def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
-        sleplet.logger.info("start computing wavelet coefficients")
+        logger.info("start computing wavelet coefficients")
         self.wavelets, self.wavelet_coefficients = self._create_wavelet_coefficients()
-        sleplet.logger.info("finish computing wavelet coefficients")
+        logger.info("finish computing wavelet coefficients")
         jth = 0 if self.j is None else self.j + 1
         return self.wavelet_coefficients[jth]
 
