+++ src/sleplet/meshes/mesh_basis_functions.py
@@ -1,13 +1,16 @@
 """Contains the `MeshBasisFunctions` class."""
+import logging
+
 import numpy as np
 from numpy import typing as npt
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 
-import sleplet
 import sleplet._validation
 import sleplet.harmonic_methods
 from sleplet.meshes.mesh_harmonic_coefficients import MeshHarmonicCoefficients
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
@@ -24,7 +27,7 @@ class MeshBasisFunctions(MeshHarmonicCoefficients):
 
     def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
         """Compute field on the vertices of the mesh."""
-        sleplet.logger.info(
+        logger.info(
             f"Mesh eigenvalue {self.rank}: "
             f"{self.mesh.mesh_eigenvalues[self.rank]:e}",
         )
