+++ src/sleplet/meshes/mesh_slepian_wavelets.py
@@ -1,15 +1,18 @@
 """Contains the `MeshSlepianWavelets` class."""
+import logging
+
 import numpy as np
 from numpy import typing as npt
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 from pys2let import pys2let_j_max
 
-import sleplet
 import sleplet._string_methods
 import sleplet._validation
 import sleplet.wavelet_methods
 from sleplet.meshes.mesh_slepian_coefficients import MeshSlepianCoefficients
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
@@ -28,9 +31,9 @@ class MeshSlepianWavelets(MeshSlepianCoefficients):
         super().__post_init_post_parse__()
 
     def _create_coefficients(self) -> npt.NDArray[np.complex_ | np.float_]:
-        sleplet.logger.info("start computing wavelets")
+        logger.info("start computing wavelets")
         self.wavelets = self._create_wavelets()
-        sleplet.logger.info("finish computing wavelets")
+        logger.info("finish computing wavelets")
         jth = 0 if self.j is None else self.j + 1
         return self.wavelets[jth]
 
