+++ src/sleplet/__init__.py
@@ -7,9 +7,14 @@ import logging
 
 from ._version import __version__  # noqa: F401
 
-logging.basicConfig(
-    format="[%(asctime)s] [%(levelname)s] --- %(message)s (%(filename)s:%(lineno)s)",
-    datefmt="%Y-%m-%d %H:%M:%S",
-    level=logging.INFO,
-)
 logger = logging.getLogger(__name__)
+
+formatter = logging.Formatter(
+    "[%(asctime)s] [%(levelname)s] --- %(message)s (%(filename)s:%(lineno)s)",
+    datefmt="%Y-%m-%d %H:%M:%S",
+)
+console_handler = logging.StreamHandler()
+console_handler.setFormatter(formatter)
+logger.addHandler(console_handler)
+logger.setLevel("INFO")
+logger.propagate = False
