+++ src/sleplet/meshes/_mesh_slepian_decomposition.py
@@ -1,14 +1,16 @@
+import logging
 from dataclasses import KW_ONLY
 
 import numpy as np
 from numpy import typing as npt
 from pydantic.dataclasses import dataclass
 
-import sleplet
 import sleplet._integration_methods
 import sleplet._validation
 import sleplet.harmonic_methods
 from sleplet.meshes.mesh_slepian import MeshSlepian
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass(config=sleplet._validation.Validation)
@@ -93,13 +95,13 @@ class MeshSlepianDecomposition:
     def _detect_method(self) -> None:
         """Detects what method is used to perform the decomposition."""
         if isinstance(self.u_i, np.ndarray):
-            sleplet.logger.info("harmonic sum method selected")
+            logger.info("harmonic sum method selected")
             self.method = "harmonic_sum"
         elif isinstance(self.u, np.ndarray) and not self.mask:
-            sleplet.logger.info("integrating the whole mesh method selected")
+            logger.info("integrating the whole mesh method selected")
             self.method = "integrate_mesh"
         elif isinstance(self.u, np.ndarray):
-            sleplet.logger.info("integrating a region on the mesh method selected")
+            logger.info("integrating a region on the mesh method selected")
             self.method = "integrate_region"
         else:
             raise RuntimeError(
