+++ src/sleplet/slepian/region.py
@@ -1,4 +1,6 @@
 """Contains the `Region` class."""
+import logging
+
 from pydantic import validator
 from pydantic.dataclasses import dataclass
 
@@ -6,6 +8,8 @@ import sleplet._bool_methods
 import sleplet._string_methods
 import sleplet._validation
 import sleplet._vars
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass(config=sleplet._validation.Validation, kw_only=True)
@@ -38,7 +42,7 @@ class Region:
 
     def _identify_region(self) -> None:
         """Identify region type based on the angle inputs or a mask name."""
-        sleplet.logger.info(
+        logger.info(
             "Slepian region values detected: "
             f"POLAR_GAP={self.gap}, "
             f"THETA_MAX={self.theta_max}, "

[…]
