### 2023-04-03 3f4cbdd
Go back to reading in NCPU when needed
