+++ tests/test_mesh.py
@@ -2,8 +2,8 @@ import numpy as np
 import pytest
 from numpy.testing import assert_allclose
 
+import sleplet
 from sleplet._integration_methods import integrate_whole_mesh
-from sleplet.harmonic_methods import mesh_forward, mesh_inverse
 
 
 def test_inverse_forward_transform_recovery(mesh_field_region) -> None:
@@ -11,8 +11,11 @@ def test_inverse_forward_transform_recovery(mesh_field_region) -> None:
     Tests that a given function is recovered after an
     inverse and forward transform on the mesh.
     """
-    u = mesh_inverse(mesh_field_region.mesh, mesh_field_region.coefficients)
-    kernel_recov = mesh_forward(mesh_field_region.mesh, u)
+    u = sleplet.harmonic_methods.mesh_inverse(
+        mesh_field_region.mesh,
+        mesh_field_region.coefficients,
+    )
+    kernel_recov = sleplet.harmonic_methods.mesh_forward(mesh_field_region.mesh, u)
     assert_allclose(
         np.abs(mesh_field_region.coefficients - kernel_recov).mean(),
         0,
