+++ tests/conftest.py
@@ -3,11 +3,11 @@ import pytest
 from numpy import typing as npt
 from numpy.random import default_rng
 
+import sleplet
 import sleplet.functions
 import sleplet.meshes
 import sleplet.slepian
 from sleplet._vars import RANDOM_SEED
-from sleplet.harmonic_methods import compute_random_signal
 
 ARRAY_DIM = 10
 L = 16
@@ -110,7 +110,7 @@ def slepian_wavelets_lim_lat_lon(
 def random_flm() -> npt.NDArray[np.complex_]:
     """Creates random flm."""
     rng = default_rng(RANDOM_SEED)
-    return compute_random_signal(L, rng, var_signal=1)
+    return sleplet.harmonic_methods.compute_random_signal(L, rng, var_signal=1)
 
 
 @pytest.fixture(scope="session")
@@ -118,7 +118,10 @@ def random_nd_flm() -> npt.NDArray[np.complex_]:
     """Creates multiple random flm."""
     rng = default_rng(RANDOM_SEED)
     return np.array(
-        [compute_random_signal(L, rng, var_signal=1) for _ in range(ARRAY_DIM)],
+        [
+            sleplet.harmonic_methods.compute_random_signal(L, rng, var_signal=1)
+            for _ in range(ARRAY_DIM)
+        ],
     )
 
 
