+++ tests/test_decomposition.py
@@ -2,10 +2,10 @@ import numpy as np
 import pyssht as ssht
 from numpy.testing import assert_allclose, assert_raises
 
+import sleplet
+import sleplet.slepian
 from sleplet._mask_methods import create_mask_region
 from sleplet._vars import SAMPLING_SCHEME
-from sleplet.slepian._slepian_decomposition import SlepianDecomposition
-from sleplet.slepian_methods import slepian_forward, slepian_inverse
 
 
 def test_decompose_all_polar(slepian_polar_cap, earth_polar_cap) -> None:
@@ -15,17 +15,17 @@ def test_decompose_all_polar(slepian_polar_cap, earth_polar_cap) -> None:
         slepian_polar_cap.L,
         Method=SAMPLING_SCHEME,
     )
-    harmonic_sum_p = slepian_forward(
+    harmonic_sum_p = sleplet.slepian_methods.slepian_forward(
         slepian_polar_cap.L,
         slepian_polar_cap,
         flm=earth_polar_cap.coefficients,
     )
-    integrate_sphere_p = slepian_forward(
+    integrate_sphere_p = sleplet.slepian_methods.slepian_forward(
         slepian_polar_cap.L,
         slepian_polar_cap,
         f=field,
     )
-    integrate_region_p = slepian_forward(
+    integrate_region_p = sleplet.slepian_methods.slepian_forward(
         slepian_polar_cap.L,
         slepian_polar_cap,
         f=field,
@@ -53,17 +53,17 @@ def test_decompose_all_lim_lat_lon(slepian_lim_lat_lon, earth_lim_lat_lon) -> No
         slepian_lim_lat_lon.L,
         Method=SAMPLING_SCHEME,
     )
-    harmonic_sum_p = slepian_forward(
+    harmonic_sum_p = sleplet.slepian_methods.slepian_forward(
         slepian_lim_lat_lon.L,
         slepian_lim_lat_lon,
         flm=earth_lim_lat_lon.coefficients,
     )
-    integrate_sphere_p = slepian_forward(
+    integrate_sphere_p = sleplet.slepian_methods.slepian_forward(
         slepian_lim_lat_lon.L,
         slepian_lim_lat_lon,
         f=field,
     )
-    integrate_region_p = slepian_forward(
+    integrate_region_p = sleplet.slepian_methods.slepian_forward(
         slepian_lim_lat_lon.L,
         slepian_lim_lat_lon,
         f=field,
@@ -86,12 +86,16 @@ def test_equality_to_harmonic_transform_polar(
     earth_polar_cap,
 ) -> None:
     """Tests that fp*Sp up to N is roughly equal to flm*Ylm."""
-    f_p = slepian_forward(
+    f_p = sleplet.slepian_methods.slepian_forward(
         slepian_polar_cap.L,
         slepian_polar_cap,
         flm=earth_polar_cap.coefficients,
     )
-    f_slepian = slepian_inverse(f_p, slepian_polar_cap.L, slepian_polar_cap)
+    f_slepian = sleplet.slepian_methods.slepian_inverse(
+        f_p,
+        slepian_polar_cap.L,
+        slepian_polar_cap,
+    )
     f_harmonic = ssht.inverse(
         earth_polar_cap.coefficients,
         slepian_polar_cap.L,
@@ -106,12 +110,16 @@ def test_equality_to_harmonic_transform_lim_lat_lon(
     earth_lim_lat_lon,
 ) -> None:
     """Tests that fp*Sp up to N is roughly equal to flm*Ylm."""
-    f_p = slepian_forward(
+    f_p = sleplet.slepian_methods.slepian_forward(
         slepian_lim_lat_lon.L,
         slepian_lim_lat_lon,
[…]
