+++ tests/test_harmonic.py
@@ -1,8 +1,8 @@
 import pyssht as ssht
 from numpy.testing import assert_equal
 
+import sleplet
 from sleplet._vars import SAMPLING_SCHEME
-from sleplet.harmonic_methods import _boost_coefficient_resolution, invert_flm_boosted
 
 L_LARGE = 128
 L_SMALL = 16
@@ -11,12 +11,15 @@ L_SMALL = 16
 def test_harmonic_coefficients_padded(random_flm) -> None:
     """Tests that harmonic coefficients are zero padded for plotting."""
     boost = L_LARGE**2 - L_SMALL**2
-    flm_boosted = _boost_coefficient_resolution(random_flm, boost)
+    flm_boosted = sleplet.harmonic_methods._boost_coefficient_resolution(
+        random_flm,
+        boost,
+    )
     assert_equal(len(flm_boosted), L_LARGE**2)
 
 
 def test_invert_flm_and_boost(random_flm) -> None:
     """Tests that the flm has been boosted and has right shape."""
     n_theta, n_phi = ssht.sample_shape(L_LARGE, Method=SAMPLING_SCHEME)
-    f = invert_flm_boosted(random_flm, L_SMALL, L_LARGE)
+    f = sleplet.harmonic_methods.invert_flm_boosted(random_flm, L_SMALL, L_LARGE)
     assert_equal(f.shape, (n_theta, n_phi))

[…]
