+++ src/sleplet/slepian/slepian_functions.py
@@ -9,7 +9,7 @@ from pydantic.dataclasses import dataclass
 import sleplet._validation
 import sleplet.slepian.region
 
-logger = logging.getLogger(__name__)
+_logger = logging.getLogger(__name__)
 
 
 @dataclass(config=sleplet._validation.Validation)
@@ -25,11 +25,11 @@ class SlepianFunctions:
         self.name = self._create_fn_name()
         area = self._calculate_area()
         self.N = round(area * self.L**2 / (4 * np.pi))
-        logger.info(f"Shannon number N={self.N}")
+        _logger.info(f"Shannon number N={self.N}")
         self.matrix_location = self._create_matrix_location()
-        logger.info("start solving eigenproblem")
+        _logger.info("start solving eigenproblem")
         self.eigenvalues, self.eigenvectors = self._solve_eigenproblem()
-        logger.info("finished solving eigenproblem")
+        _logger.info("finished solving eigenproblem")
 
     @abstractmethod
     def _create_fn_name(self) -> str:
