+++ tests/test_harmonic.py
@@ -2,7 +2,6 @@ import pyssht as ssht
 from numpy.testing import assert_equal
 
 import sleplet
-from sleplet._vars import SAMPLING_SCHEME
 
 L_LARGE = 128
 L_SMALL = 16
@@ -20,6 +19,6 @@ def test_harmonic_coefficients_padded(random_flm) -> None:
 
 def test_invert_flm_and_boost(random_flm) -> None:
     """Tests that the flm has been boosted and has right shape."""
-    n_theta, n_phi = ssht.sample_shape(L_LARGE, Method=SAMPLING_SCHEME)
+    n_theta, n_phi = ssht.sample_shape(L_LARGE, Method=sleplet._vars.SAMPLING_SCHEME)
     f = sleplet.harmonic_methods.invert_flm_boosted(random_flm, L_SMALL, L_LARGE)
     assert_equal(f.shape, (n_theta, n_phi))
