+++ tests/test_wavelets.py
@@ -3,7 +3,6 @@ from numpy.testing import assert_allclose, assert_array_equal, assert_equal
 from pys2let import pys2let_j_max
 
 import sleplet
-import sleplet.functions
 
 B = 2
 J_MIN = 0
