+++ tests/test_convolution.py
@@ -2,7 +2,6 @@ import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal
 
 import sleplet
-import sleplet.functions
 
 L = 128
 
