### 2023-04-03 fdd5755
Change logger and make examples print
