+++ examples/polar_cap/simons_5_1.py
@@ -3,7 +3,6 @@ import seaborn as sns
 from matplotlib import pyplot as plt
 from numpy import typing as npt
 
-from sleplet import logger
 from sleplet.harmonic_methods import invert_flm_boosted
 from sleplet.plot_methods import calc_plot_resolution, save_plot
 from sleplet.slepian import SlepianPolarCap
@@ -55,7 +54,7 @@ def _helper(  # noqa: PLR0913
     rank: int,
 ) -> None:
     """Helper which plots the required order and specified ranks."""
-    logger.info(f"plotting order={order}, rank={rank}")
+    print(f"plotting order={order}, rank={rank}")
     axs = ax[order, rank]
     flm = slepian.eigenvectors[rank] * SIGNS[order][rank]
     lam = slepian.eigenvalues[rank]
