+++ README.md
@@ -93,15 +93,18 @@ done
 ```python
 import numpy as np
 import pyssht as ssht
-
-from sleplet.functions import HarmonicGaussian
-from sleplet.plotting import PlotSphere
+import sleplet
 
 for ell in range(2, 0, -1):
-    f = HarmonicGaussian(L=128, l_sigma=10**ell, m_sigma=10)
+    f = sleplet.functions.HarmonicGaussian(L=128, l_sigma=10**ell, m_sigma=10)
     flm = f.translate(alpha=0.75 * np.pi, beta=0.125 * np.pi)
     f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
-    PlotSphere(f_sphere, f.L, f"fig_1_ell_{ell}", annotations=[]).execute()
+    sleplet.plotting.PlotSphere(
+        f_sphere,
+        f.L,
+        f"fig_1_ell_{ell}",
+        annotations=[],
+    ).execute()
 ```
 
 #### Sifting Convolution on the Sphere: Fig. 2
@@ -112,15 +115,12 @@ sphere earth -L 128
 
 ```python
 import pyssht as ssht
+import sleplet
 
-from sleplet.functions import Earth
-from sleplet.harmonic_methods import rotate_earth_to_south_america
-from sleplet.plotting import PlotSphere
-
-f = Earth(L=128)
-flm = rotate_earth_to_south_america(f.coefficients, f.L)
+f = sleplet.functions.Earth(L=128)
+flm = sleplet.harmonic_methods.rotate_earth_to_south_america(f.coefficients, f.L)
 f_sphere = ssht.inverse(flm, f.L, Method="MWSS")
-PlotSphere(f_sphere, f.L, "fig_2").execute()
+sleplet.plotting.PlotSphere(f_sphere, f.L, "fig_2").execute()
 ```
 
 #### Sifting Convolution on the Sphere: Fig. 3
@@ -133,18 +133,15 @@ done
 
 ```python
 import pyssht as ssht
-
-from sleplet.functions import Earth, HarmonicGaussian
-from sleplet.harmonic_methods import rotate_earth_to_south_america
-from sleplet.plotting import PlotSphere
+import sleplet
 
 for ell in range(2, 0, -1):
-    f = HarmonicGaussian(L=128, l_sigma=10**ell, m_sigma=10)
-    g = Earth(L=128)
+    f = sleplet.functions.HarmonicGaussian(L=128, l_sigma=10**ell, m_sigma=10)
+    g = sleplet.functions.Earth(L=128)
     flm = f.convolve(f.coefficients, g.coefficients.conj())
-    flm_rot = rotate_earth_to_south_america(flm, f.L)
+    flm_rot = sleplet.harmonic_methods.rotate_earth_to_south_america(flm, f.L)
     f_sphere = ssht.inverse(flm_rot, f.L, Method="MWSS")
-    PlotSphere(f_sphere, f.L, f"fig_3_ell_{ell}").execute()
+    sleplet.plotting.PlotSphere(f_sphere, f.L, f"fig_3_ell_{ell}").execute()
 ```
 
 ### Slepian Scale-Discretised Wavelets on the Sphere
@@ -169,23 +166,24 @@ sphere slepian_south_america -L 128 -s 2 -u
 
 ```python
 import pyssht as ssht
-
-from sleplet.functions import Earth, SlepianSouthAmerica
-from sleplet.harmonic_methods import rotate_earth_to_south_america
-from sleplet.plotting import PlotSphere
-from sleplet.slepian import Region
-from sleplet.slepian_methods import slepian_inverse
+import sleplet
 
 # a
-f = Earth(L=128, smoothing=2)
-flm = rotate_earth_to_south_america(f.coefficients, f.L)
[…]
