### 2023-04-03 e7ac0bf
CLI verison option/change env var/logging/fix submodules (#124)
A bit more involved than expected.
- CLI can now get a version with `-v` (and changed the previous `--view`
to `--perspective`
- Environment variables now only loaded when needed
- Logging is done per file rather than import `sleplet`
- Can now properly load sub-modules (and used in the tests)
- Fix import package directly
