+++ src/sleplet/_scripts/plotting_on_sphere.py
@@ -1,10 +1,10 @@
+import logging
 from argparse import ArgumentParser, Namespace
 
 import numpy as np
 import pyssht as ssht
 from numpy import typing as npt
 
-import sleplet
 import sleplet._class_lists
 import sleplet._mask_methods
 import sleplet._string_methods
@@ -14,9 +14,12 @@ import sleplet.plot_methods
 import sleplet.plotting._create_plot_sphere
 import sleplet.slepian_methods
 
-ALPHA_DEFAULT = 0.75
-ANNOTATION_COLOUR = "gold"
-ARROW_STYLE = {
+_logger = logging.getLogger(__name__)
+
+
+_ALPHA_DEFAULT = 0.75
+_ANNOTATION_COLOUR = "gold"
+_ARROW_STYLE = {
     "arrowhead": 0,
     "arrowside": "start",
     "ax": 4,
@@ -24,7 +27,7 @@ ARROW_STYLE = {
     "startarrowsize": 0.5,
     "startarrowhead": 6,
 }
-BETA_DEFAULT = 0.125
+_BETA_DEFAULT = 0.125
 
 
 def valid_maps(map_name: str) -> str:
@@ -60,16 +63,16 @@ def read_args() -> Namespace:
         "--alpha",
         "-a",
         type=float,
-        default=ALPHA_DEFAULT,
-        help=f"alpha/phi pi fraction - defaults to {ALPHA_DEFAULT}",
+        default=_ALPHA_DEFAULT,
+        help=f"alpha/phi pi fraction - defaults to {_ALPHA_DEFAULT}",
     )
     parser.add_argument("--bandlimit", "-L", type=int, default=16, help="bandlimit")
     parser.add_argument(
         "--beta",
         "-b",
         type=float,
-        default=BETA_DEFAULT,
-        help=f"beta/theta pi fraction - defaults to {BETA_DEFAULT}",
+        default=_BETA_DEFAULT,
+        help=f"beta/theta pi fraction - defaults to {_BETA_DEFAULT}",
     )
     parser.add_argument(
         "--convolve",
@@ -113,6 +116,16 @@ def read_args() -> Namespace:
         help="flag which removes any annotation",
     )
     parser.add_argument(
+        "--perspective",
+        "-p",
+        type=str,
+        nargs="?",
+        default="south_america",
+        const="south_america",
+        choices=["africa", "south_america"],
+        help="view of Earth: defaults to 'south_america'",
+    )
+    parser.add_argument(
         "--region",
         "-r",
         action="store_true",
@@ -147,14 +160,10 @@ def read_args() -> Namespace:
         help="flag turns off upsampling for plot",
     )
     parser.add_argument(
-        "--view",
+        "--version",
         "-v",
-        type=str,
-        nargs="?",
-        default="south_america",
-        const="south_america",
-        choices=["africa", "south_america"],
-        help="view of Earth: defaults to 'south_america'",
+        action="version",
+        version=sleplet.__version__,
     )
     return parser.parse_args()
 
@@ -179,13 +188,13 @@ def plot(
     coefficients = f.coefficients
 
     # turn off annotation if needed
-    sleplet.logger.info(f"annotations on: {annotations}")
+    _logger.info(f"annotations on: {annotations}")
     annotation = []
 
     # Shannon number for Slepian coefficients
     shannon = f.slepian.N if hasattr(f, "slepian") else None
 
-    sleplet.logger.info(f"plotting method: '{method}'")
+    _logger.info(f"plotting method: '{method}'")
     match method:
         case "rotate":
             coefficients, filename = _rotation_helper(
[…]
