+++ tests/test_parallel.py
@@ -1,6 +1,6 @@
 from numpy.testing import assert_allclose, assert_equal
 
-from sleplet._parallel_methods import split_arr_into_chunks
+import sleplet
 
 L_LARGE = 128
 L_SMALL = 16
@@ -9,7 +9,7 @@ NCPU = 4
 
 def test_split_L_into_chunks() -> None:  # noqa: N802
     """Ensure vector L split into appropriate number of chunks."""
-    chunks = split_arr_into_chunks(L_SMALL, NCPU)
+    chunks = sleplet._parallel_methods.split_arr_into_chunks(L_SMALL, NCPU)
     assert_equal(len(chunks), NCPU)
     chunk_length = L_SMALL // NCPU
     for chunk in chunks:

[…]
