+++ pyproject.toml
@@ -22,21 +22,21 @@ classifiers = [
     "Typing :: Typed",
 ]
 dependencies = [
-    "cmocean>=2.0.0",
+    "cmocean>=3.0.3",
     "gmpy2>=2.1.5",
-    "hypothesis>=6.67.1",
+    "hypothesis>=6.70.2",
     "kaleido>=0.2.1",
     "libigl>=2.4.1",
-    "matplotlib>=3.6.3",
+    "matplotlib>=3.7.1",
     "multiprocess>=0.70.14",
     "numba>=0.56.4",
     "numpy>=1.23.5",
-    "plotly>=5.13.0",
+    "plotly>=5.14.0",
     "pooch>=1.7.0",
-    "pydantic>=1.10.4",
-    "pys2let>=2.2.4",
-    "pyssht>=1.5.1",
-    "scipy>=1.10.0",
+    "pydantic>=1.10.7",
+    "pys2let>=2.2.5",
+    "pyssht>=1.5.2",
+    "scipy>=1.10.1",
     "seaborn>=0.12.2",
     "tomli>=2.0.1",
     "tqdm>=4.65.0",
