+++ temp.py
@@ -1,0 +1,3 @@
+import pooch
+
+print(pooch.os_cache("sleplet"))
