+++ .github/workflows/test.yml
@@ -31,6 +31,9 @@ jobs:
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
 
+      - name: Find pooch cache
+        run: python temp.py
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
