+++ .github/workflows/test.yml
@@ -31,13 +31,10 @@ jobs:
           cache: "pip"
           cache-dependency-path: "pyproject.toml"
 
-      - name: Find pooch cache
-        run: python temp.py
-
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install tox tox-gh-actions
+          python -m pip install pooch
 
-      - name: Test with tox
-        run: tox
+      - name: Find pooch cache
+        run: python temp.py
