### 2023-04-03 c319e2b
Try pwd
