+++ temp.py
@@ -1,3 +1,0 @@
-import pooch
-
-print(pooch.os_cache("sleplet"))
