+++ .github/workflows/test.yml
@@ -35,3 +35,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install tox tox-gh-actions
+
+      - name: Test with tox
+        run: tox
