+++ pyproject.toml
@@ -12,6 +12,8 @@ authors = [
 classifiers = [
     "License :: OSI Approved :: BSD License",
     "Operating System :: POSIX",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3 :: Only",
     "Programming Language :: Python :: 3.10",
     "Topic :: Scientific/Engineering :: Image Processing",
     "Topic :: Scientific/Engineering :: Mathematics",
@@ -20,21 +22,21 @@ classifiers = [
     "Typing :: Typed",
 ]
 dependencies = [
-    "cmocean>=2.0.0",
+    "cmocean>=3.0.3",
     "gmpy2>=2.1.5",
-    "hypothesis>=6.67.1",
+    "hypothesis>=6.70.2",
     "kaleido>=0.2.1",
     "libigl>=2.4.1",
-    "matplotlib>=3.6.3",
+    "matplotlib>=3.7.1",
     "multiprocess>=0.70.14",
     "numba>=0.56.4",
     "numpy>=1.23.5",
-    "plotly>=5.13.0",
+    "plotly>=5.14.0",
     "pooch>=1.7.0",
-    "pydantic>=1.10.4",
-    "pys2let>=2.2.4",
-    "pyssht>=1.5.1",
-    "scipy>=1.10.0",
+    "pydantic>=1.10.7",
+    "pys2let>=2.2.5",
+    "pyssht>=1.5.2",
+    "scipy>=1.10.1",
     "seaborn>=0.12.2",
     "tomli>=2.0.1",
     "tqdm>=4.65.0",
@@ -178,6 +180,7 @@ write_to = "src/sleplet/_version.py"
 all = true
 spaces_indent_inline_array = 4
 trailing_comma_inline_array = true
+overrides."project.classifiers".inline_arrays = false
 overrides."tool.coverage.paths.source".inline_arrays = false
 
 [tool.tox]
