+++ paper/paper.bib
@@ -88,9 +88,10 @@
   publisher = {arXiv},
   url       = {https://arxiv.org/abs/2302.06006}
 }
-@software{RoddySLEPLET,
+@software{Roddy2023a,
   title  = {{SLEPLET}},
   author = {Roddy, Patrick J.},
+  year   = {2023},
   doi    = {10.5281/zenodo.7268074}
 }
 @article{Simons2006,
@@ -104,8 +105,9 @@
   doi     = {10.1137/S0036144504445765}
 }
 @software{Simons2020,
-  author = {Frederik J Simons and Chris Harig and Alain Plattner and Max von Hippel and Albert},
   title  = {{slepian\_alpha}},
+  author = {Frederik J Simons and Chris Harig and Alain Plattner and Max von Hippel and Albert},
+  year   = 2020,
   doi    = {10.5281/zenodo.4085210}
 }
 @article{Slepian1961,
