### 2023-04-04 astro-informatics/sleplet#126 @Saran-nns
[REVIEW]: SLEPLET: Slepian scale-discretised wavelets in Python
Hi, I have few suggestions / requests. Let me know if you disagree with any of these:

#### Installation Instructions
- [x] On Windows, pip install sleplet fails https://github.com/openjournals/joss-reviews/issues/5221#issuecomment-1466668621. Additionally, it appears that sleplet is only compatible with python >=3.10 (i see badge). The installation instructions should be updated to reflect the supported operating systems and compatible python versions. Did you tried  previous stable versions of `pys2let>=2.2.5, pyssht>=1.5.2` ?

#### Usage/Examples
- [x] When using `sleplet`, generated images and PDFs are saved in the package root directory in conda. Is it possible for users to define or change the save path? _Imagine, only my admin has the right to access conda directory._
 - [x] `sleplet` saves generated figures in both PDF and PNG formats, which may be redundant. Can users choose the file type to be saved?
 - [x] `sleplet` does not report or plot the generated figures. When testing the examples in the notebook, it was somewhat inconvenient to search for the generated images.
 - [x] Also, it download the cache/save sample analysis data in conda root directory. I recommend it to be at `pwd`.

#### Documentation

- [x] The README file seems to include almost all example codes, which should be moved to the documentation. I recommend keeping only one or two sample codes with generated images in the `README` file and directing users to the documentation for further information, including the `Bandlimit` and `Environment Variables`. It would be best to host those examples at `https://astro-informatics.github.io/sleplet/index.html`. Otherwise, a separate `.md` file for documentation in the package itself could be created, and the path could be hyperlinked in the `README` file.

- [x] Here's what I suggest adding to the `README` file:
About sleplet
Installation instructions
Usage with a few example code runs and images/results with explanations instead of links to paper figures.
Documentation, hyperlink (to doc) for further information
Community guidelines

Thanks
