+++ .github/workflows/test.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   test:
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
     steps:
       - name: Checkout source
         uses: actions/checkout@v3
@@ -34,7 +34,4 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install tox tox-gh-actions
-
-      - name: Test with tox
-        run: tox
+          python -m pip install -e .
