+++ README.md
@@ -31,870 +31,4 @@ pip install -e .
 ```
 
 This will install two scripts `sphere` and `mesh` which can be used to generate
-the figures in [the figure section](#paper-figures).
-
-## Bandlimit
-
-The bandlimit is set as `L` throughout the code and the CLIs. The default value
-is set to `L=16` and the figures created in [the figure section](#paper-figures)
-all use `L=128`. The pre-computed data exists on
-[Zenodo](https://doi.org/10.5281/zenodo.7767698) for powers of two up to `L=128`.
-Other values will be computed when running the appropriate code (and saved for
-future use). Note that beyond `L=32` the code can be slow due to the
-difficulties of computing the Slepian matrix prior to the eigendecomposition, as
-such it is recommended to stick to the powers of two up to `L=128`.
-
-## Environment Variables
-
-- `NCPU`: sets the number of cores to use
-
-When it comes to selecting a Slepian region the order precedence is
-[polar cap region](https://doi.org/10.1111/j.1365-246X.2006.03065.x) >
-[limited latitude longitude region](https://doi.org/10.1109/TSP.2016.2646668) >
-arbitrary region,
-[as seen in the code](https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/utils/region.py).
-The default region is the `south_america` arbitrary region.
-
-- `POLAR_GAP`
-  - for a Slepian `polar cap region`, when set in conjunction with `THETA_MAX`
-    but without the other `PHI`/`THETA` variables
-- `THETA_MAX`
-  - for a Slepian `polar cap region`, when set without the other `PHI`/`THETA`
-    variables
-  - for a Slepian `limited latitude longitude region`
-- `THETA_MIN`
-  - for a Slepian `limited latitude longitude region`
-- `PHI_MAX`
-  - for a Slepian `limited latitude longitude region`
-- `PHI_MIN`
-  - for a Slepian `limited latitude longitude region`
-- `SLEPIAN_MASK`
-  - for an arbitrary Slepian region, currently `africa`/`south_america` supported
-
-## Paper Figures
-
-To recreate the figures from the below papers, one may use the CLI or the API.
-For those which don't use the `mesh` or `sphere` CLIs, the relevant API code
-isn't provided as it is contained within the
-[examples folder](https://github.com/astro-informatics/sleplet/tree/main/examples).
-
-### Sifting Convolution on the Sphere
-
-[![Sifting Convolution on the Sphere](https://img.shields.io/badge/DOI-10.1109/LSP.2021.3050961-pink.svg)](https://dx.doi.org/10.1109/LSP.2021.3050961)
-
-#### Sifting Convolution on the Sphere: Fig. 1
-
-```sh
-for ell in $(seq 2 -1 1); do
-    sphere harmonic_gaussian -a 0.75 -b 0.125 -e ${ell} 1 -L 128 -m translate -o
-done
-```
-
-```python
-import numpy as np
-import pyssht as ssht
-import sleplet
-
-for ell in range(2, 0, -1):
-    f = sleplet.functions.HarmonicGaussian(128, l_sigma=10**ell, m_sigma=10)
[…]
