+++ .github/workflows/examples.yml
@@ -4,7 +4,7 @@ on:
   workflow_dispatch:
 
 jobs:
-  readme:
+  codeblocks:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout source
