### 2023-04-04 59ddcb0
Remove second set of bullet points
