### 2023-04-04 06feb78
Add a note on supported platforms
