### 2023-04-04 5b37361
Add a note on supported platforms & test on mac (#129)
