+++ src/sleplet/_mask_methods.py
@@ -1,5 +1,6 @@
 import logging
 import os
+from pathlib import Path
 
 import numpy as np
 import pyssht as ssht
@@ -196,5 +197,5 @@ def create_mask(L: int, mask_name: str) -> npt.NDArray[np.float_]:
         mask = _create_south_america_mask(L, earth_flm)
     else:
         raise ValueError(f"Mask name {mask_name} not recognised")
-    np.save(sleplet._vars.DATA_PATH / f"slepian_masks_{mask_name}", mask)
+    np.save(Path.cwd() / f"slepian_masks_{mask_name}", mask)
     return mask
