+++ src/sleplet/_vars.py
@@ -2,7 +2,6 @@ from pathlib import Path
 
 import numpy as np
 
-DATA_PATH = Path(__file__).resolve().parent / "_data"
 FIG_PATH = Path(__file__).resolve().parent / "_figures"
 PHI_0 = np.pi
 PHI_MAX_DEFAULT = 2 * np.pi
