### 2023-04-04 astro-informatics/sleplet#126 @paddyroddy
Data location saving fixed in #130
