+++ src/sleplet/plotting/_create_plot_mesh.py
@@ -1,5 +1,6 @@
 import logging
 from dataclasses import KW_ONLY
+from pathlib import Path
 
 import cmocean
 import numpy as np
@@ -14,7 +15,6 @@ import sleplet._mask_methods
 import sleplet._mesh_methods
 import sleplet._plotly_methods
 import sleplet._validation
-import sleplet._vars
 import sleplet.meshes.mesh
 import sleplet.plot_methods
 
@@ -95,13 +95,13 @@ class PlotMesh:
 
         fig = Figure(data=data, layout=layout)
 
-        html_filename = str(sleplet._vars.FIG_PATH / "html" / f"{self.filename}.html")
+        html_filename = str(Path.cwd() / f"{self.filename}.html")
 
         py.plot(fig, filename=html_filename)
 
         for file_type in {"png", "pdf"}:
             filename = str(
-                sleplet._vars.FIG_PATH / file_type / f"{self.filename}.{file_type}",
+                Path.cwd() / f"{self.filename}.{file_type}",
             )
             _logger.info(f"saving {filename}")
             fig.write_image(filename, engine="kaleido")
