### 2023-04-04 2c25aa3
Save figures in documents directory
