+++ src/sleplet/plot_methods.py
@@ -1,12 +1,12 @@
 """Methods to help in creating plots."""
 import logging
-from pathlib import Path
 
 import numpy as np
 import pyssht as ssht
 from matplotlib import colors
 from matplotlib import pyplot as plt
 from numpy import typing as npt
+from platformdirs import user_documents_path
 
 import sleplet._mask_methods
 import sleplet._vars
@@ -85,7 +85,7 @@ def save_plot(name: str) -> None:
     """
     plt.tight_layout()
     for file_type in {"png", "pdf"}:
-        filename = Path.cwd() / f"{name}.{file_type}"
+        filename = user_documents_path() / f"{name}.{file_type}"
         _logger.info(f"saving {filename}")
         plt.savefig(filename, bbox_inches="tight")
     plt.show(block=False)
