+++ pyproject.toml
@@ -31,6 +31,7 @@ dependencies = [
     "multiprocess>=0.70.14",
     "numba>=0.56.4",
     "numpy>=1.23.5",
+    "platformdirs>=3.2.0",
     "plotly>=5.14.0",
     "pooch>=1.7.0",
     "pydantic>=1.10.7",
