### 2023-04-04 astro-informatics/sleplet#126 @paddyroddy
In #131, for now, I'm saving all figures to the user's documents directory
