+++ examples/misc/translation_normalisation.py
@@ -4,7 +4,6 @@ import seaborn as sns
 from matplotlib import pyplot as plt
 
 from sleplet.functions import HarmonicGaussian
-from sleplet.plot_methods import save_plot
 
 sns.set(context="paper")
 
@@ -27,7 +26,11 @@ def compute_translation_normalisation_theta() -> None:
     plt.ylabel(r"${\Vert f_{\ell m} Y_{\ell m}(\omega')\Vert}_{2}$")
     plt.xlim(0, 180)
     plt.ylim(0, np.ceil(norm.max()))
-    save_plot(f"harmonic_gaussian_translation_normalisation_L{L}")
+    # f"harmonic_gaussian_translation_normalisation_L{L}"
+    plt.tight_layout()
+    plt.show(block=False)
+    plt.pause(3)
+    plt.close()
 
 
 if __name__ == "__main__":
