+++ examples/arbitrary/south_america/tiling_south_america.py
@@ -3,7 +3,6 @@ import seaborn as sns
 from matplotlib import pyplot as plt
 from scipy.interpolate import pchip
 
-from sleplet.plot_methods import save_plot
 from sleplet.slepian import SlepianArbitrary
 from sleplet.wavelet_methods import create_kappas
 
@@ -41,7 +40,11 @@ def main() -> None:
     plt.xticks(ticks, ticks)
     plt.xlabel(r"$p$")
     plt.legend()
-    save_plot(f"south_america_slepian_tiling_L{L}")
+    # f"south_america_slepian_tiling_L{L}"
+    plt.tight_layout()
+    plt.show(block=False)
+    plt.pause(3)
+    plt.close()
 
 
 if __name__ == "__main__":
