+++ examples/polar_cap/eigenvalues.py
@@ -31,7 +31,7 @@ def main() -> None:
     # "polar_cap_eigenvalues"
     plt.tight_layout()
     plt.show(block=False)
-    plt.pause(3)
+    plt.pause(10)
     plt.close()
 
 
