+++ examples/misc/wavelet_transform.py
@@ -22,7 +22,7 @@ def main() -> None:
     # "ricker_wavelets"
     plt.tight_layout()
     plt.show(block=False)
-    plt.pause(3)
+    plt.pause(10)
     plt.close()
 
 
