+++ examples/arbitrary/africa/tiling_africa.py
@@ -43,7 +43,7 @@ def main() -> None:
     # f"africa_slepian_tiling_L{L}"
     plt.tight_layout()
     plt.show(block=False)
-    plt.pause(3)
+    plt.pause(10)
     plt.close()
 
 
