+++ examples/polar_cap/simons_5_1.py
@@ -44,7 +44,7 @@ def main() -> None:
     # "slepian_colatitude"
     plt.tight_layout()
     plt.show(block=False)
-    plt.pause(3)
+    plt.pause(10)
     plt.close()
 
 
