### 2023-04-05 dd5615a
Use default renderer, fig.show()
