### 2023-04-05 20e464b
Log filename rather than saving it
