+++ examples/misc/translation_normalisation.py
@@ -26,7 +26,7 @@ def compute_translation_normalisation_theta() -> None:
     plt.ylabel(r"${\Vert f_{\ell m} Y_{\ell m}(\omega')\Vert}_{2}$")
     plt.xlim(0, 180)
     plt.ylim(0, np.ceil(norm.max()))
-    # f"harmonic_gaussian_translation_normalisation_L{L}"
+    print(f"Saving: harmonic_gaussian_translation_normalisation_L{L}")
     plt.tight_layout()
     plt.show(block=False)
     plt.pause(10)
