### 2023-04-05 292fc2d
FIx logging
