### 2023-04-05 5bf9fb1
Fix missing logging
