### 2023-04-05 a77e463
Fix plotly visualisation in notebooks & no longer save figures (#133)
Hadn't thought to check notebooks. Think it also makes sense for users
to save their own figures.
