+++ src/sleplet/plotting/_create_plot_sphere.py
@@ -3,10 +3,8 @@ from dataclasses import KW_ONLY, field
 
 import cmocean
 import numpy as np
-import plotly.offline as py
 import pyssht as ssht
 from numpy import typing as npt
-from platformdirs import user_documents_path
 from plotly.graph_objs import Figure, Surface
 from plotly.graph_objs.surface import Lighting
 from pydantic.dataclasses import dataclass
@@ -115,18 +113,9 @@ class PlotSphere:
             annotations=self.annotations,
         )
 
+        _logger.info(f"Saving: {self.filename}")
         fig = Figure(data=data, layout=layout)
-
-        html_filename = str(user_documents_path() / f"{self.filename}.html")
-
-        py.plot(fig, filename=html_filename)
-
-        for file_type in {"png", "pdf"}:
-            filename = str(
-                user_documents_path() / f"{self.filename}.{file_type}",
-            )
-            _logger.info(f"saving {filename}")
-            fig.write_image(filename, engine="kaleido")
+        fig.show()
 
     @staticmethod
     def _setup_plot(
