+++ src/sleplet/plotting/_create_plot_mesh.py
@@ -3,10 +3,8 @@ from dataclasses import KW_ONLY
 
 import cmocean
 import numpy as np
-import plotly.offline as py
 from matplotlib.colors import LinearSegmentedColormap
 from numpy import typing as npt
-from platformdirs import user_documents_path
 from plotly.graph_objs import Figure, Mesh3d
 from plotly.graph_objs.mesh3d import Lighting
 from pydantic.dataclasses import dataclass
@@ -93,18 +91,9 @@ class PlotMesh:
 
         layout = sleplet._plotly_methods.create_layout(self.mesh.camera_view)
 
+        _logger.info(f"Saving: {self.filename}")
         fig = Figure(data=data, layout=layout)
-
-        html_filename = str(user_documents_path() / f"{self.filename}.html")
-
-        py.plot(fig, filename=html_filename)
-
-        for file_type in {"png", "pdf"}:
-            filename = str(
-                user_documents_path() / f"{self.filename}.{file_type}",
-            )
-            _logger.info(f"saving {filename}")
-            fig.write_image(filename, engine="kaleido")
+        fig.show()
 
     def _prepare_field(
         self,
