+++ examples/arbitrary/africa/eigenvalues_africa.py
@@ -2,7 +2,6 @@ import numpy as np
 import seaborn as sns
 from matplotlib import pyplot as plt
 
-from sleplet.plot_methods import save_plot
 from sleplet.slepian import SlepianArbitrary
 
 sns.set(context="paper")
@@ -28,7 +27,11 @@ def main() -> None:
     plt.xticks(ticks, ticks)
     plt.xlabel(r"$p$")
     plt.ylabel(r"$\mu$")
-    save_plot(f"africa_eigenvalues_L{L}")
+    print(f"Saving: africa_eigenvalues_L{L}")
+    plt.tight_layout()
+    plt.show(block=False)
+    plt.pause(10)
+    plt.close()
 
 
 if __name__ == "__main__":
