### 2023-04-05 astro-informatics/sleplet#126 @paddyroddy
In #133, I'm now not saving figures at all as it probably makes sense for the user to do that themselves.

I also hadn't thought to check notebooks (as I don't use them). I'm now using https://plotly.com/python/renderers/#displaying-figures-using-the-renderers-framework which defaults to the appropriate framework. This is browser normally and works in notebooks.
