### 2023-04-05 f760d6c
Restore supported platforms
