+++ README.md
@@ -33,3 +33,15 @@ pip install -e .
 This will install two scripts `sphere` and `mesh` which can be used to generate
 the figures in
 [the associated papers](https://astro-informatics.github.io/sleplet/#paper-figures).
+
+### Supported Platforms
+
+`SLEPLET` has been tested with
+[![Python](https://img.shields.io/pypi/pyversions/sleplet)](https://www.python.org).
+Windows is not currently supported as `SLEPLET` relies on
+[pyssht](https://pypi.org/project/pyssht) and
+[pys2let](https://pypi.org/project/pys2let) which do not work on Windows.
+These can hopefully be replaced with
+[s2fft](https://github.com/astro-informatics/s2fft/) and
+[s2wav](https://github.com/astro-informatics/s2wav) in the future when they
+are available on [PyPI](https://pypi.org/).
