+++ examples/arbitrary/africa/tiling_africa.py
@@ -40,7 +40,7 @@ def main() -> None:
     plt.xticks(ticks, ticks)
     plt.xlabel(r"$p$")
     plt.legend()
-    print(f"Saving: africa_slepian_tiling_L{L}")
+    print(f"Opening: africa_slepian_tiling_L{L}")
     plt.tight_layout()
     plt.show(block=False)
     plt.pause(10)
