+++ examples/arbitrary/eigenvalues_combined.py
@@ -40,7 +40,7 @@ def main() -> None:
     plt.xlabel(r"$p$")
     plt.ylabel(r"$\mu$")
     plt.legend(loc=3)
-    print(f"Saving: combined_eigenvalues_L{L}")
+    print(f"Opening: combined_eigenvalues_L{L}")
     plt.tight_layout()
     plt.show(block=False)
     plt.pause(10)
