+++ README.md
@@ -2,7 +2,7 @@
 
 [![PyPI](https://badge.fury.io/py/sleplet.svg)](https://pypi.org/project/sleplet)
 [![Python](https://img.shields.io/pypi/pyversions/sleplet)](https://www.python.org)
-[![Documentation](https://img.shields.io/badge/Documentation-API-blueviolet.svg)](https://astro-informatics.github.io/sleplet)
+[![Documentation](https://img.shields.io/badge/Documentation-SLEPLET-blueviolet.svg)](https://astro-informatics.github.io/sleplet)
 [![Zenodo](https://zenodo.org/badge/DOI/10.5281/zenodo.7268074.svg)](https://doi.org/10.5281/zenodo.7268074)
 [![Test](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml/badge.svg)](https://github.com/astro-informatics/sleplet/actions/workflows/test.yml)
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
@@ -31,7 +31,8 @@ pip install -e .
 ```
 
 This will install two scripts `sphere` and `mesh` which can be used to generate
-the figures in [the figure section](#paper-figures).
+the figures in
+[the associated papers](https://astro-informatics.github.io/sleplet/#paper-figures).
 
 ### Supported Platforms
 
@@ -45,868 +46,51 @@ These can hopefully be replaced with
 [s2wav](https://github.com/astro-informatics/s2wav) in the future when they
 are available on [PyPI](https://pypi.org/).
 
-## Bandlimit
+## Example Usage
 
-The bandlimit is set as `L` throughout the code and the CLIs. The default value
-is set to `L=16` and the figures created in [the figure section](#paper-figures)
-all use `L=128`. The pre-computed data exists on
-[Zenodo](https://doi.org/10.5281/zenodo.7767698) for powers of two up to `L=128`.
-Other values will be computed when running the appropriate code (and saved for
-future use). Note that beyond `L=32` the code can be slow due to the
-difficulties of computing the Slepian matrix prior to the eigendecomposition, as
-such it is recommended to stick to the powers of two up to `L=128`.
+`SLEPLET` may be interacted with via the API or the CLIs.
 
-## Environment Variables
+### API Usage
 
-- `NCPU`: sets the number of cores to use
-
-When it comes to selecting a Slepian region the order precedence is
-[polar cap region](https://doi.org/10.1111/j.1365-246X.2006.03065.x) >
-[limited latitude longitude region](https://doi.org/10.1109/TSP.2016.2646668) >
-arbitrary region,
-[as seen in the code](https://github.com/astro-informatics/sleplet/blob/main/src/sleplet/utils/region.py).
-The default region is the `south_america` arbitrary region.
-
-- `POLAR_GAP`
-  - for a Slepian `polar cap region`, when set in conjunction with `THETA_MAX`
-    but without the other `PHI`/`THETA` variables
-- `THETA_MAX`
-  - for a Slepian `polar cap region`, when set without the other `PHI`/`THETA`
-    variables
[…]
