### 2023-04-05 astro-informatics/sleplet#126 @paddyroddy
README changes done in #128
