+++ src/sleplet/__init__.py
@@ -1,4 +1,4 @@
-""".. include:: ./../../documentation/documentation.md"""  # noqa: D400,D415
+""".. include:: ./../../documentation/DOCUMENTATION.md"""  # noqa: D400,D415
 
 import logging
 
