### 2023-04-05 astro-informatics/sleplet#126 @Saran-nns
thanks for your prompt response and the fixes. I have a couple of minor requests regarding the README figure:

- [ ] In readme, The size of the colorbar and its font appears to be relatively large compared to the overall figure size. Can we adjust them to be more proportional?
- [ ] citation for plotly https://plotly.com/chart-studio-help/citations/.

thanks
