### 2023-04-05 #5221 @Saran-nns
@vissarion @paddyroddy Couple of [steps](https://github.com/astro-informatics/sleplet/issues/126#issuecomment-1497081457) to go. Will be happy to flag accept then
