### 2023-04-05 075ee16
Revert "Check hyperlinking HTML"
This reverts commit 33358e042622847307af1a9a79500e62c1b61def.
