### 2023-04-05 417f9f7
Include plotly and numpy citations
