+++ paper/paper.bib
@@ -1,12 +1,22 @@
 @article{Bates2017,
-  author  = {Bates, Alice P. and Khalid, Zubair and Kennedy, Rodney A.},
   title   = {{Slepian Spatial-Spectral Concentration Problem on the Sphere: Analytical Formulation for Limited Colatitude-Longitude Spatial Region}},
+  author  = {Bates, Alice P. and Khalid, Zubair and Kennedy, Rodney A.},
   year    = 2017,
   journal = {IEEE Transactions on Signal Processing},
   volume  = 65,
   number  = 6,
   pages   = {1527--1537},
   doi     = {10.1109/TSP.2016.2646668}
+}
+@article{Harris2020,
+  title   = {{Array programming with NumPy}},
+  author  = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
+  year    = 2020,
+  journal = {Nature},
+  volume  = 585,
+  number  = 7825,
+  pages   = {357--362},
+  doi     = {10.1038/s41586-020-2649-2}
 }
 @article{Landau1961,
   title   = {{Prolate spheroidal wave functions, fourier analysis and uncertainty -- II}},
@@ -56,6 +66,12 @@
   pages   = {5876--5887},
   doi     = {10.1109/TSP.2011.2166394}
 }
+@online{Plotly2015,
+  title     = {{Collaborative data science}},
+  author    = {Plotly Technologies Inc.},
+  year      = 2015,
+  url       = {https://plot.ly}
+}
 @article{Roddy2021,
   title   = {{Sifting Convolution on the Sphere}},
   author  = {Roddy, Patrick J. and McEwen, Jason D.},
@@ -75,11 +91,11 @@
   doi     = {10.1109/TSP.2022.3233309}
 }
 @phdthesis{Roddy2022a,
-  author = {Roddy, Patrick J.},
   title  = {{Slepian Wavelets for the Analysis of Incomplete Data on Manifolds}},
+  author = {Roddy, Patrick J.},
   year   = 2022,
-  school = {UCL (University College London)},
-  url    = {https://paddyroddy.github.io/thesis}
+  url    = {https://paddyroddy.github.io/thesis},
+  school = {UCL (University College London)}
 }
 @misc{Roddy2023,
   title     = {{Slepian Scale-Discretised Wavelets on Manifolds}},
@@ -91,7 +107,7 @@
 @software{Roddy2023a,
   title  = {{SLEPLET}},
   author = {Roddy, Patrick J.},
-  year   = {2023},
+  year   = 2023,
   doi    = {10.5281/zenodo.7268074}
 }
 @article{Simons2006,
@@ -105,10 +121,11 @@
   doi     = {10.1137/S0036144504445765}
 }
 @software{Simons2020,
-  title  = {{slepian\_alpha}},
-  author = {Frederik J Simons and Chris Harig and Alain Plattner and Max von Hippel and Albert},
-  year   = 2020,
-  doi    = {10.5281/zenodo.4085210}
+  title = {{slepian\_alpha}}
+},
[…]
