### 2023-04-05 66913c5
Download new figures from browser
