### 2023-04-05 a4e323b
Update figures & save plots with filename from browser (#135)
