### 2023-04-05 astro-informatics/sleplet#126 @paddyroddy
I've added those citations to the paper. I've changed the colour bar in #135.
