### 2023-04-05 11985b8
Remove `bibtex` placeholder
