+++ paper/paper.bib
@@ -67,10 +67,10 @@
   doi     = {10.1109/TSP.2011.2166394}
 }
 @online{Plotly2015,
-  title     = {{Collaborative data science}},
-  author    = {Plotly Technologies Inc.},
-  year      = 2015,
-  url       = {https://plot.ly}
+  title  = {{Collaborative data science}},
+  author = {Plotly Technologies Inc.},
+  year   = 2015,
+  url    = {https://plot.ly}
 }
 @article{Roddy2021,
   title   = {{Sifting Convolution on the Sphere}},
@@ -121,11 +121,10 @@
   doi     = {10.1137/S0036144504445765}
 }
 @software{Simons2020,
-  title = {{slepian\_alpha}}
-},
-    author       = {Frederik J Simons and Chris Harig and Alain Plattner and Max von Hippel and Albert},
-    year         = 2020,
-    doi          = {10.5281/zenodo.4085210},
+  title  = {{slepian\_alpha}},
+  author = {Frederik J Simons and Chris Harig and Alain Plattner and Max von Hippel and Albert},
+  year   = 2020,
+  doi    = {10.5281/zenodo.4085210}
 }
 @article{Slepian1961,
   title   = {{Prolate spheroidal wave functions, fourier analysis and uncertainty -- I}},
