### 2023-04-05 91a6153
JOSS submission (#72)
Fixes #46
