+++ .github/workflows/deploy.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Build distributions
         run: python -m build
 
-      - name: Publish package to PyPI
+      - name: Publish package to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           password: ${{ secrets.TEST_PYPI_API_TOKEN }}
