### 2023-04-05 1470d04
Use raw images for PyPI
