### 2023-04-05 #5221 @vissarion
Thanks! Please use the full title in the archive i.e. "SLEPLET: Slepian scale-discretised wavelets in Python" which is the same as the paper.
