### 2023-04-06 #5221 @vissarion
@paddyroddy I am copying my message again here because I am afraid it got lost after testing all the above commands

" Thanks! Please use the full title in the archive i.e. "SLEPLET: Slepian scale-discretised wavelets in Python" which is the same as the paper."
