### 2023-04-06 #5221 @vissarion
> > Thanks! Please use the full title in the archive i.e. "SLEPLET: Slepian scale-discretised wavelets in Python" which is the same as the paper.
>
> Sorry do you mean in the `paper.bib` in the archive or do you mean the archive name should be that?

I mean the zenodo archive name, thanks.
