### 2023-04-13 #5221 @Saran-nns
@paddyroddy I think @vissarion asks for the zenodo archive title to be same as the title of the paper `SLEPLET: Slepian Scale-Discretised Wavelets in Python`.  But the archive title is still `SLEPLET`.
