+++ paper/paper.bib
@@ -105,7 +105,7 @@
   url       = {https://arxiv.org/abs/2302.06006}
 }
 @software{Roddy2023a,
-  title  = {{SLEPLET}},
+  title  = {{SLEPLET: Slepian Scale-Discretised Wavelets in Python}},
   author = {Roddy, Patrick J.},
   year   = 2023,
   doi    = {10.5281/zenodo.7268074}
