### 2023-04-17 87760fd
Remove installing pip step (see actions/setup-python#506)
