+++ .github/workflows/deploy.yml
@@ -20,9 +20,7 @@ jobs:
           cache-dependency-path: "pyproject.toml"
 
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install build
+        run: python -m pip install build
 
       - name: Build distributions
         run: python -m build
