+++ .github/workflows/examples.yml
@@ -25,9 +25,7 @@ jobs:
           key: readme-${{ hashFiles('pyproject.toml') }}
 
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e .[readme]
+        run: python -m pip install -e .[readme]
 
       - name: Run examples in the README
         run: python -m pytest --codeblocks .github README.md documentation
@@ -53,9 +51,7 @@ jobs:
           key: examples-${{ hashFiles('pyproject.toml') }}
 
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e .[docs]
+        run: python -m pip install -e .[docs]
 
       - name: Run examples in the examples folder
         run: >-
