+++ .github/workflows/test.yml
@@ -35,9 +35,7 @@ jobs:
           cache-dependency-path: "pyproject.toml"
 
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install tox tox-gh-actions
+        run: python -m pip install tox tox-gh-actions
 
       - name: Test with tox
         run: tox
