+++ paper/paper.md
@@ -20,12 +20,12 @@ bibliography: paper.bib
 # Summary
 
 Wavelets are widely used in various disciplines to analyse signals both in space and scale.
-Whilst many fields measure data on manifolds (i.e. the sphere), often data are only observed on a partial region of the manifold.
-Wavelets are a typical approach to data of this form, but the wavelet coefficients which overlap with the boundary become contaminated and must be removed for accurate analysis.
+Whilst many fields measure data on manifolds (i.e., the sphere), often data are only observed on a partial region of the manifold.
+Wavelets are a typical approach to data of this form, but the wavelet coefficients that overlap with the boundary become contaminated and must be removed for accurate analysis.
 Another approach is to estimate the region of missing data and to use existing whole-manifold methods for analysis.
 However, both approaches introduce uncertainty into any analysis.
 Slepian wavelets enable one to work directly with only the data present, thus avoiding the problems discussed above.
-Possible applications of Slepian wavelets to areas of research measuring data on the partial sphere include: gravitational/magnetic fields in geodesy; ground-based measurements in astronomy; measurements of whole-planet properties in planetary science; and geomagnetism of the Earth.
+Applications of Slepian wavelets to areas of research measuring data on the partial sphere include gravitational/magnetic fields in geodesy, ground-based measurements in astronomy, measurements of whole-planet properties in planetary science, and geomagnetism of the Earth.
 Slepian wavelets have many potential applications in analyses of manifolds where data are only observed over a partial region.
 One such application is in cosmic microwave background analyses, where observations are inherently made on the celestial sphere, and foreground emissions mask the centre of the data.
 In fields such as astrophysics and cosmology, datasets are increasingly large and thus require analysis at high resolutions for accurate predictions.
@@ -37,23 +37,23 @@ Techniques developed in the Euclidean setting must be extended to other geometri
 Due to recent interest in geometric deep learning, analogues of Euclidean techniques must also handle general manifolds or graphs.
 Often, data are only observed over partial regions of manifolds, and thus standard whole-manifold techniques may not yield accurate predictions.
 Slepian wavelets are designed for datasets like these.
-Slepian wavelets are built upon the eigenfunctions of the Slepian concentration problem of the manifold [@Slepian1961; @Landau1961; @Landau1962] - a set of bandlimited functions which are maximally concentrated within a given region.
+Slepian wavelets are built upon the eigenfunctions of the Slepian concentration problem of the manifold [@Slepian1961; @Landau1961; @Landau1962]: a set of bandlimited functions that are maximally concentrated within a given region.
 Wavelets are constructed through a tiling of the Slepian harmonic line by leveraging the existing scale-discretised framework [@Wiaux2008; @Leistedt2013].
 Whilst these wavelets were inspired by spherical datasets, like in cosmology, the wavelet construction may be utilised for manifold or graph data.
 
-To the author's knowledge, there is no public software which allows one to compute Slepian wavelets
+To the author's knowledge, there is no public software that allows one to compute Slepian wavelets
 (or a similar approach) on the sphere or general manifolds/meshes.
 `SHTools` [@Wieczorek2018] is a `Python` code used for spherical harmonic transforms, which allows one to compute the Slepian functions of the spherical polar cap [@Simons2006].
-A series of `MATLAB` scripts exist in `slepian_alpha` [@Simons2020] which permits the calculation of the Slepian functions on the sphere.
+A series of `MATLAB` scripts exist in `slepian_alpha` [@Simons2020], which permits the calculation of the Slepian functions on the sphere.
 However, these scripts are very specialised and hard to generalise.
 
[…]
