+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Bates2017,
-  title   = {{Slepian Spatial-Spectral Concentration Problem on the Sphere: Analytical Formulation for Limited Colatitude-Longitude Spatial Region}},
+  title   = {Slepian Spatial-Spectral Concentration Problem on the Sphere: Analytical Formulation for Limited Colatitude-Longitude Spatial Region},
   author  = {Bates, Alice P. and Khalid, Zubair and Kennedy, Rodney A.},
   year    = 2017,
   journal = {IEEE Transactions on Signal Processing},
@@ -9,7 +9,7 @@
   doi     = {10.1109/TSP.2016.2646668}
 }
 @article{Harris2020,
-  title   = {{Array programming with NumPy}},
+  title   = {Array programming with {NumPy}},
   author  = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
   year    = 2020,
   journal = {Nature},
@@ -19,7 +19,7 @@
   doi     = {10.1038/s41586-020-2649-2}
 }
 @article{Landau1961,
-  title   = {{Prolate spheroidal wave functions, fourier analysis and uncertainty -- II}},
+  title   = {Prolate spheroidal wave functions, {F}ourier analysis and uncertainty -- {II}},
   author  = {Landau, H. J. and Pollak, H. O.},
   year    = 1961,
   journal = {The Bell System Technical Journal},
@@ -29,7 +29,7 @@
   doi     = {10.1002/j.1538-7305.1961.tb03977.x}
 }
 @article{Landau1962,
-  title   = {{Prolate spheroidal wave functions, fourier analysis and uncertainty -- III: The dimension of the space of essentially time- and band-limited signals}},
+  title   = {Prolate spheroidal wave functions, {F}ourier analysis and uncertainty -- {III}: The dimension of the space of essentially time- and band-limited signals},
   author  = {Landau, H. J. and Pollak, H. O.},
   year    = 1962,
   journal = {The Bell System Technical Journal},
@@ -39,7 +39,7 @@
   doi     = {10.1002/j.1538-7305.1962.tb03279.x}
 }
 @article{Leistedt2013,
-  title   = {{S2LET: A code to perform fast wavelet analysis on the sphere}},
+  title   = {S2LET: A code to perform fast wavelet analysis on the sphere},
   author  = {{Leistedt, B.} and {McEwen, J. D.} and {Vandergheynst, P.} and {Wiaux, Y.}},
   year    = 2013,
   journal = {Astronomy \& Astrophysics},
@@ -48,7 +48,7 @@
   doi     = {10.1051/0004-6361/201220729}
 }
 @inproceedings{Libigl2017,
-  title     = {{Libigl: Prototyping Geometry Processing Research in C++}},
+  title     = {Libigl: Prototyping Geometry Processing Research in {C}++},
   author    = {Jacobson, Alec and Panozzo, Daniele},
   year      = 2017,
[…]
