+++ paper/paper.md
@@ -25,10 +25,7 @@ Wavelets are a typical approach to data of this form, but the wavelet coefficien
 Another approach is to estimate the region of missing data and to use existing whole-manifold methods for analysis.
 However, both approaches introduce uncertainty into any analysis.
 Slepian wavelets enable one to work directly with only the data present, thus avoiding the problems discussed above.
-Applications of Slepian wavelets to areas of research measuring data on the partial sphere include gravitational/magnetic fields in geodesy, ground-based measurements in astronomy, measurements of whole-planet properties in planetary science, and geomagnetism of the Earth.
-Slepian wavelets have many potential applications in analyses of manifolds where data are only observed over a partial region.
-One such application is in cosmic microwave background analyses, where observations are inherently made on the celestial sphere, and foreground emissions mask the centre of the data.
-In fields such as astrophysics and cosmology, datasets are increasingly large and thus require analysis at high resolutions for accurate predictions.
+Applications of Slepian wavelets to areas of research measuring data on the partial sphere include gravitational/magnetic fields in geodesy, ground-based measurements in astronomy, measurements of whole-planet properties in planetary science, geomagnetism of the Earth, and cosmic microwave background analyses.
 
 # Statement of Need
 
