### 2023-04-20 #5221 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
