### 2023-04-20 652ae89
Add official citation from JOSS
