### 2023-04-20 818a1c1
Add Citation for JOSS (#138)
