+++ pyproject.toml
@@ -53,11 +53,7 @@ keywords = [
     "wavelets",
 ]
 name = "sleplet"
-readme = "README.md"
-requires-python = ">=3.10"
-scripts = {mesh = "sleplet._scripts.plotting_on_mesh:main", sphere = "sleplet._scripts.plotting_on_sphere:main"}
-license.file = "LICENCE.md"
-optional-dependencies.dev = [
+optional-dependencies = {dev = [
     "black",
     "build",
     "mypy",
@@ -65,18 +61,16 @@ optional-dependencies.dev = [
     "ruff",
     "tox",
     "twine",
-]
-optional-dependencies.docs = [
+], docs = [
     "pdoc3",
-]
-optional-dependencies.readme = [
+], readme = [
     "pytest-codeblocks",
-]
-urls.Code = "https://github.com/astro-informatics/sleplet"
-urls.Documentation = "https://astro-informatics.github.io/sleplet"
-urls.Download = "https://pypi.org/project/sleplet"
-urls.Homepage = "https://github.com/astro-informatics/sleplet"
-urls.Issues = "https://github.com/astro-informatics/sleplet/issues"
+]}
+readme = "README.md"
+requires-python = ">=3.10"
+scripts = {mesh = "sleplet._scripts.plotting_on_mesh:main", sphere = "sleplet._scripts.plotting_on_sphere:main"}
+urls = {Code = "https://github.com/astro-informatics/sleplet", Documentation = "https://astro-informatics.github.io/sleplet", Download = "https://pypi.org/project/sleplet", Homepage = "https://github.com/astro-informatics/sleplet", Issues = "https://github.com/astro-informatics/sleplet/issues"}
+license.file = "LICENCE.md"
 
 [tool.coverage]
 report = {skip_covered = true, sort = "cover"}
