+++ .github/workflows/test.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Cache tox/pooch
         uses: actions/cache@v3
         with:
-          path: |
+          path: |-
             .tox
             ~/.cache/sleplet
           key: test-${{ hashFiles('pyproject.toml') }}
@@ -35,9 +35,7 @@ jobs:
           cache-dependency-path: "pyproject.toml"
 
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install tox tox-gh-actions
+        run: python -m pip install tox tox-gh-actions
 
       - name: Test with tox
         run: tox
