+++ .github/workflows/documentation.yml
@@ -20,9 +20,7 @@ jobs:
           cache-dependency-path: "pyproject.toml"
 
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e .[docs]
+        run: python -m pip install -e .[docs]
 
       - name: Create documentation
         run: pdoc --html src/sleplet --template-dir documentation
